Module melior::dialect::ods::amdgpu

amdgpu dialect.

The AMDGPU dialect provides wrappers around AMD-specific functionality and LLVM intrinsics. These wrappers should be used in conjunction with more generic dialects, such as gpu and vector, when generating LLVM IR that will eventually be executed on AMD hardware.

Structs

LDSBarrierOperation

A lds_barrier operation. Barrier that includes a wait for LDS memory operations..

LDSBarrierOperationBuilder

A builder for a lds_barrier operation.

MFMAOperation

A mfma operation. MLIR wrapper for CDNA mfma instructions.

MFMAOperationBuilder

A builder for a mfma operation.

RawBufferAtomicCmpswapOperation

A raw_buffer_atomic_cmpswap operation. Raw Buffer Atomic compare-and-swap.

RawBufferAtomicCmpswapOperationBuilder

A builder for a raw_buffer_atomic_cmpswap operation.

RawBufferAtomicFaddOperation

A raw_buffer_atomic_fadd operation. Raw Buffer Floating-point Atomic Add (MI-* only).

RawBufferAtomicFaddOperationBuilder

A builder for a raw_buffer_atomic_fadd operation.

RawBufferAtomicFmaxOperation

A raw_buffer_atomic_fmax operation. Raw Buffer Floating-point Atomic Max (non-GFX9).

RawBufferAtomicFmaxOperationBuilder

A builder for a raw_buffer_atomic_fmax operation.

RawBufferAtomicSmaxOperation

A raw_buffer_atomic_smax operation. Raw Buffer Signed Integer Atomic Max.

RawBufferAtomicSmaxOperationBuilder

A builder for a raw_buffer_atomic_smax operation.

RawBufferAtomicUminOperation

A raw_buffer_atomic_umin operation. Raw Buffer Unsigned Integer Atomic Min.

RawBufferAtomicUminOperationBuilder

A builder for a raw_buffer_atomic_umin operation.

RawBufferLoadOperation

A raw_buffer_load operation. Raw Buffer load, exposing GCN features.

RawBufferLoadOperationBuilder

A builder for a raw_buffer_load operation.

RawBufferStoreOperation

A raw_buffer_store operation. Raw Buffer Store, exposing GCN features.

RawBufferStoreOperationBuilder

A builder for a raw_buffer_store operation.

WMMAOperation

A wmma operation. MLIR wrapper for RDNA3 wmma instructions.

WMMAOperationBuilder

A builder for a wmma operation.

Functions

lds_barrier

Creates a lds_barrier operation.

mfma

Creates a mfma operation.

raw_buffer_atomic_cmpswap

Creates a raw_buffer_atomic_cmpswap operation.

raw_buffer_atomic_fadd

Creates a raw_buffer_atomic_fadd operation.

raw_buffer_atomic_fmax

Creates a raw_buffer_atomic_fmax operation.

raw_buffer_atomic_smax

Creates a raw_buffer_atomic_smax operation.

raw_buffer_atomic_umin

Creates a raw_buffer_atomic_umin operation.

raw_buffer_load

Creates a raw_buffer_load operation.

raw_buffer_store

Creates a raw_buffer_store operation.

wmma

Creates a wmma operation.