Module melior::dialect::ods::arith

arith dialect.

The arith dialect is intended to hold basic integer and floating point mathematical operations. This includes unary, binary, and ternary arithmetic ops, bitwise and shift ops, cast ops, and compare ops. Operations in this dialect also accept vectors and tensors of integers or floats.

Structs

AddFOperation

A(n) addf operation. floating point addition operation.

AddFOperationBuilder

A builder for a(n) addf operation.

AddIOperation

A(n) addi operation. integer addition operation.

AddIOperationBuilder

A builder for a(n) addi operation.

AddUIExtendedOperation

A(n) addui_extended operation. extended unsigned integer addition operation returning sum and overflow bit .

AddUIExtendedOperationBuilder

A builder for a(n) addui_extended operation.

AndIOperation

A(n) andi operation. integer binary and.

AndIOperationBuilder

A builder for a(n) andi operation.

BitcastOperation

A(n) bitcast operation. bitcast between values of equal bit width.

BitcastOperationBuilder

A builder for a(n) bitcast operation.

CeilDivSIOperation

A(n) ceildivsi operation. signed ceil integer division operation.

CeilDivSIOperationBuilder

A builder for a(n) ceildivsi operation.

CeilDivUIOperation

A(n) ceildivui operation. unsigned ceil integer division operation.

CeilDivUIOperationBuilder

A builder for a(n) ceildivui operation.

CmpFOperation

A(n) cmpf operation. floating-point comparison operation.

CmpFOperationBuilder

A builder for a(n) cmpf operation.

CmpIOperation

A(n) cmpi operation. integer comparison operation.

CmpIOperationBuilder

A builder for a(n) cmpi operation.

ConstantOperation

A(n) constant operation. integer or floating point constant.

ConstantOperationBuilder

A builder for a(n) constant operation.

DivFOperation

A(n) divf operation. floating point division operation.

DivFOperationBuilder

A builder for a(n) divf operation.

DivSIOperation

A(n) divsi operation. signed integer division operation.

DivSIOperationBuilder

A builder for a(n) divsi operation.

DivUIOperation

A(n) divui operation. unsigned integer division operation.

DivUIOperationBuilder

A builder for a(n) divui operation.

ExtFOperation

A(n) extf operation. cast from floating-point to wider floating-point.

ExtFOperationBuilder

A builder for a(n) extf operation.

ExtSIOperation

A(n) extsi operation. integer sign extension operation.

ExtSIOperationBuilder

A builder for a(n) extsi operation.

ExtUIOperation

A(n) extui operation. integer zero extension operation.

ExtUIOperationBuilder

A builder for a(n) extui operation.

FPToSIOperation

A(n) fptosi operation. cast from floating-point type to integer type.

FPToSIOperationBuilder

A builder for a(n) fptosi operation.

FPToUIOperation

A(n) fptoui operation. cast from floating-point type to integer type.

FPToUIOperationBuilder

A builder for a(n) fptoui operation.

FloorDivSIOperation

A(n) floordivsi operation. signed floor integer division operation.

FloorDivSIOperationBuilder

A builder for a(n) floordivsi operation.

IndexCastOperation

A(n) index_cast operation. cast between index and integer types.

IndexCastOperationBuilder

A builder for a(n) index_cast operation.

IndexCastUIOperation

A(n) index_castui operation. unsigned cast between index and integer types.

IndexCastUIOperationBuilder

A builder for a(n) index_castui operation.

MaxFOperation

A(n) maxf operation. floating-point maximum operation.

MaxFOperationBuilder

A builder for a(n) maxf operation.

MaxSIOperation

A(n) maxsi operation. signed integer maximum operation.

MaxSIOperationBuilder

A builder for a(n) maxsi operation.

MaxUIOperation

A(n) maxui operation. unsigned integer maximum operation.

MaxUIOperationBuilder

A builder for a(n) maxui operation.

MinFOperation

A(n) minf operation. floating-point minimum operation.

MinFOperationBuilder

A builder for a(n) minf operation.

MinSIOperation

A(n) minsi operation. signed integer minimum operation.

MinSIOperationBuilder

A builder for a(n) minsi operation.

MinUIOperation

A(n) minui operation. unsigned integer minimum operation.

MinUIOperationBuilder

A builder for a(n) minui operation.

MulFOperation

A(n) mulf operation. floating point multiplication operation.

MulFOperationBuilder

A builder for a(n) mulf operation.

MulIOperation

A(n) muli operation. integer multiplication operation.

MulIOperationBuilder

A builder for a(n) muli operation.

MulSIExtendedOperation

A(n) mulsi_extended operation. extended signed integer multiplication operation .

MulSIExtendedOperationBuilder

A builder for a(n) mulsi_extended operation.

MulUIExtendedOperation

A(n) mului_extended operation. extended unsigned integer multiplication operation .

MulUIExtendedOperationBuilder

A builder for a(n) mului_extended operation.

NegFOperation

A(n) negf operation. floating point negation.

NegFOperationBuilder

A builder for a(n) negf operation.

OrIOperation

A(n) ori operation. integer binary or.

OrIOperationBuilder

A builder for a(n) ori operation.

RemFOperation

A(n) remf operation. floating point division remainder operation.

RemFOperationBuilder

A builder for a(n) remf operation.

RemSIOperation

A(n) remsi operation. signed integer division remainder operation.

RemSIOperationBuilder

A builder for a(n) remsi operation.

RemUIOperation

A(n) remui operation. unsigned integer division remainder operation.

RemUIOperationBuilder

A builder for a(n) remui operation.

SIToFPOperation

A(n) sitofp operation. cast from integer type to floating-point.

SIToFPOperationBuilder

A builder for a(n) sitofp operation.

SelectOperation

A(n) select operation. select operation.

SelectOperationBuilder

A builder for a(n) select operation.

ShLIOperation

A(n) shli operation. integer left-shift.

ShLIOperationBuilder

A builder for a(n) shli operation.

ShRSIOperation

A(n) shrsi operation. signed integer right-shift.

ShRSIOperationBuilder

A builder for a(n) shrsi operation.

ShRUIOperation

A(n) shrui operation. unsigned integer right-shift.

ShRUIOperationBuilder

A builder for a(n) shrui operation.

SubFOperation

A(n) subf operation. floating point subtraction operation.

SubFOperationBuilder

A builder for a(n) subf operation.

SubIOperation

A(n) subi operation. integer subtraction operation.

SubIOperationBuilder

A builder for a(n) subi operation.

TruncFOperation

A(n) truncf operation. cast from floating-point to narrower floating-point.

TruncFOperationBuilder

A builder for a(n) truncf operation.

TruncIOperation

A(n) trunci operation. integer truncation operation.

TruncIOperationBuilder

A builder for a(n) trunci operation.

UIToFPOperation

A(n) uitofp operation. cast from unsigned integer type to floating-point.

UIToFPOperationBuilder

A builder for a(n) uitofp operation.

XOrIOperation

A(n) xori operation. integer binary xor.

XOrIOperationBuilder

A builder for a(n) xori operation.

Functions

addf

Creates a(n) addf operation.

addi

Creates a(n) addi operation.

addui_extended

Creates a(n) addui_extended operation.

andi

Creates a(n) andi operation.

bitcast

Creates a(n) bitcast operation.

ceildivsi

Creates a(n) ceildivsi operation.

ceildivui

Creates a(n) ceildivui operation.

cmpf

Creates a(n) cmpf operation.

cmpi

Creates a(n) cmpi operation.

constant

Creates a(n) constant operation.

divf

Creates a(n) divf operation.

divsi

Creates a(n) divsi operation.

divui

Creates a(n) divui operation.

extf

Creates a(n) extf operation.

extsi

Creates a(n) extsi operation.

extui

Creates a(n) extui operation.

floordivsi

Creates a(n) floordivsi operation.

fptosi

Creates a(n) fptosi operation.

fptoui

Creates a(n) fptoui operation.

index_cast

Creates a(n) index_cast operation.

index_castui

Creates a(n) index_castui operation.

maxf

Creates a(n) maxf operation.

maxsi

Creates a(n) maxsi operation.

maxui

Creates a(n) maxui operation.

minf

Creates a(n) minf operation.

minsi

Creates a(n) minsi operation.

minui

Creates a(n) minui operation.

mulf

Creates a(n) mulf operation.

muli

Creates a(n) muli operation.

mulsi_extended

Creates a(n) mulsi_extended operation.

mului_extended

Creates a(n) mului_extended operation.

negf

Creates a(n) negf operation.

ori

Creates a(n) ori operation.

remf

Creates a(n) remf operation.

remsi

Creates a(n) remsi operation.

remui

Creates a(n) remui operation.

select

Creates a(n) select operation.

shli

Creates a(n) shli operation.

shrsi

Creates a(n) shrsi operation.

shrui

Creates a(n) shrui operation.

sitofp

Creates a(n) sitofp operation.

subf

Creates a(n) subf operation.

subi

Creates a(n) subi operation.

truncf

Creates a(n) truncf operation.

trunci

Creates a(n) trunci operation.

uitofp

Creates a(n) uitofp operation.

xori

Creates a(n) xori operation.