Module melior::dialect::ods::async

async dialect.

This dialect contains operations for modeling asynchronous execution.

Structs

AddToGroupOperation

A(n) add_to_group operation. adds and async token or value to the group.

AddToGroupOperationBuilder

A builder for a(n) add_to_group operation.

AwaitAllOperation

A(n) await_all operation. waits for the all async tokens or values in the group to become ready.

AwaitAllOperationBuilder

A builder for a(n) await_all operation.

AwaitOperation

A(n) await operation. waits for the argument to become ready.

AwaitOperationBuilder

A builder for a(n) await operation.

CallOperation

A(n) call operation. async call operation.

CallOperationBuilder

A builder for a(n) call operation.

CoroBeginOperation

A(n) coro.begin operation. returns a handle to the coroutine.

CoroBeginOperationBuilder

A builder for a(n) coro.begin operation.

CoroEndOperation

A(n) coro.end operation. marks the end of the coroutine in the suspend block.

CoroEndOperationBuilder

A builder for a(n) coro.end operation.

CoroFreeOperation

A(n) coro.free operation. deallocates the coroutine frame.

CoroFreeOperationBuilder

A builder for a(n) coro.free operation.

CoroIdOperation

A(n) coro.id operation. returns a switched-resume coroutine identifier.

CoroIdOperationBuilder

A builder for a(n) coro.id operation.

CoroSaveOperation

A(n) coro.save operation. saves the coroutine state.

CoroSaveOperationBuilder

A builder for a(n) coro.save operation.

CoroSuspendOperation

A(n) coro.suspend operation. suspends the coroutine.

CoroSuspendOperationBuilder

A builder for a(n) coro.suspend operation.

CreateGroupOperation

A(n) create_group operation. creates an empty async group.

CreateGroupOperationBuilder

A builder for a(n) create_group operation.

ExecuteOperation

A(n) execute operation. Asynchronous execute operation.

ExecuteOperationBuilder

A builder for a(n) execute operation.

FuncOperation

A(n) func operation. async function operation.

FuncOperationBuilder

A builder for a(n) func operation.

ReturnOperation

A(n) return operation. Async function return operation.

ReturnOperationBuilder

A builder for a(n) return operation.

RuntimeAddRefOperation

A(n) runtime.add_ref operation. adds a reference to async value.

RuntimeAddRefOperationBuilder

A builder for a(n) runtime.add_ref operation.

RuntimeAddToGroupOperation

A(n) runtime.add_to_group operation. adds and async token or value to the group.

RuntimeAddToGroupOperationBuilder

A builder for a(n) runtime.add_to_group operation.

RuntimeAwaitAndResumeOperation

A(n) runtime.await_and_resume operation. awaits the async operand and resumes the coroutine.

RuntimeAwaitAndResumeOperationBuilder

A builder for a(n) runtime.await_and_resume operation.

RuntimeAwaitOperation

A(n) runtime.await operation. blocks the caller thread until the operand becomes available.

RuntimeAwaitOperationBuilder

A builder for a(n) runtime.await operation.

RuntimeCreateGroupOperation

A(n) runtime.create_group operation. creates an async runtime group.

RuntimeCreateGroupOperationBuilder

A builder for a(n) runtime.create_group operation.

RuntimeCreateOperation

A(n) runtime.create operation. creates an async runtime token or value.

RuntimeCreateOperationBuilder

A builder for a(n) runtime.create operation.

RuntimeDropRefOperation

A(n) runtime.drop_ref operation. drops a reference to async value.

RuntimeDropRefOperationBuilder

A builder for a(n) runtime.drop_ref operation.

RuntimeIsErrorOperation

A(n) runtime.is_error operation. returns true if token, value or group is in error state.

RuntimeIsErrorOperationBuilder

A builder for a(n) runtime.is_error operation.

RuntimeLoadOperation

A(n) runtime.load operation. loads the value from the runtime async.value.

RuntimeLoadOperationBuilder

A builder for a(n) runtime.load operation.

RuntimeNumWorkerThreadsOperation

A(n) runtime.num_worker_threads operation. gets the number of threads in the threadpool from the runtime.

RuntimeNumWorkerThreadsOperationBuilder

A builder for a(n) runtime.num_worker_threads operation.

RuntimeResumeOperation

A(n) runtime.resume operation. resumes the coroutine on a thread managed by the runtime.

RuntimeResumeOperationBuilder

A builder for a(n) runtime.resume operation.

RuntimeSetAvailableOperation

A(n) runtime.set_available operation. switches token or value to available state.

RuntimeSetAvailableOperationBuilder

A builder for a(n) runtime.set_available operation.

RuntimeSetErrorOperation

A(n) runtime.set_error operation. switches token or value to error state.

RuntimeSetErrorOperationBuilder

A builder for a(n) runtime.set_error operation.

RuntimeStoreOperation

A(n) runtime.store operation. stores the value into the runtime async.value.

RuntimeStoreOperationBuilder

A builder for a(n) runtime.store operation.

YieldOperation

A(n) yield operation. terminator for Async execute operation.

YieldOperationBuilder

A builder for a(n) yield operation.

Functions

add_to_group

Creates a(n) add_to_group operation.

await

Creates a(n) await operation.

await_all

Creates a(n) await_all operation.

call

Creates a(n) call operation.

coro_begin

Creates a(n) coro.begin operation.

coro_end

Creates a(n) coro.end operation.

coro_free

Creates a(n) coro.free operation.

coro_id

Creates a(n) coro.id operation.

coro_save

Creates a(n) coro.save operation.

coro_suspend

Creates a(n) coro.suspend operation.

create_group

Creates a(n) create_group operation.

execute

Creates a(n) execute operation.

func

Creates a(n) func operation.

return

Creates a(n) return operation.

runtime_add_ref

Creates a(n) runtime.add_ref operation.

runtime_add_to_group

Creates a(n) runtime.add_to_group operation.

runtime_await

Creates a(n) runtime.await operation.

runtime_await_and_resume

Creates a(n) runtime.await_and_resume operation.

runtime_create

Creates a(n) runtime.create operation.

runtime_create_group

Creates a(n) runtime.create_group operation.

runtime_drop_ref

Creates a(n) runtime.drop_ref operation.

runtime_is_error

Creates a(n) runtime.is_error operation.

runtime_load

Creates a(n) runtime.load operation.

runtime_num_worker_threads

Creates a(n) runtime.num_worker_threads operation.

runtime_resume

Creates a(n) runtime.resume operation.

runtime_set_available

Creates a(n) runtime.set_available operation.

runtime_set_error

Creates a(n) runtime.set_error operation.

runtime_store

Creates a(n) runtime.store operation.

yield

Creates a(n) yield operation.