Module melior::dialect::ods::bufferization

bufferization dialect.

Bufferization in MLIR is the process of converting the tensor type to the memref type. Simply put, bufferization is the process of converting computations on the mathematical tensor construct to computations on physical memory buffers. The bufferization dialect contains operations/interfaces specific to the bufferization passes.

An overview of the bufferization infrastructure and important conceptual details related to using the MLIR dialect conversion infrastructure can be found in bufferization and buffer deallocation.

Structs

AllocTensorOperation

An alloc_tensor operation. Allocate buffer for a tensor.

AllocTensorOperationBuilder

A builder for an alloc_tensor operation.

CloneOperation

A clone operation. Clone a memref.

CloneOperationBuilder

A builder for a clone operation.

CopyTensorOperation

A copy_tensor operation. Copy a tensor.

CopyTensorOperationBuilder

A builder for a copy_tensor operation.

DeallocOperation

A dealloc operation. Deallocates the given memrefs if no alias is retained.

DeallocOperationBuilder

A builder for a dealloc operation.

DeallocTensorOperation

A dealloc_tensor operation. Release underlying storage format of given tensor.

DeallocTensorOperationBuilder

A builder for a dealloc_tensor operation.

ToMemrefOperation

A to_memref operation. Cast a tensor to memref.

ToMemrefOperationBuilder

A builder for a to_memref operation.

ToTensorOperation

A to_tensor operation. Create a tensor from a memref.

ToTensorOperationBuilder

A builder for a to_tensor operation.

Functions

alloc_tensor

Creates an alloc_tensor operation.

clone

Creates a clone operation.

copy_tensor

Creates a copy_tensor operation.

dealloc

Creates a dealloc operation.

dealloc_tensor

Creates a dealloc_tensor operation.

to_memref

Creates a to_memref operation.

to_tensor

Creates a to_tensor operation.