Module melior::dialect::ods::func

func dialect.

Structs

CallIndirectOp

call_indirect operation. Indirect call operation.

CallIndirectOpBuilder

A builder for call_indirect operation. Indirect call operation.

CallOp

call operation. Call operation.

CallOpBuilder

A builder for call operation. Call operation.

ConstantOp

constant operation. Constant.

ConstantOpBuilder

A builder for constant operation. Constant.

FuncOp

func operation. An operation with a name containing a single SSACFG region.

FuncOpBuilder

A builder for func operation. An operation with a name containing a single SSACFG region.

ReturnOp

return operation. Function return operation.

ReturnOpBuilder

A builder for return operation. Function return operation.

Functions

call

Creates a call operation. Call operation.

call_indirect

Creates a call_indirect operation. Indirect call operation.

constant

Creates a constant operation. Constant.

func

Creates a func operation. An operation with a name containing a single SSACFG region.

return

Creates a return operation. Function return operation.