Module melior::dialect::ods::gpu

gpu dialect.

Structs

AllReduceOperation

An all_reduce operation. Reduce values among workgroup..

AllReduceOperationBuilder

A builder for an all_reduce operation.

AllocOperation

An alloc operation. GPU memory allocation operation..

AllocOperationBuilder

A builder for an alloc operation.

BarrierOperation

A barrier operation. Synchronizes all work items of a workgroup..

BarrierOperationBuilder

A builder for a barrier operation.

BlockDimOperation

A block_dim operation.

BlockDimOperationBuilder

A builder for a block_dim operation.

BlockIdOperation

A block_id operation.

BlockIdOperationBuilder

A builder for a block_id operation.

Create2To4SpMatOperation

A create_2to4_spmat operation. Create sparse matrix with 2:4 sparsity operation.

Create2To4SpMatOperationBuilder

A builder for a create_2to4_spmat operation.

CreateCooAoSOperation

A create_coo_aos operation. Create sparse matrix in COO format operation (AoS).

CreateCooAoSOperationBuilder

A builder for a create_coo_aos operation.

CreateCooOperation

A create_coo operation. Create sparse matrix in COO format operation.

CreateCooOperationBuilder

A builder for a create_coo operation.

CreateCsrOperation

A create_csr operation. Create sparse matrix in CSR format operation.

CreateCsrOperationBuilder

A builder for a create_csr operation.

CreateDnTensorOperation

A create_dn_tensor operation. Create dense tensor operation.

CreateDnTensorOperationBuilder

A builder for a create_dn_tensor operation.

DeallocOperation

A dealloc operation. GPU memory deallocation operation.

DeallocOperationBuilder

A builder for a dealloc operation.

DestroyDnTensorOperation

A destroy_dn_tensor operation. Destroy dense tensor operation.

DestroyDnTensorOperationBuilder

A builder for a destroy_dn_tensor operation.

DestroySpMatOperation

A destroy_sp_mat operation. Destroy sparse matrix operation.

DestroySpMatOperationBuilder

A builder for a destroy_sp_mat operation.

GPUFuncOperation

A func operation. Function executable on a GPU.

GPUFuncOperationBuilder

A builder for a func operation.

GPUModuleOperation

A module operation. A top level compilation unit containing code to be run on a GPU..

GPUModuleOperationBuilder

A builder for a module operation.

GlobalIdOperation

A global_id operation.

GlobalIdOperationBuilder

A builder for a global_id operation.

GridDimOperation

A grid_dim operation.

GridDimOperationBuilder

A builder for a grid_dim operation.

HostRegisterOperation

A host_register operation. Registers a memref for access from device..

HostRegisterOperationBuilder

A builder for a host_register operation.

HostUnregisterOperation

A host_unregister operation. Unregisters a memref for access from device..

HostUnregisterOperationBuilder

A builder for a host_unregister operation.

LaneIdOperation

A lane_id operation.

LaneIdOperationBuilder

A builder for a lane_id operation.

LaunchFuncOperation

A launch_func operation. Launches a function as a GPU kernel.

LaunchFuncOperationBuilder

A builder for a launch_func operation.

LaunchOperation

A launch operation. GPU kernel launch operation.

LaunchOperationBuilder

A builder for a launch operation.

MemcpyOperation

A memcpy operation. GPU memcpy operation.

MemcpyOperationBuilder

A builder for a memcpy operation.

MemsetOperation

A memset operation. GPU memset operation.

MemsetOperationBuilder

A builder for a memset operation.

ModuleEndOperation

A module_end operation. A pseudo op that marks the end of a gpu.module..

ModuleEndOperationBuilder

A builder for a module_end operation.

NumSubgroupsOperation

A num_subgroups operation.

NumSubgroupsOperationBuilder

A builder for a num_subgroups operation.

PrintfOperation

A printf operation. Device-side printf, as in CUDA or OpenCL, for debugging.

PrintfOperationBuilder

A builder for a printf operation.

ReturnOperation

A return operation. Terminator for GPU functions..

ReturnOperationBuilder

A builder for a return operation.

SDDMMBufferSizeOperation

A sddmm_buffer_size operation. Precompute buffersize for SDDMM operation.

SDDMMBufferSizeOperationBuilder

A builder for a sddmm_buffer_size operation.

SDDMMOperation

A sddmm operation. SDDMM operation.

SDDMMOperationBuilder

A builder for a sddmm operation.

SetDefaultDeviceOperation

A set_default_device operation. Set default GPU for operations after this by index.

SetDefaultDeviceOperationBuilder

A builder for a set_default_device operation.

ShuffleOperation

A shuffle operation. Shuffles values within a subgroup..

ShuffleOperationBuilder

A builder for a shuffle operation.

SpMMBufferSizeOperation

A spmm_buffer_size operation. Precompute buffersize for SpMM operation.

SpMMBufferSizeOperationBuilder

A builder for a spmm_buffer_size operation.

SpMMOperation

A spmm operation. SpMM operation.

SpMMOperationBuilder

A builder for a spmm operation.

SpMVBufferSizeOperation

A spmv_buffer_size operation. Precompute buffersize for SpMV operation.

SpMVBufferSizeOperationBuilder

A builder for a spmv_buffer_size operation.

SpMVOperation

A spmv operation. SpMV operation.

SpMVOperationBuilder

A builder for a spmv operation.

SubgroupIdOperation

A subgroup_id operation.

SubgroupIdOperationBuilder

A builder for a subgroup_id operation.

SubgroupMmaComputeOperation

A subgroup_mma_compute operation. GPU warp synchronous matrix multiply accumulate.

SubgroupMmaComputeOperationBuilder

A builder for a subgroup_mma_compute operation.

SubgroupMmaConstantMatrixOperation

A subgroup_mma_constant_matrix operation. GPU warp synchronous constant matrix.

SubgroupMmaConstantMatrixOperationBuilder

A builder for a subgroup_mma_constant_matrix operation.

SubgroupMmaElementwiseOperation

A subgroup_mma_elementwise operation. GPU warp elementwise operation on a matrix.

SubgroupMmaElementwiseOperationBuilder

A builder for a subgroup_mma_elementwise operation.

SubgroupMmaLoadMatrixOperation

A subgroup_mma_load_matrix operation. GPU warp synchronous matrix load.

SubgroupMmaLoadMatrixOperationBuilder

A builder for a subgroup_mma_load_matrix operation.

SubgroupMmaStoreMatrixOperation

A subgroup_mma_store_matrix operation. GPU warp synchronous matrix store.

SubgroupMmaStoreMatrixOperationBuilder

A builder for a subgroup_mma_store_matrix operation.

SubgroupReduceOperation

A subgroup_reduce operation. Reduce values among subgroup..

SubgroupReduceOperationBuilder

A builder for a subgroup_reduce operation.

SubgroupSizeOperation

A subgroup_size operation.

SubgroupSizeOperationBuilder

A builder for a subgroup_size operation.

TerminatorOperation

A terminator operation. Terminator for GPU launch regions..

TerminatorOperationBuilder

A builder for a terminator operation.

ThreadIdOperation

A thread_id operation.

ThreadIdOperationBuilder

A builder for a thread_id operation.

WaitOperation

A wait operation. Wait for async gpu ops to complete..

WaitOperationBuilder

A builder for a wait operation.

YieldOperation

A yield operation. GPU yield operation.

YieldOperationBuilder

A builder for a yield operation.

Functions

all_reduce

Creates an all_reduce operation.

alloc

Creates an alloc operation.

barrier

Creates a barrier operation.

block_dim

Creates a block_dim operation.

block_id

Creates a block_id operation.

create_2_to_4_spmat

Creates a create_2to4_spmat operation.

create_coo

Creates a create_coo operation.

create_coo_aos

Creates a create_coo_aos operation.

create_csr

Creates a create_csr operation.

create_dn_tensor

Creates a create_dn_tensor operation.

dealloc

Creates a dealloc operation.

destroy_dn_tensor

Creates a destroy_dn_tensor operation.

destroy_sp_mat

Creates a destroy_sp_mat operation.

func

Creates a func operation.

global_id

Creates a global_id operation.

grid_dim

Creates a grid_dim operation.

host_register

Creates a host_register operation.

host_unregister

Creates a host_unregister operation.

lane_id

Creates a lane_id operation.

launch

Creates a launch operation.

launch_func

Creates a launch_func operation.

memcpy

Creates a memcpy operation.

memset

Creates a memset operation.

module

Creates a module operation.

module_end

Creates a module_end operation.

num_subgroups

Creates a num_subgroups operation.

printf

Creates a printf operation.

return

Creates a return operation.

sddmm

Creates a sddmm operation.

sddmm_buffer_size

Creates a sddmm_buffer_size operation.

set_default_device

Creates a set_default_device operation.

shuffle

Creates a shuffle operation.

spmm

Creates a spmm operation.

spmm_buffer_size

Creates a spmm_buffer_size operation.

spmv

Creates a spmv operation.

spmv_buffer_size

Creates a spmv_buffer_size operation.

subgroup_id

Creates a subgroup_id operation.

subgroup_mma_compute

Creates a subgroup_mma_compute operation.

subgroup_mma_constant_matrix

Creates a subgroup_mma_constant_matrix operation.

subgroup_mma_elementwise

Creates a subgroup_mma_elementwise operation.

subgroup_mma_load_matrix

Creates a subgroup_mma_load_matrix operation.

subgroup_mma_store_matrix

Creates a subgroup_mma_store_matrix operation.

subgroup_reduce

Creates a subgroup_reduce operation.

subgroup_size

Creates a subgroup_size operation.

terminator

Creates a terminator operation.

thread_id

Creates a thread_id operation.

wait

Creates a wait operation.

yield

Creates a yield operation.