Struct melior::dialect::ods::gpu::SpMMOp

pub struct SpMMOp<'c> { /* private fields */ }

spmm operation: SpMM operation

The gpu.spmm operation performs the SpMM operation on the given sparse and dense matrix, and buffer. The operation expects handles returned by previous sparse operations to construct an environment and the operands for SpMM. The buffer must have been allocated on the device.

If the async keyword is present, the op is executed asynchronously (i.e. it does not block until the execution has finished on the device). In that case, it returns a !gpu.async.token in addition to the environment.

The matrix arguments can also be associated with one of the following operators: NON_TRANSPOSE, TRANSPOSE, CONJUGATE_TRANSPOSE. The default value is NON_TRANSPOSE.

Example:

%token = gpu.spmm async [%dep] %spmatA{TRANSPOSE}, %dnmatB{TRANSPOSE}, %dnmatC, %buffers : type($buffers) into f32

Implementations

impl<'c> SpMMOp<'c>

pub fn name() -> &'static str

pub fn operation(&self) -> &Operation<'c>

pub fn builder( context: &'c Context, location: Location<'c> ) -> SpMMOpBuilder<'c, Unset, Unset, Unset, Unset, Unset, Unset>

pub fn async_token( &self, context: &'c Context ) -> Result<OperationResult<'c, '_>, Error>

pub fn async_dependencies( &self, context: &'c Context ) -> Result<impl Iterator<Item = Value<'c, '_>>, Error>

pub fn spmat_a(&self, context: &'c Context) -> Result<Value<'c, '_>, Error>

pub fn dnmat_b(&self, context: &'c Context) -> Result<Value<'c, '_>, Error>

pub fn dnmat_c(&self, context: &'c Context) -> Result<Value<'c, '_>, Error>

pub fn buffers( &self, context: &'c Context ) -> Result<impl Iterator<Item = Value<'c, '_>>, Error>

pub fn mode_a(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_mode_a(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn mode_b(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_mode_b(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn compute_type( &self, context: &'c Context ) -> Result<TypeAttribute<'c>, Error>

pub fn set_compute_type( &mut self, context: &'c Context, value: TypeAttribute<'c> )