Module melior::dialect::ods::index

index dialect.

The Index dialect contains operations for manipulating values of the builtin index type. The index type models target-specific values of pointer width, like intptr_t. Index values are typically used as loop bounds, array subscripts, tensor dimensions, etc.

The operations in this dialect operate exclusively on scalar index types. The dialect and its operations treat the index type as signless and contains signed and unsigned versions of certain operations where the distinction is meaningful. In particular, the operations and transformations are careful to be aware of the target-independent-ness of the index type, such as when folding.

The folding semantics of the Index dialect operations ensure that folding produces the same results irrespective of the eventual target pointer width. All index constants are stored in APInts of maximum index bitwidth: 64. Operations are folded using 64-bit integer arithmetic.

For operations where the values of the upper 32 bits don’t impact the values of the lower 32 bits, no additional handling is required because if the target is 32-bit, the truncated folded result will be the same as if the operation were computed with 32-bit arithmetic, and if the target is 64-bit, the fold result is valid by default.

Consider addition: an overflow in 32-bit is the same as truncating the result computed in 64-bit. For example, add(0x800000008, 0x800000008) is 0x1000000010 in 64-bit, which truncates to 0x10, the same result as truncating the operands first: add(0x08, 0x08). Specifically, an operation f can always be folded if it satisfies the following for all 64-bit values of a and b:

trunc(f(a, b)) = f(trunc(a), trunc(b))

When materializing target-specific code, constants just need to be truncated as appropriate.

Operations where the values of the upper 32 bits do impact the values of the lower 32 bits are not folded if the results would be different in 32-bit. These are operations that right shift – division, remainder, etc. These operations are only folded for subsets of a and b for which the above property is satisfied. This is checked per fold attempt.

Consider division: the 32-bit computation will differ from 64-bit if the latter results in a high bit shifted into the lower 32 bits. For example, div(0x100000002, 2) is 0x80000001 in 64-bit but 0x01 in 32-bit; it cannot be folded. However, div(0x200000002, 2) can be folded. The 64-bit result is 0x100000001, which truncated to 32 bits is 0x01. The 32-bit result of the operation with truncated operands div(0x02, 2) which is 0x01, the same as truncating the 64-bit result.

Structs

AddOp

add operation. Index addition.

AddOpBuilder

A builder for add operation. Index addition.

AndOp

and operation. Index bitwise and.

AndOpBuilder

A builder for and operation. Index bitwise and.

BoolConstantOp

bool.constant operation. Boolean constant.

BoolConstantOpBuilder

A builder for bool.constant operation. Boolean constant.

CastSOp

casts operation. Index signed cast.

CastSOpBuilder

A builder for casts operation. Index signed cast.

CastUOp

castu operation. Index unsigned cast.

CastUOpBuilder

A builder for castu operation. Index unsigned cast.

CeilDivSOp

ceildivs operation. Index signed ceil division.

CeilDivSOpBuilder

A builder for ceildivs operation. Index signed ceil division.

CeilDivUOp

ceildivu operation. Index unsigned ceil division.

CeilDivUOpBuilder

A builder for ceildivu operation. Index unsigned ceil division.

CmpOp

cmp operation. Index compare.

CmpOpBuilder

A builder for cmp operation. Index compare.

ConstantOp

constant operation. Index constant.

ConstantOpBuilder

A builder for constant operation. Index constant.

DivSOp

divs operation. Index signed division.

DivSOpBuilder

A builder for divs operation. Index signed division.

DivUOp

divu operation. Index unsigned division.

DivUOpBuilder

A builder for divu operation. Index unsigned division.

FloorDivSOp

floordivs operation. Index signed floor division.

FloorDivSOpBuilder

A builder for floordivs operation. Index signed floor division.

MaxSOp

maxs operation. Index signed maximum.

MaxSOpBuilder

A builder for maxs operation. Index signed maximum.

MaxUOp

maxu operation. Index unsigned maximum.

MaxUOpBuilder

A builder for maxu operation. Index unsigned maximum.

MinSOp

mins operation. Index signed minimum.

MinSOpBuilder

A builder for mins operation. Index signed minimum.

MinUOp

minu operation. Index unsigned minimum.

MinUOpBuilder

A builder for minu operation. Index unsigned minimum.

MulOp

mul operation. Index multiplication.

MulOpBuilder

A builder for mul operation. Index multiplication.

OrOp

or operation. Index bitwise or.

OrOpBuilder

A builder for or operation. Index bitwise or.

RemSOp

rems operation. Index signed remainder.

RemSOpBuilder

A builder for rems operation. Index signed remainder.

RemUOp

remu operation. Index unsigned remainder.

RemUOpBuilder

A builder for remu operation. Index unsigned remainder.

ShlOp

shl operation. Index shift left.

ShlOpBuilder

A builder for shl operation. Index shift left.

ShrSOp

shrs operation. Signed index shift right.

ShrSOpBuilder

A builder for shrs operation. Signed index shift right.

ShrUOp

shru operation. Unsigned index shift right.

ShrUOpBuilder

A builder for shru operation. Unsigned index shift right.

SizeOfOp

sizeof operation. Size in bits of the index type.

SizeOfOpBuilder

A builder for sizeof operation. Size in bits of the index type.

SubOp

sub operation. Index subtraction.

SubOpBuilder

A builder for sub operation. Index subtraction.

XOrOp

xor operation. Index bitwise xor.

XOrOpBuilder

A builder for xor operation. Index bitwise xor.

Functions

add

Creates a new add operation. Index addition.

and

Creates a new and operation. Index bitwise and.

bool_constant

Creates a new bool.constant operation. Boolean constant.

casts

Creates a new casts operation. Index signed cast.

castu

Creates a new castu operation. Index unsigned cast.

ceildivs

Creates a new ceildivs operation. Index signed ceil division.

ceildivu

Creates a new ceildivu operation. Index unsigned ceil division.

cmp

Creates a new cmp operation. Index compare.

constant

Creates a new constant operation. Index constant.

divs

Creates a new divs operation. Index signed division.

divu

Creates a new divu operation. Index unsigned division.

floordivs

Creates a new floordivs operation. Index signed floor division.

maxs

Creates a new maxs operation. Index signed maximum.

maxu

Creates a new maxu operation. Index unsigned maximum.

mins

Creates a new mins operation. Index signed minimum.

minu

Creates a new minu operation. Index unsigned minimum.

mul

Creates a new mul operation. Index multiplication.

or

Creates a new or operation. Index bitwise or.

rems

Creates a new rems operation. Index signed remainder.

remu

Creates a new remu operation. Index unsigned remainder.

shl

Creates a new shl operation. Index shift left.

shrs

Creates a new shrs operation. Signed index shift right.

shru

Creates a new shru operation. Unsigned index shift right.

sizeof

Creates a new sizeof operation. Size in bits of the index type.

sub

Creates a new sub operation. Index subtraction.

xor

Creates a new xor operation. Index bitwise xor.