Module melior::dialect::ods::linalg

linalg dialect.

The linalg dialect groups together a set of types, operations and transformations that are useful to implement a structured abstraction on buffers and tensors. These abstractions are useful for transformations and can lower to scalar load/store and other operations or to more general library calls.

Additional Linalg Dialect Documentation and a Rationale Document are are also available and should be read first before going in the details of the op semantics.

Structs

IndexOp

index operation: Linalg index operation

IndexOpBuilder

Builder for index operation: Linalg index operation

YieldOp

yield operation: Linalg yield operation

YieldOpBuilder

Builder for yield operation: Linalg yield operation

Functions

index

Creates a new index operation: Linalg index operation

yield

Creates a new yield operation: Linalg yield operation