Module melior::dialect::ods::llvm

llvm dialect.

Structs

AShrOperation

ashr operation.

AShrOperationBuilder

A builder for ashr operation.

AddOperation

add operation.

AddOperationBuilder

A builder for add operation.

AddrSpaceCastOperation

addrspacecast operation.

AddrSpaceCastOperationBuilder

A builder for addrspacecast operation.

AddressOfOperation

mlir.addressof operation. Creates a pointer pointing to a global or a function.

AddressOfOperationBuilder

A builder for mlir.addressof operation. Creates a pointer pointing to a global or a function.

AllocaOperation

alloca operation.

AllocaOperationBuilder

A builder for alloca operation.

AndOperation

and operation.

AndOperationBuilder

A builder for and operation.

AtomicCmpXchgOperation

cmpxchg operation.

AtomicCmpXchgOperationBuilder

A builder for cmpxchg operation.

AtomicRMWOperation

atomicrmw operation.

AtomicRMWOperationBuilder

A builder for atomicrmw operation.

BitcastOperation

bitcast operation.

BitcastOperationBuilder

A builder for bitcast operation.

BrOperation

br operation.

BrOperationBuilder

A builder for br operation.

CallOperation

call operation. Call to an LLVM function..

CallOperationBuilder

A builder for call operation. Call to an LLVM function..

ComdatOperation

comdat operation. LLVM dialect comdat region.

ComdatOperationBuilder

A builder for comdat operation. LLVM dialect comdat region.

ComdatSelectorOperation

comdat_selector operation. LLVM dialect comdat selector declaration.

ComdatSelectorOperationBuilder

A builder for comdat_selector operation. LLVM dialect comdat selector declaration.

CondBrOperation

cond_br operation.

CondBrOperationBuilder

A builder for cond_br operation.

ConstantOperation

mlir.constant operation. Defines a constant of LLVM type..

ConstantOperationBuilder

A builder for mlir.constant operation. Defines a constant of LLVM type..

ExtractElementOperation

extractelement operation. Extract an element from an LLVM vector..

ExtractElementOperationBuilder

A builder for extractelement operation. Extract an element from an LLVM vector..

ExtractValueOperation

extractvalue operation. Extract a value from an LLVM struct..

ExtractValueOperationBuilder

A builder for extractvalue operation. Extract a value from an LLVM struct..

FAddOperation

fadd operation.

FAddOperationBuilder

A builder for fadd operation.

FCmpOperation

fcmp operation.

FCmpOperationBuilder

A builder for fcmp operation.

FDivOperation

fdiv operation.

FDivOperationBuilder

A builder for fdiv operation.

FMulOperation

fmul operation.

FMulOperationBuilder

A builder for fmul operation.

FNegOperation

fneg operation.

FNegOperationBuilder

A builder for fneg operation.

FPExtOperation

fpext operation.

FPExtOperationBuilder

A builder for fpext operation.

FPToSIOperation

fptosi operation.

FPToSIOperationBuilder

A builder for fptosi operation.

FPToUIOperation

fptoui operation.

FPToUIOperationBuilder

A builder for fptoui operation.

FPTruncOperation

fptrunc operation.

FPTruncOperationBuilder

A builder for fptrunc operation.

FRemOperation

frem operation.

FRemOperationBuilder

A builder for frem operation.

FSubOperation

fsub operation.

FSubOperationBuilder

A builder for fsub operation.

FenceOperation

fence operation.

FenceOperationBuilder

A builder for fence operation.

FreezeOperation

freeze operation.

FreezeOperationBuilder

A builder for freeze operation.

GEPOperation

getelementptr operation.

GEPOperationBuilder

A builder for getelementptr operation.

GlobalCtorsOperation

mlir.global_ctors operation. LLVM dialect global_ctors..

GlobalCtorsOperationBuilder

A builder for mlir.global_ctors operation. LLVM dialect global_ctors..

GlobalDtorsOperation

mlir.global_dtors operation. LLVM dialect global_dtors..

GlobalDtorsOperationBuilder

A builder for mlir.global_dtors operation. LLVM dialect global_dtors..

GlobalOperation

mlir.global operation. LLVM dialect global..

GlobalOperationBuilder

A builder for mlir.global operation. LLVM dialect global..

ICmpOperation

icmp operation.

ICmpOperationBuilder

A builder for icmp operation.

InlineAsmOperation

inline_asm operation.

InlineAsmOperationBuilder

A builder for inline_asm operation.

InsertElementOperation

insertelement operation. Insert an element into an LLVM vector..

InsertElementOperationBuilder

A builder for insertelement operation. Insert an element into an LLVM vector..

InsertValueOperation

insertvalue operation. Insert a value into an LLVM struct..

InsertValueOperationBuilder

A builder for insertvalue operation. Insert a value into an LLVM struct..

IntToPtrOperation

inttoptr operation.

IntToPtrOperationBuilder

A builder for inttoptr operation.

InvokeOperation

invoke operation.

InvokeOperationBuilder

A builder for invoke operation.

LLVMFuncOperation

func operation. LLVM dialect function..

LLVMFuncOperationBuilder

A builder for func operation. LLVM dialect function..

LShrOperation

lshr operation.

LShrOperationBuilder

A builder for lshr operation.

LandingpadOperation

landingpad operation.

LandingpadOperationBuilder

A builder for landingpad operation.

LoadOperation

load operation.

LoadOperationBuilder

A builder for load operation.

MetadataOperation

metadata operation. LLVM dialect metadata..

MetadataOperationBuilder

A builder for metadata operation. LLVM dialect metadata..

MulOperation

mul operation.

MulOperationBuilder

A builder for mul operation.

NullOperation

mlir.null operation. Defines a value containing a null pointer to LLVM type..

NullOperationBuilder

A builder for mlir.null operation. Defines a value containing a null pointer to LLVM type..

OrOperation

or operation.

OrOperationBuilder

A builder for or operation.

PoisonOperation

mlir.poison operation. Creates a poison value of LLVM dialect type..

PoisonOperationBuilder

A builder for mlir.poison operation. Creates a poison value of LLVM dialect type..

PtrToIntOperation

ptrtoint operation.

PtrToIntOperationBuilder

A builder for ptrtoint operation.

ResumeOperation

resume operation.

ResumeOperationBuilder

A builder for resume operation.

ReturnOperation

return operation.

ReturnOperationBuilder

A builder for return operation.

SDivOperation

sdiv operation.

SDivOperationBuilder

A builder for sdiv operation.

SExtOperation

sext operation.

SExtOperationBuilder

A builder for sext operation.

SIToFPOperation

sitofp operation.

SIToFPOperationBuilder

A builder for sitofp operation.

SRemOperation

srem operation.

SRemOperationBuilder

A builder for srem operation.

SelectOperation

select operation.

SelectOperationBuilder

A builder for select operation.

ShlOperation

shl operation.

ShlOperationBuilder

A builder for shl operation.

ShuffleVectorOperation

shufflevector operation. Construct a permutation of two vectors..

ShuffleVectorOperationBuilder

A builder for shufflevector operation. Construct a permutation of two vectors..

StoreOperation

store operation.

StoreOperationBuilder

A builder for store operation.

SubOperation

sub operation.

SubOperationBuilder

A builder for sub operation.

SwitchOperation

switch operation.

SwitchOperationBuilder

A builder for switch operation.

TruncOperation

trunc operation.

TruncOperationBuilder

A builder for trunc operation.

UDivOperation

udiv operation.

UDivOperationBuilder

A builder for udiv operation.

UIToFPOperation

uitofp operation.

UIToFPOperationBuilder

A builder for uitofp operation.

URemOperation

urem operation.

URemOperationBuilder

A builder for urem operation.

UndefOperation

mlir.undef operation. Creates an undefined value of LLVM dialect type..

UndefOperationBuilder

A builder for mlir.undef operation. Creates an undefined value of LLVM dialect type..

UnreachableOperation

unreachable operation.

UnreachableOperationBuilder

A builder for unreachable operation.

XOrOperation

xor operation.

XOrOperationBuilder

A builder for xor operation.

ZExtOperation

zext operation.

ZExtOperationBuilder

A builder for zext operation.

Functions

add

Creates a add operation.

addrspacecast

Creates a addrspacecast operation.

alloca

Creates a alloca operation.

and

Creates a and operation.

ashr

Creates a ashr operation.

atomicrmw

Creates a atomicrmw operation.

bitcast

Creates a bitcast operation.

br

Creates a br operation.

call

Creates a call operation. Call to an LLVM function..

cmpxchg

Creates a cmpxchg operation.

comdat

Creates a comdat operation. LLVM dialect comdat region.

comdat_selector

Creates a comdat_selector operation. LLVM dialect comdat selector declaration.

cond_br

Creates a cond_br operation.

extractelement

Creates a extractelement operation. Extract an element from an LLVM vector..

extractvalue

Creates a extractvalue operation. Extract a value from an LLVM struct..

fadd

Creates a fadd operation.

fcmp

Creates a fcmp operation.

fdiv

Creates a fdiv operation.

fence

Creates a fence operation.

fmul

Creates a fmul operation.

fneg

Creates a fneg operation.

fpext

Creates a fpext operation.

fptosi

Creates a fptosi operation.

fptoui

Creates a fptoui operation.

fptrunc

Creates a fptrunc operation.

freeze

Creates a freeze operation.

frem

Creates a frem operation.

fsub

Creates a fsub operation.

func

Creates a func operation. LLVM dialect function..

getelementptr

Creates a getelementptr operation.

icmp

Creates a icmp operation.

inline_asm

Creates a inline_asm operation.

insertelement

Creates a insertelement operation. Insert an element into an LLVM vector..

insertvalue

Creates a insertvalue operation. Insert a value into an LLVM struct..

inttoptr

Creates a inttoptr operation.

invoke

Creates a invoke operation.

landingpad

Creates a landingpad operation.

load

Creates a load operation.

lshr

Creates a lshr operation.

metadata

Creates a metadata operation. LLVM dialect metadata..

mlir_addressof

Creates a mlir.addressof operation. Creates a pointer pointing to a global or a function.

mlir_constant

Creates a mlir.constant operation. Defines a constant of LLVM type..

mlir_global

Creates a mlir.global operation. LLVM dialect global..

mlir_global_ctors

Creates a mlir.global_ctors operation. LLVM dialect global_ctors..

mlir_global_dtors

Creates a mlir.global_dtors operation. LLVM dialect global_dtors..

mlir_null

Creates a mlir.null operation. Defines a value containing a null pointer to LLVM type..

mlir_poison

Creates a mlir.poison operation. Creates a poison value of LLVM dialect type..

mlir_undef

Creates a mlir.undef operation. Creates an undefined value of LLVM dialect type..

mul

Creates a mul operation.

or

Creates a or operation.

ptrtoint

Creates a ptrtoint operation.

resume

Creates a resume operation.

return

Creates a return operation.

sdiv

Creates a sdiv operation.

select

Creates a select operation.

sext

Creates a sext operation.

shl

Creates a shl operation.

shufflevector

Creates a shufflevector operation. Construct a permutation of two vectors..

sitofp

Creates a sitofp operation.

srem

Creates a srem operation.

store

Creates a store operation.

sub

Creates a sub operation.

switch

Creates a switch operation.

trunc

Creates a trunc operation.

udiv

Creates a udiv operation.

uitofp

Creates a uitofp operation.

unreachable

Creates a unreachable operation.

urem

Creates a urem operation.

xor

Creates a xor operation.

zext

Creates a zext operation.