Struct melior::dialect::ods::llvm::LLVMFuncOp

pub struct LLVMFuncOp<'c> { /* private fields */ }

func operation. LLVM dialect function..

MLIR functions are defined by an operation that is not built into the IR itself. The LLVM dialect provides an llvm.func operation to define functions compatible with LLVM IR. These functions have LLVM dialect function type but use MLIR syntax to express it. They are required to have exactly one result type. LLVM function operation is intended to capture additional properties of LLVM functions, such as linkage and calling convention, that may be modeled differently by the built-in MLIR function.

// The type of @bar is !llvm<"i64 (i64)">
llvm.func @bar(%arg0: i64) -> i64 {
  llvm.return %arg0 : i64
}

// Type type of @foo is !llvm<"void (i64)">
// !llvm.void type is omitted
llvm.func @foo(%arg0: i64) {
  llvm.return
}

// A function with `internal` linkage.
llvm.func internal @internal_func() {
  llvm.return
}

Implementations

impl<'c> LLVMFuncOp<'c>

pub fn name() -> &'static str

pub fn operation(&self) -> &Operation<'c>

pub fn builder( context: &'c Context, location: Location<'c> ) -> LLVMFuncOpBuilder<'c, Unset, Unset, Unset>

pub fn body(&self, context: &'c Context) -> Result<RegionRef<'c, '_>, Error>

pub fn sym_name( &self, context: &'c Context ) -> Result<StringAttribute<'c>, Error>

pub fn set_sym_name(&mut self, context: &'c Context, value: StringAttribute<'c>)

pub fn function_type( &self, context: &'c Context ) -> Result<TypeAttribute<'c>, Error>

pub fn set_function_type( &mut self, context: &'c Context, value: TypeAttribute<'c> )

pub fn linkage(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_linkage(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn dso_local(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_dso_local(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn remove_dso_local(&mut self, context: &'c Context) -> Result<(), Error>

pub fn c_conv(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_c_conv(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn comdat(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_comdat(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn remove_comdat(&mut self, context: &'c Context) -> Result<(), Error>

pub fn personality( &self, context: &'c Context ) -> Result<FlatSymbolRefAttribute<'c>, Error>

pub fn set_personality( &mut self, context: &'c Context, value: FlatSymbolRefAttribute<'c> )

pub fn remove_personality(&mut self, context: &'c Context) -> Result<(), Error>

pub fn garbage_collector( &self, context: &'c Context ) -> Result<StringAttribute<'c>, Error>

pub fn set_garbage_collector( &mut self, context: &'c Context, value: StringAttribute<'c> )

pub fn remove_garbage_collector( &mut self, context: &'c Context ) -> Result<(), Error>

pub fn passthrough( &self, context: &'c Context ) -> Result<ArrayAttribute<'c>, Error>

pub fn set_passthrough( &mut self, context: &'c Context, value: ArrayAttribute<'c> )

pub fn remove_passthrough(&mut self, context: &'c Context) -> Result<(), Error>

pub fn arg_attrs( &self, context: &'c Context ) -> Result<ArrayAttribute<'c>, Error>

pub fn set_arg_attrs(&mut self, context: &'c Context, value: ArrayAttribute<'c>)

pub fn remove_arg_attrs(&mut self, context: &'c Context) -> Result<(), Error>

pub fn res_attrs( &self, context: &'c Context ) -> Result<ArrayAttribute<'c>, Error>

pub fn set_res_attrs(&mut self, context: &'c Context, value: ArrayAttribute<'c>)

pub fn remove_res_attrs(&mut self, context: &'c Context) -> Result<(), Error>

pub fn function_entry_count( &self, context: &'c Context ) -> Result<IntegerAttribute<'c>, Error>

pub fn set_function_entry_count( &mut self, context: &'c Context, value: IntegerAttribute<'c> )

pub fn remove_function_entry_count( &mut self, context: &'c Context ) -> Result<(), Error>

pub fn memory(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_memory(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn remove_memory(&mut self, context: &'c Context) -> Result<(), Error>

pub fn visibility(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_visibility(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn arm_streaming( &self, context: &'c Context ) -> Result<Attribute<'c>, Error>

pub fn set_arm_streaming(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn remove_arm_streaming( &mut self, context: &'c Context ) -> Result<(), Error>

pub fn arm_locally_streaming( &self, context: &'c Context ) -> Result<Attribute<'c>, Error>

pub fn set_arm_locally_streaming( &mut self, context: &'c Context, value: Attribute<'c> )

pub fn remove_arm_locally_streaming( &mut self, context: &'c Context ) -> Result<(), Error>

pub fn section( &self, context: &'c Context ) -> Result<StringAttribute<'c>, Error>

pub fn set_section(&mut self, context: &'c Context, value: StringAttribute<'c>)

pub fn remove_section(&mut self, context: &'c Context) -> Result<(), Error>

pub fn unnamed_addr(&self, context: &'c Context) -> Result<Attribute<'c>, Error>

pub fn set_unnamed_addr(&mut self, context: &'c Context, value: Attribute<'c>)

pub fn remove_unnamed_addr(&mut self, context: &'c Context) -> Result<(), Error>

pub fn alignment( &self, context: &'c Context ) -> Result<IntegerAttribute<'c>, Error>

pub fn set_alignment( &mut self, context: &'c Context, value: IntegerAttribute<'c> )

pub fn remove_alignment(&mut self, context: &'c Context) -> Result<(), Error>