Module melior::dialect::ods::math

math dialect.

The math dialect is intended to hold mathematical operations on integer and floating types beyond simple arithmetics. Each operation works on scalar, vector or tensor type. On vector and tensor type operations apply elementwise unless explicitly specified otherwise. As an example, the floating point absolute value can be expressed as:

// Scalar absolute value.
%a = math.absf %b : f64

// Vector elementwise absolute value.
%f = math.absf %g : vector<4xf32>

// Tensor elementwise absolute value.
%x = math.absf %y : tensor<4x?xf8>

Structs

AbsFOperation

An absf operation. Floating point absolute-value operation.

AbsFOperationBuilder

A builder for an absf operation.

AbsIOperation

An absi operation. Integer absolute-value operation.

AbsIOperationBuilder

A builder for an absi operation.

Atan2Operation

An atan2 operation. 2-argument arcus tangent of the given values.

Atan2OperationBuilder

A builder for an atan2 operation.

AtanOperation

An atan operation. Arcus tangent of the given value.

AtanOperationBuilder

A builder for an atan operation.

CbrtOperation

A cbrt operation. Cube root of the specified value.

CbrtOperationBuilder

A builder for a cbrt operation.

CeilOperation

A ceil operation. Ceiling of the specified value.

CeilOperationBuilder

A builder for a ceil operation.

CopySignOperation

A copysign operation. A copysign operation.

CopySignOperationBuilder

A builder for a copysign operation.

CosOperation

A cos operation. Cosine of the specified value.

CosOperationBuilder

A builder for a cos operation.

CountLeadingZerosOperation

A ctlz operation. Counts the leading zeros an integer value.

CountLeadingZerosOperationBuilder

A builder for a ctlz operation.

CountTrailingZerosOperation

A cttz operation. Counts the trailing zeros an integer value.

CountTrailingZerosOperationBuilder

A builder for a cttz operation.

CtPopOperation

A ctpop operation. Counts the number of set bits of an integer value.

CtPopOperationBuilder

A builder for a ctpop operation.

ErfOperation

An erf operation. Error function of the specified value.

ErfOperationBuilder

A builder for an erf operation.

Exp2Operation

An exp2 operation. Base-2 exponential of the specified value.

Exp2OperationBuilder

A builder for an exp2 operation.

ExpM1Operation

An expm1 operation. Base-e exponential of the specified value minus 1.

ExpM1OperationBuilder

A builder for an expm1 operation.

ExpOperation

An exp operation. Base-e exponential of the specified value.

ExpOperationBuilder

A builder for an exp operation.

FPowIOperation

A fpowi operation. Floating point raised to the signed integer power.

FPowIOperationBuilder

A builder for a fpowi operation.

FloorOperation

A floor operation. Floor of the specified value.

FloorOperationBuilder

A builder for a floor operation.

FmaOperation

A fma operation. Floating point fused multipy-add operation.

FmaOperationBuilder

A builder for a fma operation.

IPowIOperation

An ipowi operation. Signed integer raised to the power of operation.

IPowIOperationBuilder

A builder for an ipowi operation.

Log1pOperation

A log1p operation. Computes the natural logarithm of one plus the given value.

Log1pOperationBuilder

A builder for a log1p operation.

Log2Operation

A log2 operation. Base-2 logarithm of the specified value.

Log2OperationBuilder

A builder for a log2 operation.

Log10Operation

A log10 operation. Base-10 logarithm of the specified value.

Log10OperationBuilder

A builder for a log10 operation.

LogOperation

A log operation. Base-e logarithm of the specified value.

LogOperationBuilder

A builder for a log operation.

PowFOperation

A powf operation. Floating point raised to the power of operation.

PowFOperationBuilder

A builder for a powf operation.

RoundEvenOperation

A roundeven operation. Round of the specified value with halfway cases to even.

RoundEvenOperationBuilder

A builder for a roundeven operation.

RoundOperation

A round operation. Round of the specified value.

RoundOperationBuilder

A builder for a round operation.

RsqrtOperation

A rsqrt operation. Reciprocal of sqrt (1 / sqrt of the specified value).

RsqrtOperationBuilder

A builder for a rsqrt operation.

SinOperation

A sin operation. Sine of the specified value.

SinOperationBuilder

A builder for a sin operation.

SqrtOperation

A sqrt operation. Sqrt of the specified value.

SqrtOperationBuilder

A builder for a sqrt operation.

TanOperation

A tan operation. Tangent of the specified value.

TanOperationBuilder

A builder for a tan operation.

TanhOperation

A tanh operation. Hyperbolic tangent of the specified value.

TanhOperationBuilder

A builder for a tanh operation.

TruncOperation

A trunc operation. Trunc of the specified value.

TruncOperationBuilder

A builder for a trunc operation.

Functions

absf

Creates an absf operation.

absi

Creates an absi operation.

atan

Creates an atan operation.

atan_2

Creates an atan2 operation.

cbrt

Creates a cbrt operation.

ceil

Creates a ceil operation.

copysign

Creates a copysign operation.

cos

Creates a cos operation.

ctlz

Creates a ctlz operation.

ctpop

Creates a ctpop operation.

cttz

Creates a cttz operation.

erf

Creates an erf operation.

exp

Creates an exp operation.

exp_2

Creates an exp2 operation.

expm_1

Creates an expm1 operation.

floor

Creates a floor operation.

fma

Creates a fma operation.

fpowi

Creates a fpowi operation.

ipowi

Creates an ipowi operation.

log

Creates a log operation.

log_1_p

Creates a log1p operation.

log_2

Creates a log2 operation.

log_10

Creates a log10 operation.

powf

Creates a powf operation.

round

Creates a round operation.

roundeven

Creates a roundeven operation.

rsqrt

Creates a rsqrt operation.

sin

Creates a sin operation.

sqrt

Creates a sqrt operation.

tan

Creates a tan operation.

tanh

Creates a tanh operation.

trunc

Creates a trunc operation.