Module melior::dialect::ods::pdl

pdl dialect.

PDL presents a high level abstraction for the rewrite pattern infrastructure available in MLIR. This abstraction allows for representing patterns transforming MLIR, as MLIR. This allows for applying all of the benefits that the general MLIR infrastructure provides, to the infrastructure itself. This means that pattern matching can be more easily verified for correctness, targeted by frontends, and optimized.

PDL abstracts over various different aspects of patterns and core MLIR data structures. Patterns are specified via a pdl.pattern operation. These operations contain a region body for the “matcher” code, and terminate with a pdl.rewrite that either dispatches to an external rewriter or contains a region for the rewrite specified via pdl. The types of values in pdl are handle types to MLIR C++ types, with !pdl.attribute, !pdl.operation, !pdl.value, and !pdl.type directly mapping to mlir::Attribute, mlir::Operation*, mlir::Value, and mlir::Type respectively.

An example pattern is shown below:

// pdl.pattern contains metadata similarly to a `RewritePattern`.
pdl.pattern : benefit(1) {
  // External input operand values are specified via `pdl.operand` operations.
  // Result types are constrainted via `pdl.type` operations.

  %resultType = pdl.type
  %inputOperand = pdl.operand
  %root = pdl.operation "foo.op"(%inputOperand) -> %resultType
  pdl.rewrite %root {
    pdl.replace %root with (%inputOperand)
  }
}

The above pattern simply replaces an operation with its first operand. Note how the input operation is specified structurally, similarly to how it would look in memory. This is a simple example and pdl provides support for many other features such as applying external constraints or external generator methods. These features and more are detailed below.

Structs

ApplyNativeConstraintOp

apply_native_constraint operation. Apply a native constraint to a set of provided entities.

ApplyNativeConstraintOpBuilder

Builder for apply_native_constraint operation. Apply a native constraint to a set of provided entities.

ApplyNativeRewriteOp

apply_native_rewrite operation. Apply a native rewrite method inside of pdl.rewrite region.

ApplyNativeRewriteOpBuilder

Builder for apply_native_rewrite operation. Apply a native rewrite method inside of pdl.rewrite region.

AttributeOp

attribute operation. Define an input attribute in a pattern.

AttributeOpBuilder

Builder for attribute operation. Define an input attribute in a pattern.

EraseOp

erase operation. Mark an input operation as erased.

EraseOpBuilder

Builder for erase operation. Mark an input operation as erased.

OperandOp

operand operation. Define an external input operand in a pattern.

OperandOpBuilder

Builder for operand operation. Define an external input operand in a pattern.

OperandsOp

operands operation. Define a range of input operands in a pattern.

OperandsOpBuilder

Builder for operands operation. Define a range of input operands in a pattern.

OperationOp

operation operation. Define an operation within a pattern.

OperationOpBuilder

Builder for operation operation. Define an operation within a pattern.

PatternOp

pattern operation. Define a rewrite pattern.

PatternOpBuilder

Builder for pattern operation. Define a rewrite pattern.

RangeOp

range operation. Construct a range of pdl entities.

RangeOpBuilder

Builder for range operation. Construct a range of pdl entities.

ReplaceOp

replace operation. Mark an input operation as replaced.

ReplaceOpBuilder

Builder for replace operation. Mark an input operation as replaced.

ResultOp

result operation. Extract a result from an operation.

ResultOpBuilder

Builder for result operation. Extract a result from an operation.

ResultsOp

results operation. Extract a result group from an operation.

ResultsOpBuilder

Builder for results operation. Extract a result group from an operation.

RewriteOp

rewrite operation. Specify the rewrite of a matched pattern.

RewriteOpBuilder

Builder for rewrite operation. Specify the rewrite of a matched pattern.

TypeOp

type operation. Define a type handle within a pattern.

TypeOpBuilder

Builder for type operation. Define a type handle within a pattern.

TypesOp

types operation. Define a range of type handles within a pattern.

TypesOpBuilder

Builder for types operation. Define a range of type handles within a pattern.

Functions

_operation

Creates a new operation operation. Define an operation within a pattern.

apply_native_constraint

Creates a new apply_native_constraint operation. Apply a native constraint to a set of provided entities.

apply_native_rewrite

Creates a new apply_native_rewrite operation. Apply a native rewrite method inside of pdl.rewrite region.

attribute

Creates a new attribute operation. Define an input attribute in a pattern.

erase

Creates a new erase operation. Mark an input operation as erased.

operand

Creates a new operand operation. Define an external input operand in a pattern.

operands

Creates a new operands operation. Define a range of input operands in a pattern.

pattern

Creates a new pattern operation. Define a rewrite pattern.

range

Creates a new range operation. Construct a range of pdl entities.

replace

Creates a new replace operation. Mark an input operation as replaced.

result

Creates a new result operation. Extract a result from an operation.

results

Creates a new results operation. Extract a result group from an operation.

rewrite

Creates a new rewrite operation. Specify the rewrite of a matched pattern.

type

Creates a new type operation. Define a type handle within a pattern.

types

Creates a new types operation. Define a range of type handles within a pattern.