Module melior::dialect::ods::pdl_interp

pdl_interp dialect.

The PDL Interpreter dialect provides a lower level abstraction compared to the PDL dialect, and is targeted towards low level optimization and interpreter code generation. The dialect operations encapsulates low-level pattern match and rewrite “primitives”, such as navigating the IR (Operation::getOperand), creating new operations (OpBuilder::create), etc. Many of the operations within this dialect also fuse branching control flow with some form of a predicate comparison operation. This type of fusion reduces the amount of work that an interpreter must do when executing.

Structs

ApplyConstraintOperation

A(n) apply_constraint operation. Apply a constraint to a set of positional values.

ApplyConstraintOperationBuilder

A builder for a(n) apply_constraint operation.

ApplyRewriteOperation

A(n) apply_rewrite operation. Invoke and apply an externally registered rewrite method.

ApplyRewriteOperationBuilder

A builder for a(n) apply_rewrite operation.

AreEqualOperation

A(n) are_equal operation. Check if two positional values or ranges are equivalent.

AreEqualOperationBuilder

A builder for a(n) are_equal operation.

BranchOperation

A(n) branch operation. General branch operation.

BranchOperationBuilder

A builder for a(n) branch operation.

CheckAttributeOperation

A(n) check_attribute operation. Check the value of an Attribute.

CheckAttributeOperationBuilder

A builder for a(n) check_attribute operation.

CheckOperandCountOperation

A(n) check_operand_count operation. Check the number of operands of an Operation.

CheckOperandCountOperationBuilder

A builder for a(n) check_operand_count operation.

CheckOperationNameOperation

A(n) check_operation_name operation. Check the OperationName of an Operation.

CheckOperationNameOperationBuilder

A builder for a(n) check_operation_name operation.

CheckResultCountOperation

A(n) check_result_count operation. Check the number of results of an Operation.

CheckResultCountOperationBuilder

A builder for a(n) check_result_count operation.

CheckTypeOperation

A(n) check_type operation. Compare a type to a known value.

CheckTypeOperationBuilder

A builder for a(n) check_type operation.

CheckTypesOperation

A(n) check_types operation. Compare a range of types to a range of known values.

CheckTypesOperationBuilder

A builder for a(n) check_types operation.

ContinueOperation

A(n) continue operation. Breaks the current iteration.

ContinueOperationBuilder

A builder for a(n) continue operation.

CreateAttributeOperation

A(n) create_attribute operation. Create an interpreter handle to a constant Attribute.

CreateAttributeOperationBuilder

A builder for a(n) create_attribute operation.

CreateOperationOperation

A(n) create_operation operation. Create an instance of a specific Operation.

CreateOperationOperationBuilder

A builder for a(n) create_operation operation.

CreateRangeOperation

A(n) create_range operation. Construct a range of PDL entities.

CreateRangeOperationBuilder

A builder for a(n) create_range operation.

CreateTypeOperation

A(n) create_type operation. Create an interpreter handle to a constant Type.

CreateTypeOperationBuilder

A builder for a(n) create_type operation.

CreateTypesOperation

A(n) create_types operation. Create an interpreter handle to a range of constant Types.

CreateTypesOperationBuilder

A builder for a(n) create_types operation.

EraseOperation

A(n) erase operation. Mark an operation as erased.

EraseOperationBuilder

A builder for a(n) erase operation.

ExtractOperation

A(n) extract operation. Extract the item at the specified index in a range.

ExtractOperationBuilder

A builder for a(n) extract operation.

FinalizeOperation

A(n) finalize operation. Finalize a pattern match or rewrite sequence.

FinalizeOperationBuilder

A builder for a(n) finalize operation.

ForEachOperation

A(n) foreach operation. Iterates over a range of values or ranges.

ForEachOperationBuilder

A builder for a(n) foreach operation.

FuncOperation

A(n) func operation. PDL Interpreter Function Operation.

FuncOperationBuilder

A builder for a(n) func operation.

GetAttributeOperation

A(n) get_attribute operation. Get a specified attribute value from an Operation.

GetAttributeOperationBuilder

A builder for a(n) get_attribute operation.

GetAttributeTypeOperation

A(n) get_attribute_type operation. Get the result type of a specified Attribute.

GetAttributeTypeOperationBuilder

A builder for a(n) get_attribute_type operation.

GetDefiningOperation

A(n) get_defining_op operation. Get the defining operation of a Value.

GetDefiningOperationBuilder

A builder for a(n) get_defining_op operation.

GetOperandOperation

A(n) get_operand operation. Get a specified operand from an Operation.

GetOperandOperationBuilder

A builder for a(n) get_operand operation.

GetOperandsOperation

A(n) get_operands operation. Get a specified operand group from an Operation.

GetOperandsOperationBuilder

A builder for a(n) get_operands operation.

GetResultOperation

A(n) get_result operation. Get a specified result from an Operation.

GetResultOperationBuilder

A builder for a(n) get_result operation.

GetResultsOperation

A(n) get_results operation. Get a specified result group from an Operation.

GetResultsOperationBuilder

A builder for a(n) get_results operation.

GetUsersOperation

A(n) get_users operation. Get the users of a Value.

GetUsersOperationBuilder

A builder for a(n) get_users operation.

GetValueTypeOperation

A(n) get_value_type operation. Get the result type of a specified Value.

GetValueTypeOperationBuilder

A builder for a(n) get_value_type operation.

IsNotNullOperation

A(n) is_not_null operation. Check if a positional value is non-null.

IsNotNullOperationBuilder

A builder for a(n) is_not_null operation.

RecordMatchOperation

A(n) record_match operation. Record the metadata for a successful pattern match.

RecordMatchOperationBuilder

A builder for a(n) record_match operation.

ReplaceOperation

A(n) replace operation. Mark an operation as replaced.

ReplaceOperationBuilder

A builder for a(n) replace operation.

SwitchAttributeOperation

A(n) switch_attribute operation. Switch on the value of an Attribute.

SwitchAttributeOperationBuilder

A builder for a(n) switch_attribute operation.

SwitchOperandCountOperation

A(n) switch_operand_count operation. Switch on the operand count of an Operation.

SwitchOperandCountOperationBuilder

A builder for a(n) switch_operand_count operation.

SwitchOperationNameOperation

A(n) switch_operation_name operation. Switch on the OperationName of an Operation.

SwitchOperationNameOperationBuilder

A builder for a(n) switch_operation_name operation.

SwitchResultCountOperation

A(n) switch_result_count operation. Switch on the result count of an Operation.

SwitchResultCountOperationBuilder

A builder for a(n) switch_result_count operation.

SwitchTypeOperation

A(n) switch_type operation. Switch on a Type value.

SwitchTypeOperationBuilder

A builder for a(n) switch_type operation.

SwitchTypesOperation

A(n) switch_types operation. Switch on a range of Type values.

SwitchTypesOperationBuilder

A builder for a(n) switch_types operation.

Functions

apply_constraint

Creates a(n) apply_constraint operation.

apply_rewrite

Creates a(n) apply_rewrite operation.

are_equal

Creates a(n) are_equal operation.

branch

Creates a(n) branch operation.

check_attribute

Creates a(n) check_attribute operation.

check_operand_count

Creates a(n) check_operand_count operation.

check_operation_name

Creates a(n) check_operation_name operation.

check_result_count

Creates a(n) check_result_count operation.

check_type

Creates a(n) check_type operation.

check_types

Creates a(n) check_types operation.

continue

Creates a(n) continue operation.

create_attribute

Creates a(n) create_attribute operation.

create_operation

Creates a(n) create_operation operation.

create_range

Creates a(n) create_range operation.

create_type

Creates a(n) create_type operation.

create_types

Creates a(n) create_types operation.

erase

Creates a(n) erase operation.

extract

Creates a(n) extract operation.

finalize

Creates a(n) finalize operation.

foreach

Creates a(n) foreach operation.

func

Creates a(n) func operation.

get_attribute

Creates a(n) get_attribute operation.

get_attribute_type

Creates a(n) get_attribute_type operation.

get_defining_op

Creates a(n) get_defining_op operation.

get_operand

Creates a(n) get_operand operation.

get_operands

Creates a(n) get_operands operation.

get_result

Creates a(n) get_result operation.

get_results

Creates a(n) get_results operation.

get_users

Creates a(n) get_users operation.

get_value_type

Creates a(n) get_value_type operation.

is_not_null

Creates a(n) is_not_null operation.

record_match

Creates a(n) record_match operation.

replace

Creates a(n) replace operation.

switch_attribute

Creates a(n) switch_attribute operation.

switch_operand_count

Creates a(n) switch_operand_count operation.

switch_operation_name

Creates a(n) switch_operation_name operation.

switch_result_count

Creates a(n) switch_result_count operation.

switch_type

Creates a(n) switch_type operation.

switch_types

Creates a(n) switch_types operation.