Module melior::dialect::ods::shape

shape dialect.

This dialect contains operations for shape inference.

Note: Unless explicitly stated, all functions that return a shape and take shapes as input, return the invalid shape if one of its operands is an invalid shape. This avoids flagging multiple errors for one verification failure. The dialect itself does not specify how errors should be combined (there are multiple different options, from always choosing first operand, concatting etc. on how to combine them).

Structs

AddOperation

A(n) add operation. Addition of sizes and indices.

AddOperationBuilder

A builder for a(n) add operation.

AnyOperation

A(n) any operation. Return any combination of the input shapes.

AnyOperationBuilder

A builder for a(n) any operation.

AssumingAllOperation

A(n) assuming_all operation. Return a logical AND of all witnesses.

AssumingAllOperationBuilder

A builder for a(n) assuming_all operation.

AssumingOperation

A(n) assuming operation. Execute the region.

AssumingOperationBuilder

A builder for a(n) assuming operation.

AssumingYieldOperation

A(n) assuming_yield operation. Yield operation.

AssumingYieldOperationBuilder

A builder for a(n) assuming_yield operation.

BroadcastOperation

A(n) broadcast operation. Returns the broadcasted output shape of two or more inputs.

BroadcastOperationBuilder

A builder for a(n) broadcast operation.

ConcatOperation

A(n) concat operation. Concatenates two shapes.

ConcatOperationBuilder

A builder for a(n) concat operation.

ConstShapeOperation

A(n) const_shape operation. Creates a constant shape or extent tensor.

ConstShapeOperationBuilder

A builder for a(n) const_shape operation.

ConstSizeOperation

A(n) const_size operation. Creates a constant of type shape.size.

ConstSizeOperationBuilder

A builder for a(n) const_size operation.

ConstWitnessOperation

A(n) const_witness operation. An operation that returns a statically known witness value.

ConstWitnessOperationBuilder

A builder for a(n) const_witness operation.

CstrBroadcastableOperation

A(n) cstr_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

CstrBroadcastableOperationBuilder

A builder for a(n) cstr_broadcastable operation.

CstrEqOperation

A(n) cstr_eq operation. Determines if all input shapes are equal.

CstrEqOperationBuilder

A builder for a(n) cstr_eq operation.

CstrRequireOperation

A(n) cstr_require operation. Represents a runtime assertion that an i1 is true.

CstrRequireOperationBuilder

A builder for a(n) cstr_require operation.

DebugPrintOperation

A(n) debug_print operation. Prints the input shape or size.

DebugPrintOperationBuilder

A builder for a(n) debug_print operation.

DimOperation

A(n) dim operation. Gets the specified extent from the shape of a shaped input.

DimOperationBuilder

A builder for a(n) dim operation.

DivOperation

A(n) div operation. Division of sizes and indices.

DivOperationBuilder

A builder for a(n) div operation.

FromExtentTensorOperation

A(n) from_extent_tensor operation. Creates a shape from a tensor of extents.

FromExtentTensorOperationBuilder

A builder for a(n) from_extent_tensor operation.

FromExtentsOperation

A(n) from_extents operation. Creates a shape from extents.

FromExtentsOperationBuilder

A builder for a(n) from_extents operation.

FuncOperation

A(n) func operation. Shape function.

FuncOperationBuilder

A builder for a(n) func operation.

FunctionLibraryOperation

A(n) function_library operation. Represents shape functions and corresponding ops.

FunctionLibraryOperationBuilder

A builder for a(n) function_library operation.

GetExtentOperation

A(n) get_extent operation. Gets the specified extent from a shape or extent tensor.

GetExtentOperationBuilder

A builder for a(n) get_extent operation.

IndexToSizeOperation

A(n) index_to_size operation. Converts a standard index to a shape size.

IndexToSizeOperationBuilder

A builder for a(n) index_to_size operation.

IsBroadcastableOperation

A(n) is_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

IsBroadcastableOperationBuilder

A builder for a(n) is_broadcastable operation.

MaxOperation

A(n) max operation. Elementwise maximum.

MaxOperationBuilder

A builder for a(n) max operation.

MeetOperation

A(n) meet operation. Returns the least general shape or size of its operands.

MeetOperationBuilder

A builder for a(n) meet operation.

MinOperation

A(n) min operation. Elementwise minimum.

MinOperationBuilder

A builder for a(n) min operation.

MulOperation

A(n) mul operation. Multiplication of sizes and indices.

MulOperationBuilder

A builder for a(n) mul operation.

NumElementsOperation

A(n) num_elements operation. Returns the number of elements for a given shape.

NumElementsOperationBuilder

A builder for a(n) num_elements operation.

RankOperation

A(n) rank operation. Gets the rank of a shape.

RankOperationBuilder

A builder for a(n) rank operation.

ReduceOperation

A(n) reduce operation. Returns an expression reduced over a shape or extent tensor.

ReduceOperationBuilder

A builder for a(n) reduce operation.

ReturnOperation

A(n) return operation. Shape function return operation.

ReturnOperationBuilder

A builder for a(n) return operation.

ShapeEqOperation

A(n) shape_eq operation. Returns whether the input shapes or extent tensors are equal.

ShapeEqOperationBuilder

A builder for a(n) shape_eq operation.

ShapeOfOperation

A(n) shape_of operation. Returns shape of a value or shaped type operand.

ShapeOfOperationBuilder

A builder for a(n) shape_of operation.

SizeToIndexOperation

A(n) size_to_index operation. Casts between index types of the shape and standard dialect.

SizeToIndexOperationBuilder

A builder for a(n) size_to_index operation.

SplitAtOperation

A(n) split_at operation. Splits a shape at a given index.

SplitAtOperationBuilder

A builder for a(n) split_at operation.

ToExtentTensorOperation

A(n) to_extent_tensor operation. Creates a dimension tensor from a shape.

ToExtentTensorOperationBuilder

A builder for a(n) to_extent_tensor operation.

ValueAsShapeOperation

A(n) value_as_shape operation. Returns value as a shape.

ValueAsShapeOperationBuilder

A builder for a(n) value_as_shape operation.

ValueOfOperation

A(n) value_of operation. Returns value of a !shape.value_shape operand.

ValueOfOperationBuilder

A builder for a(n) value_of operation.

WithOperation

A(n) with_shape operation. Returns ValueShape with given shape.

WithOperationBuilder

A builder for a(n) with_shape operation.

YieldOperation

A(n) yield operation. Returns the value to parent op.

YieldOperationBuilder

A builder for a(n) yield operation.

Functions

add

Creates a(n) add operation.

any

Creates a(n) any operation.

assuming

Creates a(n) assuming operation.

assuming_all

Creates a(n) assuming_all operation.

assuming_yield

Creates a(n) assuming_yield operation.

broadcast

Creates a(n) broadcast operation.

concat

Creates a(n) concat operation.

const_shape

Creates a(n) const_shape operation.

const_size

Creates a(n) const_size operation.

const_witness

Creates a(n) const_witness operation.

cstr_broadcastable

Creates a(n) cstr_broadcastable operation.

cstr_eq

Creates a(n) cstr_eq operation.

cstr_require

Creates a(n) cstr_require operation.

debug_print

Creates a(n) debug_print operation.

dim

Creates a(n) dim operation.

div

Creates a(n) div operation.

from_extent_tensor

Creates a(n) from_extent_tensor operation.

from_extents

Creates a(n) from_extents operation.

func

Creates a(n) func operation.

function_library

Creates a(n) function_library operation.

get_extent

Creates a(n) get_extent operation.

index_to_size

Creates a(n) index_to_size operation.

is_broadcastable

Creates a(n) is_broadcastable operation.

max

Creates a(n) max operation.

meet

Creates a(n) meet operation.

min

Creates a(n) min operation.

mul

Creates a(n) mul operation.

num_elements

Creates a(n) num_elements operation.

rank

Creates a(n) rank operation.

reduce

Creates a(n) reduce operation.

return

Creates a(n) return operation.

shape_eq

Creates a(n) shape_eq operation.

shape_of

Creates a(n) shape_of operation.

size_to_index

Creates a(n) size_to_index operation.

split_at

Creates a(n) split_at operation.

to_extent_tensor

Creates a(n) to_extent_tensor operation.

value_as_shape

Creates a(n) value_as_shape operation.

value_of

Creates a(n) value_of operation.

with_shape

Creates a(n) with_shape operation.

yield

Creates a(n) yield operation.