Module melior::dialect::ods::sparse_tensor

sparse_tensor dialect.

The SparseTensor dialect supports all the attributes, types, operations, and passes that are required to make sparse tensor types first class citizens within the MLIR compiler infrastructure. The dialect forms a bridge between high-level operations on sparse tensors types and lower-level operations on the actual sparse storage schemes consisting of positions, coordinates, and values. Lower-level support may consist of fully generated code or may be provided by means of a small sparse runtime support library.

The concept of treating sparsity as a property, not a tedious implementation detail, by letting a sparse compiler generate sparse code automatically was pioneered for linear algebra by [Bik96] in MT1 (see https://www.aartbik.com/sparse.php) and formalized to tensor algebra by [Kjolstad17,Kjolstad20] in the Sparse Tensor Algebra Compiler (TACO) project (see http://tensor-compiler.org).

The MLIR implementation [Biketal22] closely follows the “sparse iteration theory” that forms the foundation of TACO. A rewriting rule is applied to each tensor expression in the Linalg dialect (MLIR’s tensor index notation) where the sparsity of tensors is indicated using the per-level level-types (e.g., dense, compressed, singleton) together with a specification of the order on the levels (see [Chou18] for an in-depth discussions and possible extensions to these level-types). Subsequently, a topologically sorted iteration graph, reflecting the required order on coordinates with respect to the levels of each tensor, is constructed to ensure that all tensors are visited in natural level-coordinate order. Next, iteration lattices are constructed for the tensor expression for every index in topological order. Each iteration lattice point consists of a conjunction of tensor coordinates together with a tensor (sub)expression that needs to be evaluated for that conjunction. Within the lattice, iteration points are ordered according to the way coordinates are exhausted. As such these iteration lattices drive actual sparse code generation, which consists of a relatively straightforward one-to-one mapping from iteration lattices to combinations of for-loops, while-loops, and if-statements. Sparse tensor outputs that materialize uninitialized are handled with direct insertions if all parallel loops are outermost or insertions that indirectly go through a 1-dimensional access pattern expansion (a.k.a. workspace) where feasible [Gustavson72,Bik96,Kjolstad19].

[Bik96] Aart J.C. Bik. Compiler Support for Sparse Matrix Computations. PhD thesis, Leiden University, May 1996.

[Biketal22] Aart J.C. Bik, Penporn Koanantakool, Tatiana Shpeisman, Nicolas Vasilache, Bixia Zheng, and Fredrik Kjolstad. Compiler Support for Sparse Tensor Computations in MLIR. ACM Transactions on Architecture and Code Optimization, June, 2022. See: https://dl.acm.org/doi/10.1145/3544559

[Chou18] Stephen Chou, Fredrik Berg Kjolstad, and Saman Amarasinghe. Format Abstraction for Sparse Tensor Algebra Compilers. Proceedings of the ACM on Programming Languages, October 2018.

[Chou20] Stephen Chou, Fredrik Berg Kjolstad, and Saman Amarasinghe. Automatic Generation of Efficient Sparse Tensor Format Conversion Routines. Proceedings of the 41st ACM SIGPLAN Conference on Programming Language Design and Implementation, June, 2020.

[Gustavson72] Fred G. Gustavson. Some basic techniques for solving sparse systems of linear equations. In Sparse Matrices and Their Applications, pages 41–52. Plenum Press, New York, 1972.

[Kjolstad17] Fredrik Berg Kjolstad, Shoaib Ashraf Kamil, Stephen Chou, David Lugato, and Saman Amarasinghe. The Tensor Algebra Compiler. Proceedings of the ACM on Programming Languages, October 2017.

[Kjolstad19] Fredrik Berg Kjolstad, Peter Ahrens, Shoaib Ashraf Kamil, and Saman Amarasinghe. Tensor Algebra Compilation with Workspaces, Proceedings of the IEEE/ACM International Symposium on Code Generation and Optimization, 2019.

[Kjolstad20] Fredrik Berg Kjolstad. Sparse Tensor Algebra Compilation. PhD thesis, MIT, February, 2020.

Structs

BinaryOp

binary operation. Binary set operation utilized within linalg.generic.

BinaryOpBuilder

A builder for binary operation. Binary set operation utilized within linalg.generic.

CompressOp

compress operation. Compressed an access pattern for insertion.

CompressOpBuilder

A builder for compress operation. Compressed an access pattern for insertion.

ConcatenateOp

concatenate operation. Concatenates a list of tensors into a single tensor..

ConcatenateOpBuilder

A builder for concatenate operation. Concatenates a list of tensors into a single tensor..

ConvertOp

convert operation. Converts between different tensor types.

ConvertOpBuilder

A builder for convert operation. Converts between different tensor types.

ExpandOp

expand operation. Expands an access pattern for insertion.

ExpandOpBuilder

A builder for expand operation. Expands an access pattern for insertion.

ForeachOp

foreach operation. Iterates over elements in a tensor.

ForeachOpBuilder

A builder for foreach operation. Iterates over elements in a tensor.

GetStorageSpecifierOp

storage_specifier.get operation.

GetStorageSpecifierOpBuilder

A builder for storage_specifier.get operation.

InsertOp

insert operation. Inserts a value into the sparse tensor.

InsertOpBuilder

A builder for insert operation. Inserts a value into the sparse tensor.

LoadOp

load operation. Rematerializes tensor from underlying sparse storage format.

LoadOpBuilder

A builder for load operation. Rematerializes tensor from underlying sparse storage format.

NewOp

new operation. Materializes a new sparse tensor from given source.

NewOpBuilder

A builder for new operation. Materializes a new sparse tensor from given source.

NumberOfEntriesOp

number_of_entries operation. Returns the number of entries that are stored in the tensor..

NumberOfEntriesOpBuilder

A builder for number_of_entries operation. Returns the number of entries that are stored in the tensor..

OutOp

out operation. Outputs a sparse tensor to the given destination.

OutOpBuilder

A builder for out operation. Outputs a sparse tensor to the given destination.

PackOp

pack operation. Returns a sparse tensor from the given values, levels.

PackOpBuilder

A builder for pack operation. Returns a sparse tensor from the given values, levels.

PushBackOp

push_back operation. Pushes a value to the back of a given buffer.

PushBackOpBuilder

A builder for push_back operation. Pushes a value to the back of a given buffer.

ReduceOp

reduce operation. Custom reduction operation utilized within linalg.generic.

ReduceOpBuilder

A builder for reduce operation. Custom reduction operation utilized within linalg.generic.

SelectOp

select operation. Select operation utilized within linalg.generic.

SelectOpBuilder

A builder for select operation. Select operation utilized within linalg.generic.

SetStorageSpecifierOp

storage_specifier.set operation.

SetStorageSpecifierOpBuilder

A builder for storage_specifier.set operation.

SortCooOp

sort_coo operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

SortCooOpBuilder

A builder for sort_coo operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

SortOp

sort operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

SortOpBuilder

A builder for sort operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

StorageSpecifierInitOp

storage_specifier.init operation.

StorageSpecifierInitOpBuilder

A builder for storage_specifier.init operation.

ToCoordinatesBufferOp

coordinates_buffer operation. Extracts the linear coordinates array from a tensor.

ToCoordinatesBufferOpBuilder

A builder for coordinates_buffer operation. Extracts the linear coordinates array from a tensor.

ToCoordinatesOp

coordinates operation. Extracts the level-th coordinates array of the tensor.

ToCoordinatesOpBuilder

A builder for coordinates operation. Extracts the level-th coordinates array of the tensor.

ToPositionsOp

positions operation. Extracts the level-th positions array of the tensor.

ToPositionsOpBuilder

A builder for positions operation. Extracts the level-th positions array of the tensor.

ToSliceOffsetOp

slice.offset operation. Extracts the offset of the sparse tensor slice at the given dimension.

ToSliceOffsetOpBuilder

A builder for slice.offset operation. Extracts the offset of the sparse tensor slice at the given dimension.

ToSliceStrideOp

slice.stride operation. Extracts the stride of the sparse tensor slice at the given dimension.

ToSliceStrideOpBuilder

A builder for slice.stride operation. Extracts the stride of the sparse tensor slice at the given dimension.

ToValuesOp

values operation. Extracts numerical values array from a tensor.

ToValuesOpBuilder

A builder for values operation. Extracts numerical values array from a tensor.

UnaryOp

unary operation. Unary set operation utilized within linalg.generic.

UnaryOpBuilder

A builder for unary operation. Unary set operation utilized within linalg.generic.

UnpackOp

unpack operation. Returns the (values, coordinates) pair unpacked from the input tensor.

UnpackOpBuilder

A builder for unpack operation. Returns the (values, coordinates) pair unpacked from the input tensor.

YieldOp

yield operation. Yield from sparse_tensor set-like operations.

YieldOpBuilder

A builder for yield operation. Yield from sparse_tensor set-like operations.

Functions

binary

Creates a new binary operation. Binary set operation utilized within linalg.generic.

compress

Creates a new compress operation. Compressed an access pattern for insertion.

concatenate

Creates a new concatenate operation. Concatenates a list of tensors into a single tensor..

convert

Creates a new convert operation. Converts between different tensor types.

coordinates

Creates a new coordinates operation. Extracts the level-th coordinates array of the tensor.

coordinates_buffer

Creates a new coordinates_buffer operation. Extracts the linear coordinates array from a tensor.

expand

Creates a new expand operation. Expands an access pattern for insertion.

foreach

Creates a new foreach operation. Iterates over elements in a tensor.

insert

Creates a new insert operation. Inserts a value into the sparse tensor.

load

Creates a new load operation. Rematerializes tensor from underlying sparse storage format.

new

Creates a new new operation. Materializes a new sparse tensor from given source.

number_of_entries

Creates a new number_of_entries operation. Returns the number of entries that are stored in the tensor..

out

Creates a new out operation. Outputs a sparse tensor to the given destination.

pack

Creates a new pack operation. Returns a sparse tensor from the given values, levels.

positions

Creates a new positions operation. Extracts the level-th positions array of the tensor.

push_back

Creates a new push_back operation. Pushes a value to the back of a given buffer.

reduce

Creates a new reduce operation. Custom reduction operation utilized within linalg.generic.

select

Creates a new select operation. Select operation utilized within linalg.generic.

slice_offset

Creates a new slice.offset operation. Extracts the offset of the sparse tensor slice at the given dimension.

slice_stride

Creates a new slice.stride operation. Extracts the stride of the sparse tensor slice at the given dimension.

sort

Creates a new sort operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

sort_coo

Creates a new sort_coo operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

storage_specifier_get

Creates a new storage_specifier.get operation.

storage_specifier_init

Creates a new storage_specifier.init operation.

storage_specifier_set

Creates a new storage_specifier.set operation.

unary

Creates a new unary operation. Unary set operation utilized within linalg.generic.

unpack

Creates a new unpack operation. Returns the (values, coordinates) pair unpacked from the input tensor.

values

Creates a new values operation. Extracts numerical values array from a tensor.

yield

Creates a new yield operation. Yield from sparse_tensor set-like operations.