Module melior::dialect::ods::tensor

tensor dialect.

The tensor dialect is intended to hold core tensor creation and manipulation ops, which are not strongly associated with any particular other dialect or domain abstraction. The primary smoke test of this is ops that make sense for any tensor element type.

We leave it to other dialects to hold the vast swath of possible computations one might want to do on a tensor.

The tensor type is (for better or for worse) used to represent all kinds of things, and supports an open-ended set of element types. Examples:

representing large, dense aggregations of primitive types, suitable for high-performance numerical computing.

representing shapes in the shape dialect, which consist of small 1D tensors of index data type.

representing aggregations of strings or “variant” types.

representing large, sparse aggregations of primitive types, suitable for high-performance numerical computing.

Thus, for the tensor dialect, we prefer for now to constrain the scope as much as possible. The expectation is that at some point in the future, the tensor dialect’s scope may be broadened through a careful discussion of the tradeoffs.

The tensor type is actually a builtin type (it lives in the builtin dialect), and does not live in this dialect.

Structs

BitcastOperation

A(n) bitcast operation. tensor bitcast operation.

BitcastOperationBuilder

A builder for a(n) bitcast operation.

CastOperation

A(n) cast operation. tensor cast operation.

CastOperationBuilder

A builder for a(n) cast operation.

CollapseShapeOperation

A(n) collapse_shape operation. operation to produce a tensor with a smaller rank.

CollapseShapeOperationBuilder

A builder for a(n) collapse_shape operation.

DimOperation

A(n) dim operation. dimension index operation.

DimOperationBuilder

A builder for a(n) dim operation.

EmptyOperation

A(n) empty operation. empty tensor operation.

EmptyOperationBuilder

A builder for a(n) empty operation.

ExpandShapeOperation

A(n) expand_shape operation. operation to produce a tensor with a higher rank.

ExpandShapeOperationBuilder

A builder for a(n) expand_shape operation.

ExtractOperation

A(n) extract operation. element extraction operation.

ExtractOperationBuilder

A builder for a(n) extract operation.

ExtractSliceOperation

A(n) extract_slice operation. extract slice operation.

ExtractSliceOperationBuilder

A builder for a(n) extract_slice operation.

FromElementsOperation

A(n) from_elements operation. tensor from elements operation..

FromElementsOperationBuilder

A builder for a(n) from_elements operation.

GatherOperation

A(n) gather operation. gather a subset of a tensor at specified indices.

GatherOperationBuilder

A builder for a(n) gather operation.

GenerateOperation

A(n) generate operation. Creates a dynamically sized tensor from elements.

GenerateOperationBuilder

A builder for a(n) generate operation.

InsertOperation

A(n) insert operation. element insertion operation.

InsertOperationBuilder

A builder for a(n) insert operation.

InsertSliceOperation

A(n) insert_slice operation. insert_slice operation.

InsertSliceOperationBuilder

A builder for a(n) insert_slice operation.

PackOperation

A(n) pack operation. tensor pack operation.

PackOperationBuilder

A builder for a(n) pack operation.

PadOperation

A(n) pad operation. tensor pad operation.

PadOperationBuilder

A builder for a(n) pad operation.

ParallelInsertSliceOperation

A(n) parallel_insert_slice operation. Specify the tensor slice update of a single thread of a parent ParallelCombiningOpInterface op. .

ParallelInsertSliceOperationBuilder

A builder for a(n) parallel_insert_slice operation.

RankOperation

A(n) rank operation. rank operation.

RankOperationBuilder

A builder for a(n) rank operation.

ReshapeOperation

A(n) reshape operation. tensor reshape operation.

ReshapeOperationBuilder

A builder for a(n) reshape operation.

ScatterOperation

A(n) scatter operation. scatter a tensor into a destination tensor at specified indices.

ScatterOperationBuilder

A builder for a(n) scatter operation.

SplatOperation

A(n) splat operation. tensor splat or broadcast operation.

SplatOperationBuilder

A builder for a(n) splat operation.

UnPackOperation

A(n) unpack operation. tensor unpack operation.

UnPackOperationBuilder

A builder for a(n) unpack operation.

YieldOperation

A(n) yield operation. Yield a value from a region.

YieldOperationBuilder

A builder for a(n) yield operation.

Functions

bitcast

Creates a(n) bitcast operation.

cast

Creates a(n) cast operation.

collapse_shape

Creates a(n) collapse_shape operation.

dim

Creates a(n) dim operation.

empty

Creates a(n) empty operation.

expand_shape

Creates a(n) expand_shape operation.

extract

Creates a(n) extract operation.

extract_slice

Creates a(n) extract_slice operation.

from_elements

Creates a(n) from_elements operation.

gather

Creates a(n) gather operation.

generate

Creates a(n) generate operation.

insert

Creates a(n) insert operation.

insert_slice

Creates a(n) insert_slice operation.

pack

Creates a(n) pack operation.

pad

Creates a(n) pad operation.

parallel_insert_slice

Creates a(n) parallel_insert_slice operation.

rank

Creates a(n) rank operation.

reshape

Creates a(n) reshape operation.

scatter

Creates a(n) scatter operation.

splat

Creates a(n) splat operation.

unpack

Creates a(n) unpack operation.

yield

Creates a(n) yield operation.