Module melior::dialect::ods::tosa

tosa dialect.

The Tensor Operator Set Architecture (TOSA) dialect.

This dialect implements the TOSA standard described at https://www.mlplatform.org/tosa/tosa_spec.html .

Tensor Operator Set Architecture (TOSA) provides a set of whole-tensor operations commonly employed by Deep Neural Networks. The intent is to enable a variety of implementations running on a diverse range of processors, with the results at the TOSA level consistent across those implementations. Applications or frameworks which target TOSA can therefore be deployed on a wide range of different processors, such as CPUs or GPUs, with defined accuracy and compatibility constraints. Most operators from the common ML frameworks should be expressible in TOSA. It is expected that there will be tools to lower from the ML frameworks into TOSA.

Structs

AbsOperation

An abs operation. Elementwise abs op.

AbsOperationBuilder

A builder for an abs operation.

AddOperation

An add operation. Elementwise addition operator.

AddOperationBuilder

A builder for an add operation.

ApplyScaleOperation

An apply_scale operation. Rescale scalar operator for Tosa tensor operators.

ApplyScaleOperationBuilder

A builder for an apply_scale operation.

ArgMaxOperation

An argmax operation. Perform argmax on the input..

ArgMaxOperationBuilder

A builder for an argmax operation.

ArithmeticRightShiftOperation

An arithmetic_right_shift operation. Elementwise Arithmetic Right Shift.

ArithmeticRightShiftOperationBuilder

A builder for an arithmetic_right_shift operation.

AvgPool2dOperation

An avg_pool2d operation. Performs max pooling on the input..

AvgPool2dOperationBuilder

A builder for an avg_pool2d operation.

BitwiseAndOperation

A bitwise_and operation. Bitwise AND operator.

BitwiseAndOperationBuilder

A builder for a bitwise_and operation.

BitwiseNotOperation

A bitwise_not operation. Bitwise NOT operator.

BitwiseNotOperationBuilder

A builder for a bitwise_not operation.

BitwiseOrOperation

A bitwise_or operation. Bitwise OR operator.

BitwiseOrOperationBuilder

A builder for a bitwise_or operation.

BitwiseXorOperation

A bitwise_xor operation. Bitwise XOR operator.

BitwiseXorOperationBuilder

A builder for a bitwise_xor operation.

CastOperation

A cast operation. Cast operation.

CastOperationBuilder

A builder for a cast operation.

CeilOperation

A ceil operation. Elementwise ceil op.

CeilOperationBuilder

A builder for a ceil operation.

ClampOperation

A clamp operation. Computes clamp(features, min, max)..

ClampOperationBuilder

A builder for a clamp operation.

ClzOperation

A clz operation. Elementwise count leading zero op.

ClzOperationBuilder

A builder for a clz operation.

ConcatOperation

A concat operation. Concatenates tensors along one dimension..

ConcatOperationBuilder

A builder for a concat operation.

ConstOperation

A const operation. Constant op..

ConstOperationBuilder

A builder for a const operation.

Conv2DOperation

A conv2d operation. 2D Convolution Operator.

Conv2DOperationBuilder

A builder for a conv2d operation.

Conv3DOperation

A conv3d operation. 3D Convolution operator.

Conv3DOperationBuilder

A builder for a conv3d operation.

CustomOperation

A custom operation. Custom operator wrapper for Tosa.

CustomOperationBuilder

A builder for a custom operation.

DepthwiseConv2DOperation

A depthwise_conv2d operation. Depthwise 2D Convolution operator.

DepthwiseConv2DOperationBuilder

A builder for a depthwise_conv2d operation.

DivOperation

A div operation. Integer divide operator.

DivOperationBuilder

A builder for a div operation.

EqualOperation

An equal operation. Returns the truth value of (x == y) element-wise..

EqualOperationBuilder

A builder for an equal operation.

ErfOperation

An erf operation. Computes gauss error function of input.

ErfOperationBuilder

A builder for an erf operation.

ExpOperation

An exp operation. Elementwise exp op.

ExpOperationBuilder

A builder for an exp operation.

FFT2dOperation

A fft2d operation. Performs FFT2D operation on the input..

FFT2dOperationBuilder

A builder for a fft2d operation.

FloorOperation

A floor operation. Elementwise floor op.

FloorOperationBuilder

A builder for a floor operation.

FullyConnectedOperation

A fully_connected operation. Fully Connected operator.

FullyConnectedOperationBuilder

A builder for a fully_connected operation.

GatherOperation

A gather operation. Gather operation,.

GatherOperationBuilder

A builder for a gather operation.

GreaterEqualOperation

A greater_equal operation. Returns the truth value of (x >= y) element-wise..

GreaterEqualOperationBuilder

A builder for a greater_equal operation.

GreaterOperation

A greater operation. Returns the truth value of (x > y) element-wise..

GreaterOperationBuilder

A builder for a greater operation.

IdentityOperation

An identity operation. Identity operator.

IdentityOperationBuilder

A builder for an identity operation.

IfOperation

A cond_if operation. Conditional if operator.

IfOperationBuilder

A builder for a cond_if operation.

LogOperation

A log operation. Elementwise log op.

LogOperationBuilder

A builder for a log operation.

LogicalAndOperation

A logical_and operation. Returns the truth value of x AND y element-wise..

LogicalAndOperationBuilder

A builder for a logical_and operation.

LogicalLeftShiftOperation

A logical_left_shift operation. Elementwise Logical Left Shift.

LogicalLeftShiftOperationBuilder

A builder for a logical_left_shift operation.

LogicalNotOperation

A logical_not operation. Returns the truth value of NOT x element-wise..

LogicalNotOperationBuilder

A builder for a logical_not operation.

LogicalOrOperation

A logical_or operation. Returns the truth value of x OR y element-wise..

LogicalOrOperationBuilder

A builder for a logical_or operation.

LogicalRightShiftOperation

A logical_right_shift operation. Elementwise Logical Right Shift.

LogicalRightShiftOperationBuilder

A builder for a logical_right_shift operation.

LogicalXorOperation

A logical_xor operation. Returns the truth value of x XOR y element-wise..

LogicalXorOperationBuilder

A builder for a logical_xor operation.

MatMulOperation

A matmul operation. Matrix multiplication with bias.

MatMulOperationBuilder

A builder for a matmul operation.

MaxPool2dOperation

A max_pool2d operation. Performs max pooling on the input..

MaxPool2dOperationBuilder

A builder for a max_pool2d operation.

MaximumOperation

A maximum operation. Elementwise Maximum.

MaximumOperationBuilder

A builder for a maximum operation.

MinimumOperation

A minimum operation. Elementwise Minimum.

MinimumOperationBuilder

A builder for a minimum operation.

MulOperation

A mul operation. Multiplication operator.

MulOperationBuilder

A builder for a mul operation.

NegateOperation

A negate operation. Elementwise negate op.

NegateOperationBuilder

A builder for a negate operation.

PadOperation

A pad operation. Pads a tensor with value specified..

PadOperationBuilder

A builder for a pad operation.

PowOperation

A pow operation. Computes the power of one value to another..

PowOperationBuilder

A builder for a pow operation.

RFFT2dOperation

A rfft2d operation. Performs RFFT2D operation on the input..

RFFT2dOperationBuilder

A builder for a rfft2d operation.

ReciprocalOperation

A reciprocal operation. Elementwise reciprocal op.

ReciprocalOperationBuilder

A builder for a reciprocal operation.

ReduceAllOperation

A reduce_all operation. Reduce All operator.

ReduceAllOperationBuilder

A builder for a reduce_all operation.

ReduceAnyOperation

A reduce_any operation. Reduce Any operator.

ReduceAnyOperationBuilder

A builder for a reduce_any operation.

ReduceMaxOperation

A reduce_max operation. Reduce Max operator.

ReduceMaxOperationBuilder

A builder for a reduce_max operation.

ReduceMinOperation

A reduce_min operation. Reduce Min operator.

ReduceMinOperationBuilder

A builder for a reduce_min operation.

ReduceProdOperation

A reduce_prod operation. Reduce Prod operator.

ReduceProdOperationBuilder

A builder for a reduce_prod operation.

ReduceSumOperation

A reduce_sum operation. Reduce Sum operator.

ReduceSumOperationBuilder

A builder for a reduce_sum operation.

RescaleOperation

A rescale operation. Tosa rescale operator.

RescaleOperationBuilder

A builder for a rescale operation.

ReshapeOperation

A reshape operation. Reshape operator.

ReshapeOperationBuilder

A builder for a reshape operation.

ResizeOperation

A resize operation. Resize operation, supports various resize/upsample modes.

ResizeOperationBuilder

A builder for a resize operation.

ReverseOperation

A reverse operation. Reverse operator.

ReverseOperationBuilder

A builder for a reverse operation.

RsqrtOperation

A rsqrt operation. Elementwise 1/sqrt op.

RsqrtOperationBuilder

A builder for a rsqrt operation.

ScatterOperation

A scatter operation. Scatter operation,.

ScatterOperationBuilder

A builder for a scatter operation.

SelectOperation

A select operation. Elementwise select operator.

SelectOperationBuilder

A builder for a select operation.

SigmoidOperation

A sigmoid operation. Computes elementwise sigmoid of input..

SigmoidOperationBuilder

A builder for a sigmoid operation.

SliceOperation

A slice operation. Slice operator.

SliceOperationBuilder

A builder for a slice operation.

SubOperation

A sub operation. Elementwise subtraction operator.

SubOperationBuilder

A builder for a sub operation.

TableOperation

A table operation. Table lookup op.

TableOperationBuilder

A builder for a table operation.

TanhOperation

A tanh operation. Computes elementwise hyperbolic tangent of input.

TanhOperationBuilder

A builder for a tanh operation.

TileOperation

A tile operation. Tile operator.

TileOperationBuilder

A builder for a tile operation.

TransposeConv2DOperation

A transpose_conv2d operation. Transpose 2D Convolution operator..

TransposeConv2DOperationBuilder

A builder for a transpose_conv2d operation.

TransposeOperation

A transpose operation. Transpose operator.

TransposeOperationBuilder

A builder for a transpose operation.

WhileOperation

A while_loop operation. Output = input; While (Cond(output)) {output = Body(output)}.

WhileOperationBuilder

A builder for a while_loop operation.

YieldOperation

A yield operation. Yield operator.

YieldOperationBuilder

A builder for a yield operation.

Functions

abs

Creates an abs operation.

add

Creates an add operation.

apply_scale

Creates an apply_scale operation.

argmax

Creates an argmax operation.

arithmetic_right_shift

Creates an arithmetic_right_shift operation.

avg_pool_2_d

Creates an avg_pool2d operation.

bitwise_and

Creates a bitwise_and operation.

bitwise_not

Creates a bitwise_not operation.

bitwise_or

Creates a bitwise_or operation.

bitwise_xor

Creates a bitwise_xor operation.

cast

Creates a cast operation.

ceil

Creates a ceil operation.

clamp

Creates a clamp operation.

clz

Creates a clz operation.

concat

Creates a concat operation.

cond_if

Creates a cond_if operation.

const

Creates a const operation.

conv_2_d

Creates a conv2d operation.

conv_3_d

Creates a conv3d operation.

custom

Creates a custom operation.

depthwise_conv_2_d

Creates a depthwise_conv2d operation.

div

Creates a div operation.

equal

Creates an equal operation.

erf

Creates an erf operation.

exp

Creates an exp operation.

fft_2_d

Creates a fft2d operation.

floor

Creates a floor operation.

fully_connected

Creates a fully_connected operation.

gather

Creates a gather operation.

greater

Creates a greater operation.

greater_equal

Creates a greater_equal operation.

identity

Creates an identity operation.

log

Creates a log operation.

logical_and

Creates a logical_and operation.

logical_left_shift

Creates a logical_left_shift operation.

logical_not

Creates a logical_not operation.

logical_or

Creates a logical_or operation.

logical_right_shift

Creates a logical_right_shift operation.

logical_xor

Creates a logical_xor operation.

matmul

Creates a matmul operation.

max_pool_2_d

Creates a max_pool2d operation.

maximum

Creates a maximum operation.

minimum

Creates a minimum operation.

mul

Creates a mul operation.

negate

Creates a negate operation.

pad

Creates a pad operation.

pow

Creates a pow operation.

reciprocal

Creates a reciprocal operation.

reduce_all

Creates a reduce_all operation.

reduce_any

Creates a reduce_any operation.

reduce_max

Creates a reduce_max operation.

reduce_min

Creates a reduce_min operation.

reduce_prod

Creates a reduce_prod operation.

reduce_sum

Creates a reduce_sum operation.

rescale

Creates a rescale operation.

reshape

Creates a reshape operation.

resize

Creates a resize operation.

reverse

Creates a reverse operation.

rfft_2_d

Creates a rfft2d operation.

rsqrt

Creates a rsqrt operation.

scatter

Creates a scatter operation.

select

Creates a select operation.

sigmoid

Creates a sigmoid operation.

slice

Creates a slice operation.

sub

Creates a sub operation.

table

Creates a table operation.

tanh

Creates a tanh operation.

tile

Creates a tile operation.

transpose

Creates a transpose operation.

transpose_conv_2_d

Creates a transpose_conv2d operation.

while_loop

Creates a while_loop operation.

yield

Creates a yield operation.