Module melior::dialect::ods::transform

transform dialect.

Structs

AlternativesOp

alternatives operation: Attempts sequences of transforms until one succeeds

AlternativesOpBuilder

Builder for alternatives operation: Attempts sequences of transforms until one succeeds

CastOp

cast operation

CastOpBuilder

Builder for cast operation

ForeachOp

foreach operation: Executes the body for each payload op

ForeachOpBuilder

Builder for foreach operation: Executes the body for each payload op

GetClosestIsolatedParentOp

get_closest_isolated_parent operation: Gets handles to the closest isolated-from-above parents

GetClosestIsolatedParentOpBuilder

Builder for get_closest_isolated_parent operation: Gets handles to the closest isolated-from-above parents

GetConsumersOfResult

get_consumers_of_result operation: Get handle to the consumers of this operation’s result number

GetConsumersOfResultBuilder

Builder for get_consumers_of_result operation: Get handle to the consumers of this operation’s result number

GetProducerOfOperand

get_producer_of_operand operation: Get handle to the producer of this operation’s operand number

GetProducerOfOperandBuilder

Builder for get_producer_of_operand operation: Get handle to the producer of this operation’s operand number

MergeHandlesOp

merge_handles operation: Merges handles into one pointing to the union of payload ops

MergeHandlesOpBuilder

Builder for merge_handles operation: Merges handles into one pointing to the union of payload ops

PDLMatchOp

pdl_match operation: Finds ops that match the named PDL pattern

PDLMatchOpBuilder

Builder for pdl_match operation: Finds ops that match the named PDL pattern

PrintOp

print operation: Dump each payload op

PrintOpBuilder

Builder for print operation: Dump each payload op

ReplicateOp

replicate operation: Lists payload ops multiple times in the new handle

ReplicateOpBuilder

Builder for replicate operation: Lists payload ops multiple times in the new handle

SequenceOp

sequence operation: Contains a sequence of other transform ops to apply

SequenceOpBuilder

Builder for sequence operation: Contains a sequence of other transform ops to apply

SplitHandlesOp

split_handles operation: Splits handles from a union of payload ops to a list

SplitHandlesOpBuilder

Builder for split_handles operation: Splits handles from a union of payload ops to a list

WithPDLPatternsOp

with_pdl_patterns operation: Contains PDL patterns available for use in transforms

WithPDLPatternsOpBuilder

Builder for with_pdl_patterns operation: Contains PDL patterns available for use in transforms

YieldOp

yield operation: Yields operation handles from a transform IR region

YieldOpBuilder

Builder for yield operation: Yields operation handles from a transform IR region

Functions

alternatives

Creates a new alternatives operation: Attempts sequences of transforms until one succeeds

cast

Creates a new cast operation

foreach

Creates a new foreach operation: Executes the body for each payload op

get_closest_isolated_parent

Creates a new get_closest_isolated_parent operation: Gets handles to the closest isolated-from-above parents

get_consumers_of_result

Creates a new get_consumers_of_result operation: Get handle to the consumers of this operation’s result number

get_producer_of_operand

Creates a new get_producer_of_operand operation: Get handle to the producer of this operation’s operand number

merge_handles

Creates a new merge_handles operation: Merges handles into one pointing to the union of payload ops

pdl_match

Creates a new pdl_match operation: Finds ops that match the named PDL pattern

print

Creates a new print operation: Dump each payload op

replicate

Creates a new replicate operation: Lists payload ops multiple times in the new handle

sequence

Creates a new sequence operation: Contains a sequence of other transform ops to apply

split_handles

Creates a new split_handles operation: Splits handles from a union of payload ops to a list

with_pdl_patterns

Creates a new with_pdl_patterns operation: Contains PDL patterns available for use in transforms

yield

Creates a new yield operation: Yields operation handles from a transform IR region