Module melior::dialect::ods::vector

vector dialect.

Structs

BitCastOperation

A(n) bitcast operation. bitcast casts between vectors.

BitCastOperationBuilder

A builder for a(n) bitcast operation.

BroadcastOperation

A(n) broadcast operation. broadcast operation.

BroadcastOperationBuilder

A builder for a(n) broadcast operation.

CompressStoreOperation

A(n) compressstore operation. writes elements selectively from a vector as defined by a mask.

CompressStoreOperationBuilder

A builder for a(n) compressstore operation.

ConstantMaskOperation

A(n) constant_mask operation. creates a constant vector mask.

ConstantMaskOperationBuilder

A builder for a(n) constant_mask operation.

ContractionOperation

A(n) contract operation. vector contraction operation.

ContractionOperationBuilder

A builder for a(n) contract operation.

CreateMaskOperation

A(n) create_mask operation. creates a vector mask.

CreateMaskOperationBuilder

A builder for a(n) create_mask operation.

ExpandLoadOperation

A(n) expandload operation. reads elements from memory and spreads them into a vector as defined by a mask.

ExpandLoadOperationBuilder

A builder for a(n) expandload operation.

ExtractElementOperation

A(n) extractelement operation. extractelement operation.

ExtractElementOperationBuilder

A builder for a(n) extractelement operation.

ExtractOperation

A(n) extract operation. extract operation.

ExtractOperationBuilder

A builder for a(n) extract operation.

ExtractStridedSliceOperation

A(n) extract_strided_slice operation. extract_strided_slice operation.

ExtractStridedSliceOperationBuilder

A builder for a(n) extract_strided_slice operation.

FMAOperation

A(n) fma operation. vector fused multiply-add.

FMAOperationBuilder

A builder for a(n) fma operation.

FlatTransposeOperation

A(n) flat_transpose operation. Vector matrix transposition on flattened 1-D MLIR vectors.

FlatTransposeOperationBuilder

A builder for a(n) flat_transpose operation.

GatherOperation

A(n) gather operation. gathers elements from memory or ranked tensor into a vector as defined by an index vector and a mask vector .

GatherOperationBuilder

A builder for a(n) gather operation.

InsertElementOperation

A(n) insertelement operation. insertelement operation.

InsertElementOperationBuilder

A builder for a(n) insertelement operation.

InsertOperation

A(n) insert operation. insert operation.

InsertOperationBuilder

A builder for a(n) insert operation.

InsertStridedSliceOperation

A(n) insert_strided_slice operation. strided_slice operation.

InsertStridedSliceOperationBuilder

A builder for a(n) insert_strided_slice operation.

LoadOperation

A(n) load operation. reads an n-D slice of memory into an n-D vector.

LoadOperationBuilder

A builder for a(n) load operation.

MaskOperation

A(n) mask operation. Predicates a maskable vector operation.

MaskOperationBuilder

A builder for a(n) mask operation.

MaskedLoadOperation

A(n) maskedload operation. loads elements from memory into a vector as defined by a mask vector.

MaskedLoadOperationBuilder

A builder for a(n) maskedload operation.

MaskedStoreOperation

A(n) maskedstore operation. stores elements from a vector into memory as defined by a mask vector.

MaskedStoreOperationBuilder

A builder for a(n) maskedstore operation.

MatmulOperation

A(n) matrix_multiply operation. Vector matrix multiplication op that operates on flattened 1-D MLIR vectors.

MatmulOperationBuilder

A builder for a(n) matrix_multiply operation.

MultiDimReductionOperation

A(n) multi_reduction operation. Multi-dimensional reduction operation.

MultiDimReductionOperationBuilder

A builder for a(n) multi_reduction operation.

OuterProductOperation

A(n) outerproduct operation. vector outerproduct with optional fused add.

OuterProductOperationBuilder

A builder for a(n) outerproduct operation.

PrintOperation

A(n) print operation. print operation (for testing and debugging).

PrintOperationBuilder

A builder for a(n) print operation.

ReductionOperation

A(n) reduction operation. reduction operation.

ReductionOperationBuilder

A builder for a(n) reduction operation.

ReshapeOperation

A(n) reshape operation. vector reshape operation.

ReshapeOperationBuilder

A builder for a(n) reshape operation.

ScalableExtractOperation

A(n) scalable.extract operation. extract subvector from scalable vector operation.

ScalableExtractOperationBuilder

A builder for a(n) scalable.extract operation.

ScalableInsertOperation

A(n) scalable.insert operation. insert subvector into scalable vector operation.

ScalableInsertOperationBuilder

A builder for a(n) scalable.insert operation.

ScanOperation

A(n) scan operation. Scan operation.

ScanOperationBuilder

A builder for a(n) scan operation.

ScatterOperation

A(n) scatter operation. scatters elements from a vector into memory as defined by an index vector and a mask vector .

ScatterOperationBuilder

A builder for a(n) scatter operation.

ShapeCastOperation

A(n) shape_cast operation. shape_cast casts between vector shapes.

ShapeCastOperationBuilder

A builder for a(n) shape_cast operation.

ShuffleOperation

A(n) shuffle operation. shuffle operation.

ShuffleOperationBuilder

A builder for a(n) shuffle operation.

SplatOperation

A(n) splat operation. vector splat or broadcast operation.

SplatOperationBuilder

A builder for a(n) splat operation.

StoreOperation

A(n) store operation. writes an n-D vector to an n-D slice of memory.

StoreOperationBuilder

A builder for a(n) store operation.

TransferReadOperation

A(n) transfer_read operation. Reads a supervector from memory into an SSA vector value..

TransferReadOperationBuilder

A builder for a(n) transfer_read operation.

TransferWriteOperation

A(n) transfer_write operation. The vector.transfer_write op writes a supervector to memory..

TransferWriteOperationBuilder

A builder for a(n) transfer_write operation.

TransposeOperation

A(n) transpose operation. vector transpose operation.

TransposeOperationBuilder

A builder for a(n) transpose operation.

TypeCastOperation

A(n) type_cast operation. type_cast op converts a scalar memref to a vector memref.

TypeCastOperationBuilder

A builder for a(n) type_cast operation.

VectorScaleOperation

A(n) vscale operation. Load vector scale size.

VectorScaleOperationBuilder

A builder for a(n) vscale operation.

WarpExecuteOnLane0Operation

A(n) warp_execute_on_lane_0 operation. Executes operations in the associated region on thread #0 of aSPMD program.

WarpExecuteOnLane0OperationBuilder

A builder for a(n) warp_execute_on_lane_0 operation.

YieldOperation

A(n) yield operation. Terminates and yields values from vector regions..

YieldOperationBuilder

A builder for a(n) yield operation.

Functions

bitcast

Creates a(n) bitcast operation.

broadcast

Creates a(n) broadcast operation.

compressstore

Creates a(n) compressstore operation.

constant_mask

Creates a(n) constant_mask operation.

contract

Creates a(n) contract operation.

create_mask

Creates a(n) create_mask operation.

expandload

Creates a(n) expandload operation.

extract

Creates a(n) extract operation.

extract_strided_slice

Creates a(n) extract_strided_slice operation.

extractelement

Creates a(n) extractelement operation.

flat_transpose

Creates a(n) flat_transpose operation.

fma

Creates a(n) fma operation.

gather

Creates a(n) gather operation.

insert

Creates a(n) insert operation.

insert_strided_slice

Creates a(n) insert_strided_slice operation.

insertelement

Creates a(n) insertelement operation.

load

Creates a(n) load operation.

mask

Creates a(n) mask operation.

maskedload

Creates a(n) maskedload operation.

maskedstore

Creates a(n) maskedstore operation.

matrix_multiply

Creates a(n) matrix_multiply operation.

multi_reduction

Creates a(n) multi_reduction operation.

outerproduct

Creates a(n) outerproduct operation.

print

Creates a(n) print operation.

reduction

Creates a(n) reduction operation.

reshape

Creates a(n) reshape operation.

scalable_extract

Creates a(n) scalable.extract operation.

scalable_insert

Creates a(n) scalable.insert operation.

scan

Creates a(n) scan operation.

scatter

Creates a(n) scatter operation.

shape_cast

Creates a(n) shape_cast operation.

shuffle

Creates a(n) shuffle operation.

splat

Creates a(n) splat operation.

store

Creates a(n) store operation.

transfer_read

Creates a(n) transfer_read operation.

transfer_write

Creates a(n) transfer_write operation.

transpose

Creates a(n) transpose operation.

type_cast

Creates a(n) type_cast operation.

vscale

Creates a(n) vscale operation.

warp_execute_on_lane_0

Creates a(n) warp_execute_on_lane_0 operation.

yield

Creates a(n) yield operation.