Trait melior::ir::attribute::AttributeLike

pub trait AttributeLike<'c> {

    // Required method
    fn to_raw(&self) -> MlirAttribute;

    // Provided methods
    fn context(&self) -> ContextRef<'c> { ... }
    fn type(&self) -> Type<'_> { ... }
    fn type_id(&self) -> TypeId<'c> { ... }
    fn dump(&self) { ... }
    fn is_affine_map(&self) -> bool { ... }
    fn is_array(&self) -> bool { ... }
    fn is_bool(&self) -> bool { ... }
    fn is_dense_bool_array(&self) -> bool { ... }
    fn is_dense_elements(&self) -> bool { ... }
    fn is_dense_f32_array(&self) -> bool { ... }
    fn is_dense_f64_array(&self) -> bool { ... }
    fn is_dense_fp_elements(&self) -> bool { ... }
    fn is_dense_i16_array(&self) -> bool { ... }
    fn is_dense_i32_array(&self) -> bool { ... }
    fn is_dense_i64_array(&self) -> bool { ... }
    fn is_dense_i8_array(&self) -> bool { ... }
    fn is_dense_int_elements(&self) -> bool { ... }
    fn is_dictionary(&self) -> bool { ... }
    fn is_elements(&self) -> bool { ... }
    fn is_flat_symbol_ref(&self) -> bool { ... }
    fn is_float(&self) -> bool { ... }
    fn is_integer(&self) -> bool { ... }
    fn is_integer_set(&self) -> bool { ... }
    fn is_opaque(&self) -> bool { ... }
    fn is_sparse_elements(&self) -> bool { ... }
    fn is_sparse_tensor_encoding_attr(&self) -> bool { ... }
    fn is_strided_layout(&self) -> bool { ... }
    fn is_string(&self) -> bool { ... }
    fn is_symbol_ref(&self) -> bool { ... }
    fn is_type(&self) -> bool { ... }
    fn is_unit(&self) -> bool { ... }

}

Trait for attribute-like types.

Required Methods

fn to_raw(&self) -> MlirAttribute

Converts a attribute into a raw object.

Provided Methods

fn context(&self) -> ContextRef<'c>

Gets a context.

fn type(&self) -> Type<'_>

Gets a type.

fn type_id(&self) -> TypeId<'c>

Gets a type ID.

fn dump(&self)

Dumps a attribute.

fn is_affine_map(&self) -> bool

Returns true if an attribute is affine map.

fn is_array(&self) -> bool

Returns true if an attribute is array.

fn is_bool(&self) -> bool

Returns true if an attribute is bool.

fn is_dense_bool_array(&self) -> bool

Returns true if an attribute is dense bool array.

fn is_dense_elements(&self) -> bool

Returns true if an attribute is dense elements.

fn is_dense_f32_array(&self) -> bool

Returns true if an attribute is dense f32 array.

fn is_dense_f64_array(&self) -> bool

Returns true if an attribute is dense f64 array.

fn is_dense_fp_elements(&self) -> bool

Returns true if an attribute is dense fp elements.

fn is_dense_i16_array(&self) -> bool

Returns true if an attribute is dense i16 array.

fn is_dense_i32_array(&self) -> bool

Returns true if an attribute is dense i32 array.

fn is_dense_i64_array(&self) -> bool

Returns true if an attribute is dense i64 array.

fn is_dense_i8_array(&self) -> bool

Returns true if an attribute is dense i8 array.

fn is_dense_int_elements(&self) -> bool

Returns true if an attribute is dense int elements.

fn is_dictionary(&self) -> bool

Returns true if an attribute is dictionary.

fn is_elements(&self) -> bool

Returns true if an attribute is elements.

fn is_flat_symbol_ref(&self) -> bool

Returns true if an attribute is flat symbol ref.

fn is_float(&self) -> bool

Returns true if an attribute is float.

fn is_integer(&self) -> bool

Returns true if an attribute is integer.

fn is_integer_set(&self) -> bool

Returns true if an attribute is integer set.

fn is_opaque(&self) -> bool

Returns true if an attribute is opaque.

fn is_sparse_elements(&self) -> bool

Returns true if an attribute is sparse elements.

fn is_sparse_tensor_encoding_attr(&self) -> bool

Returns true if an attribute is sparse tensor encoding attr.

fn is_strided_layout(&self) -> bool

Returns true if an attribute is strided layout.

fn is_string(&self) -> bool

Returns true if an attribute is string.

fn is_symbol_ref(&self) -> bool

Returns true if an attribute is symbol ref.

fn is_type(&self) -> bool

Returns true if an attribute is type.

fn is_unit(&self) -> bool

Returns true if an attribute is unit.