Struct melior::ir::block::Block

pub struct Block<'c> { /* private fields */ }

A block.

Implementations

impl<'c> Block<'c>

pub fn new(arguments: &[(Type<'c>, Location<'c>)]) -> Self

Creates a block.

pub fn argument(&self, position: usize) -> Result<Argument<'_>, Error>

Gets an argument at a position.

pub fn argument_count(&self) -> usize

Gets a number of arguments.

pub fn first_operation(&self) -> Option<OperationRef<'_>>

Gets the first operation.

pub fn terminator(&self) -> Option<OperationRef<'_>>

Gets a terminator operation.

pub fn parent_region(&self) -> Option<RegionRef<'_>>

Gets a parent region.

pub fn parent_operation(&self) -> Option<OperationRef<'_>>

Gets a parent operation.

pub fn add_argument(&self, type: Type<'c>, location: Location<'c>) -> Value<'_>

Adds an argument.

pub fn append_operation(&self, operation: Operation<'_>) -> OperationRef<'_>

Appends an operation.

pub fn insert_operation(
&self,
position: usize,
operation: Operation<'_>
) -> OperationRef<'_>

Inserts an operation.

pub fn insert_operation_after(
&self,
one: OperationRef<'_>,
other: Operation<'_>
) -> OperationRef<'_>

Inserts an operation after another.

pub fn insert_operation_before(
&self,
one: OperationRef<'_>,
other: Operation<'_>
) -> OperationRef<'_>

Inserts an operation before another.

pub unsafe fn detach(&self) -> Option<Block<'_>>

Detaches a block from a region and assumes its ownership.

Safety

This function might invalidate existing references to the block if you drop it too early.

pub fn next_in_region(&self) -> Option<BlockRef<'_>>

Gets a next block in a region.