Struct melior::ir::operation::OperationRef

pub struct OperationRef<'c, 'a> { /* private fields */ }

A reference to an operation.

Implementations

impl<'c, 'a> OperationRef<'c, 'a>

pub fn result(self, index: usize) -> Result<OperationResult<'c, 'a>, Error>

Gets a result at a position.

pub unsafe fn to_ref(&self) -> &'a Operation<'c>

Gets an operation.

This function is different from deref because the correct lifetime is kept for the return type.

Safety

The returned reference is safe to use only in the lifetime scope of the operation reference.

pub const fn to_raw(self) -> MlirOperation

Converts an operation reference into a raw object.

pub unsafe fn from_raw(raw: MlirOperation) -> Self

Creates an operation reference from a raw object.

Safety

A raw object must be valid.

pub unsafe fn from_option_raw(raw: MlirOperation) -> Option<Self>

Creates an optional operation reference from a raw object.

Safety

A raw object must be valid.