Struct melior::ir::Attribute

pub struct Attribute<'c> { /* private fields */ }

An attribute.

Implementations

impl<'c> Attribute<'c>

pub fn parse(context: &'c Context, source: &str) -> Option<Self>

Parses an attribute.

pub fn null() -> Self

Creates a null attribute.

pub fn type(&self) -> Option<Type<'c>>

Gets a type.

pub fn type_id(&self) -> Option<Id>

Gets a type ID.

pub fn is_null(&self) -> bool

Returns true if an attribute is null.

pub fn is_affine_map(&self) -> bool

Returns true if an attribute is a affine map.

pub fn is_array(&self) -> bool

Returns true if an attribute is a array.

pub fn is_bool(&self) -> bool

Returns true if an attribute is a bool.

pub fn is_dense_elements(&self) -> bool

Returns true if an attribute is dense elements.

pub fn is_dense_integer_elements(&self) -> bool

Returns true if an attribute is dense integer elements.

pub fn is_dense_float_elements(&self) -> bool

Returns true if an attribute is dense float elements.

pub fn is_dictionary(&self) -> bool

Returns true if an attribute is a dictionary.

pub fn is_elements(&self) -> bool

Returns true if an attribute is elements.

pub fn is_float(&self) -> bool

Returns true if an attribute is a float.

pub fn is_integer(&self) -> bool

Returns true if an attribute is an integer.

pub fn is_integer_set(&self) -> bool

Returns true if an attribute is an integer set.

pub fn is_opaque(&self) -> bool

Returns true if an attribute is opaque.

pub fn is_opaque_elements(&self) -> bool

Returns true if an attribute is opaque elements.

pub fn is_sparse_elements(&self) -> bool

Returns true if an attribute is sparse elements.

pub fn is_string(&self) -> bool

Returns true if an attribute is a string.

pub fn is_symbol(&self) -> bool

Returns true if an attribute is a symbol.

pub fn is_type(&self) -> bool

Returns true if an attribute is a type.

pub fn is_unit(&self) -> bool

Returns true if an attribute is a unit.

pub fn context(&self) -> ContextRef<'c>

Gets a context.

pub fn dump(&self)

Dumps an attribute.