Trait melior::ir::TypeLike

pub trait TypeLike<'c> {

    // Required method
    fn to_raw(&self) -> MlirType;

    // Provided methods
    fn context(&self) -> ContextRef<'c> { ... }
    fn id(&self) -> TypeId { ... }
    fn dump(&self) { ... }
    fn is_any_quantized_type(&self) -> bool { ... }
    fn is_bf16(&self) -> bool { ... }
    fn is_calibrated_quantized_type(&self) -> bool { ... }
    fn is_complex(&self) -> bool { ... }
    fn is_f16(&self) -> bool { ... }
    fn is_f32(&self) -> bool { ... }
    fn is_f64(&self) -> bool { ... }
    fn is_float8e4m3fn(&self) -> bool { ... }
    fn is_float8e5m2(&self) -> bool { ... }
    fn is_function(&self) -> bool { ... }
    fn is_index(&self) -> bool { ... }
    fn is_integer(&self) -> bool { ... }
    fn is_mem_ref(&self) -> bool { ... }
    fn is_none(&self) -> bool { ... }
    fn is_opaque(&self) -> bool { ... }
    fn is_pdl_attribute_type(&self) -> bool { ... }
    fn is_pdl_operation_type(&self) -> bool { ... }
    fn is_pdl_range_type(&self) -> bool { ... }
    fn is_pdl_type(&self) -> bool { ... }
    fn is_pdl_type_type(&self) -> bool { ... }
    fn is_pdl_value_type(&self) -> bool { ... }
    fn is_quantized_type(&self) -> bool { ... }
    fn is_ranked_tensor(&self) -> bool { ... }
    fn is_shaped(&self) -> bool { ... }
    fn is_tensor(&self) -> bool { ... }
    fn is_transform_any_op_type(&self) -> bool { ... }
    fn is_transform_operation_type(&self) -> bool { ... }
    fn is_tuple(&self) -> bool { ... }
    fn is_uniform_quantized_per_axis_type(&self) -> bool { ... }
    fn is_uniform_quantized_type(&self) -> bool { ... }
    fn is_unranked_mem_ref(&self) -> bool { ... }
    fn is_unranked_tensor(&self) -> bool { ... }
    fn is_vector(&self) -> bool { ... }

}

Trait for type-like types.

Required Methods

fn to_raw(&self) -> MlirType

Converts a type into a raw type.

Provided Methods

fn context(&self) -> ContextRef<'c>

Gets a context.

fn id(&self) -> TypeId

Gets an ID.

fn dump(&self)

Dumps a type.

fn is_any_quantized_type(&self) -> bool

Returns true if a type is any quantized type.

fn is_bf16(&self) -> bool

Returns true if a type is bf16.

fn is_calibrated_quantized_type(&self) -> bool

Returns true if a type is calibrated quantized type.

fn is_complex(&self) -> bool

Returns true if a type is complex.

fn is_f16(&self) -> bool

Returns true if a type is f16.

fn is_f32(&self) -> bool

Returns true if a type is f32.

fn is_f64(&self) -> bool

Returns true if a type is f64.

fn is_float8e4m3fn(&self) -> bool

Returns true if a type is float8e4m3fn.

fn is_float8e5m2(&self) -> bool

Returns true if a type is float8e5m2.

fn is_function(&self) -> bool

Returns true if a type is function.

fn is_index(&self) -> bool

Returns true if a type is index.

fn is_integer(&self) -> bool

Returns true if a type is integer.

fn is_mem_ref(&self) -> bool

Returns true if a type is mem ref.

fn is_none(&self) -> bool

Returns true if a type is none.

fn is_opaque(&self) -> bool

Returns true if a type is opaque.

fn is_pdl_attribute_type(&self) -> bool

Returns true if a type is pdl attribute type.

fn is_pdl_operation_type(&self) -> bool

Returns true if a type is pdl operation type.

fn is_pdl_range_type(&self) -> bool

Returns true if a type is pdl range type.

fn is_pdl_type(&self) -> bool

Returns true if a type is pdl type.

fn is_pdl_type_type(&self) -> bool

Returns true if a type is pdl type type.

fn is_pdl_value_type(&self) -> bool

Returns true if a type is pdl value type.

fn is_quantized_type(&self) -> bool

Returns true if a type is quantized type.

fn is_ranked_tensor(&self) -> bool

Returns true if a type is ranked tensor.

fn is_shaped(&self) -> bool

Returns true if a type is shaped.

fn is_tensor(&self) -> bool

Returns true if a type is tensor.

fn is_transform_any_op_type(&self) -> bool

Returns true if a type is transform any op type.

fn is_transform_operation_type(&self) -> bool

Returns true if a type is transform operation type.

fn is_tuple(&self) -> bool

Returns true if a type is tuple.

fn is_uniform_quantized_per_axis_type(&self) -> bool

Returns true if a type is uniform quantized per axis type.

fn is_uniform_quantized_type(&self) -> bool

Returns true if a type is uniform quantized type.

fn is_unranked_mem_ref(&self) -> bool

Returns true if a type is unranked mem ref.

fn is_unranked_tensor(&self) -> bool

Returns true if a type is unranked tensor.

fn is_vector(&self) -> bool

Returns true if a type is vector.