Module melior::pass::conversion

Dialect conversion passes.

Functions

convert_arithmetic_to_llvm

Creates a pass to convert the arith dialect to the llvm dialect.

convert_func_to_llvm

Creates a pass to convert the func dialect to the llvm dialect.

convert_math_to_libm

Creates a pass to convert the math dialect to the libm dialect.

convert_math_to_llvm

Creates a pass to convert the math dialect to the llvm dialect.

convert_math_to_spirv

Creates a pass to convert the math dialect to the spirv dialect.

convert_scf_to_llvm

Creates a pass to convert the cf dialect to the llvm dialect.

convert_scf_to_spirv

Creates a pass to convert the cf dialect to the spirv dialect.