Struct melior::Context

pub struct Context { /* private fields */ }

A context of IR, dialects, and passes.

Contexts own various objects, such as types, locations, and dialect instances.

Implementations

impl Context

pub fn new() -> Self

Creates a context.

pub fn registered_dialect_count(&self) -> usize

Gets a number of registered dialects.

pub fn loaded_dialect_count(&self) -> usize

Gets a number of loaded dialects.

pub fn get_or_load_dialect(&self, name: &str) -> Dialect<'_>

Gets or loads a dialect.

pub fn append_dialect_registry(&self, registry: &DialectRegistry)

Appends a dialect registry.

pub fn load_all_available_dialects(&self)

Loads all available dialects.

pub fn enable_multi_threading(&self, enabled: bool)

Enables multi-threading.

pub fn allow_unregistered_dialects(&self) -> bool

Returns true if unregistered dialects are allowed.

pub fn set_allow_unregistered_dialects(&self, allowed: bool)

Sets if unregistered dialects are allowed.

pub fn is_registered_operation(&self, name: &str) -> bool

Returns true if a given operation is registered in a context.

pub const fn to_raw(&self) -> MlirContext

Converts a context into a raw object.

pub fn attach_diagnostic_handler<F: FnMut(Diagnostic<'_>) -> bool>( &self, handler: F ) -> DiagnosticHandlerId

Attaches a diagnostic handler.

pub fn detach_diagnostic_handler(&self, id: DiagnosticHandlerId)

Detaches a diagnostic handler.