Module melior::dialect::ods::amdgpu

amdgpu dialect.

The AMDGPU dialect provides wrappers around AMD-specific functionality and LLVM intrinsics. These wrappers should be used in conjunction with more generic dialects, such as gpu and vector, when generating LLVM IR that will eventually be executed on AMD hardware.

Structs

LDSBarrierOp

lds_barrier operation. Barrier that includes a wait for LDS memory operations..

LDSBarrierOpBuilder

Builder for lds_barrier operation. Barrier that includes a wait for LDS memory operations..

MFMAOp

mfma operation. MLIR wrapper for CDNA mfma instructions.

MFMAOpBuilder

Builder for mfma operation. MLIR wrapper for CDNA mfma instructions.

RawBufferAtomicCmpswapOp

raw_buffer_atomic_cmpswap operation. Raw Buffer Atomic compare-and-swap.

RawBufferAtomicCmpswapOpBuilder

Builder for raw_buffer_atomic_cmpswap operation. Raw Buffer Atomic compare-and-swap.

RawBufferAtomicFaddOp

raw_buffer_atomic_fadd operation. Raw Buffer Floating-point Atomic Add (MI-* only).

RawBufferAtomicFaddOpBuilder

Builder for raw_buffer_atomic_fadd operation. Raw Buffer Floating-point Atomic Add (MI-* only).

RawBufferAtomicFmaxOp

raw_buffer_atomic_fmax operation. Raw Buffer Floating-point Atomic Max (non-GFX9).

RawBufferAtomicFmaxOpBuilder

Builder for raw_buffer_atomic_fmax operation. Raw Buffer Floating-point Atomic Max (non-GFX9).

RawBufferAtomicSmaxOp

raw_buffer_atomic_smax operation. Raw Buffer Signed Integer Atomic Max.

RawBufferAtomicSmaxOpBuilder

Builder for raw_buffer_atomic_smax operation. Raw Buffer Signed Integer Atomic Max.

RawBufferAtomicUminOp

raw_buffer_atomic_umin operation. Raw Buffer Unsigned Integer Atomic Min.

RawBufferAtomicUminOpBuilder

Builder for raw_buffer_atomic_umin operation. Raw Buffer Unsigned Integer Atomic Min.

RawBufferLoadOp

raw_buffer_load operation. Raw Buffer load, exposing GCN features.

RawBufferLoadOpBuilder

Builder for raw_buffer_load operation. Raw Buffer load, exposing GCN features.

RawBufferStoreOp

raw_buffer_store operation. Raw Buffer Store, exposing GCN features.

RawBufferStoreOpBuilder

Builder for raw_buffer_store operation. Raw Buffer Store, exposing GCN features.

WMMAOp

wmma operation. MLIR wrapper for RDNA3 wmma instructions.

WMMAOpBuilder

Builder for wmma operation. MLIR wrapper for RDNA3 wmma instructions.

Functions

lds_barrier

Creates a new lds_barrier operation. Barrier that includes a wait for LDS memory operations..

mfma

Creates a new mfma operation. MLIR wrapper for CDNA mfma instructions.

raw_buffer_atomic_cmpswap

Creates a new raw_buffer_atomic_cmpswap operation. Raw Buffer Atomic compare-and-swap.

raw_buffer_atomic_fadd

Creates a new raw_buffer_atomic_fadd operation. Raw Buffer Floating-point Atomic Add (MI-* only).

raw_buffer_atomic_fmax

Creates a new raw_buffer_atomic_fmax operation. Raw Buffer Floating-point Atomic Max (non-GFX9).

raw_buffer_atomic_smax

Creates a new raw_buffer_atomic_smax operation. Raw Buffer Signed Integer Atomic Max.

raw_buffer_atomic_umin

Creates a new raw_buffer_atomic_umin operation. Raw Buffer Unsigned Integer Atomic Min.

raw_buffer_load

Creates a new raw_buffer_load operation. Raw Buffer load, exposing GCN features.

raw_buffer_store

Creates a new raw_buffer_store operation. Raw Buffer Store, exposing GCN features.

wmma

Creates a new wmma operation. MLIR wrapper for RDNA3 wmma instructions.