Module melior::dialect::ods::arith

arith dialect.

The arith dialect is intended to hold basic integer and floating point mathematical operations. This includes unary, binary, and ternary arithmetic ops, bitwise and shift ops, cast ops, and compare ops. Operations in this dialect also accept vectors and tensors of integers or floats.

Structs

AddFOp

addf operation: Floating point addition operation

AddFOpBuilder

Builder for addf operation: Floating point addition operation

AddIOp

addi operation: Integer addition operation

AddIOpBuilder

Builder for addi operation: Integer addition operation

AddUIExtendedOp

addui_extended operation: extended unsigned integer addition operation returning sum and overflow bit

AddUIExtendedOpBuilder

Builder for addui_extended operation: extended unsigned integer addition operation returning sum and overflow bit

AndIOp

andi operation: Integer binary and

AndIOpBuilder

Builder for andi operation: Integer binary and

BitcastOp

bitcast operation: Bitcast between values of equal bit width

BitcastOpBuilder

Builder for bitcast operation: Bitcast between values of equal bit width

CeilDivSIOp

ceildivsi operation: Signed ceil integer division operation

CeilDivSIOpBuilder

Builder for ceildivsi operation: Signed ceil integer division operation

CeilDivUIOp

ceildivui operation: Unsigned ceil integer division operation

CeilDivUIOpBuilder

Builder for ceildivui operation: Unsigned ceil integer division operation

CmpFOp

cmpf operation: Floating-point comparison operation

CmpFOpBuilder

Builder for cmpf operation: Floating-point comparison operation

CmpIOp

cmpi operation: Integer comparison operation

CmpIOpBuilder

Builder for cmpi operation: Integer comparison operation

ConstantOp

constant operation: Integer or floating point constant

ConstantOpBuilder

Builder for constant operation: Integer or floating point constant

DivFOp

divf operation: Floating point division operation

DivFOpBuilder

Builder for divf operation: Floating point division operation

DivSIOp

divsi operation: Signed integer division operation

DivSIOpBuilder

Builder for divsi operation: Signed integer division operation

DivUIOp

divui operation: Unsigned integer division operation

DivUIOpBuilder

Builder for divui operation: Unsigned integer division operation

ExtFOp

extf operation: Cast from floating-point to wider floating-point

ExtFOpBuilder

Builder for extf operation: Cast from floating-point to wider floating-point

ExtSIOp

extsi operation: Integer sign extension operation

ExtSIOpBuilder

Builder for extsi operation: Integer sign extension operation

ExtUIOp

extui operation: Integer zero extension operation

ExtUIOpBuilder

Builder for extui operation: Integer zero extension operation

FPToSIOp

fptosi operation: Cast from floating-point type to integer type

FPToSIOpBuilder

Builder for fptosi operation: Cast from floating-point type to integer type

FPToUIOp

fptoui operation: Cast from floating-point type to integer type

FPToUIOpBuilder

Builder for fptoui operation: Cast from floating-point type to integer type

FloorDivSIOp

floordivsi operation: Signed floor integer division operation

FloorDivSIOpBuilder

Builder for floordivsi operation: Signed floor integer division operation

IndexCastOp

index_cast operation: Cast between index and integer types

IndexCastOpBuilder

Builder for index_cast operation: Cast between index and integer types

IndexCastUIOp

index_castui operation: Unsigned cast between index and integer types

IndexCastUIOpBuilder

Builder for index_castui operation: Unsigned cast between index and integer types

MaxFOp

maxf operation: Floating-point maximum operation

MaxFOpBuilder

Builder for maxf operation: Floating-point maximum operation

MaxSIOp

maxsi operation: Signed integer maximum operation

MaxSIOpBuilder

Builder for maxsi operation: Signed integer maximum operation

MaxUIOp

maxui operation: Unsigned integer maximum operation

MaxUIOpBuilder

Builder for maxui operation: Unsigned integer maximum operation

MinFOp

minf operation: Floating-point minimum operation

MinFOpBuilder

Builder for minf operation: Floating-point minimum operation

MinSIOp

minsi operation: Signed integer minimum operation

MinSIOpBuilder

Builder for minsi operation: Signed integer minimum operation

MinUIOp

minui operation: Unsigned integer minimum operation

MinUIOpBuilder

Builder for minui operation: Unsigned integer minimum operation

MulFOp

mulf operation: Floating point multiplication operation

MulFOpBuilder

Builder for mulf operation: Floating point multiplication operation

MulIOp

muli operation: Integer multiplication operation

MulIOpBuilder

Builder for muli operation: Integer multiplication operation

MulSIExtendedOp

mulsi_extended operation: extended signed integer multiplication operation

MulSIExtendedOpBuilder

Builder for mulsi_extended operation: extended signed integer multiplication operation

MulUIExtendedOp

mului_extended operation: extended unsigned integer multiplication operation

MulUIExtendedOpBuilder

Builder for mului_extended operation: extended unsigned integer multiplication operation

NegFOp

negf operation: Floating point negation

NegFOpBuilder

Builder for negf operation: Floating point negation

OrIOp

ori operation: Integer binary or

OrIOpBuilder

Builder for ori operation: Integer binary or

RemFOp

remf operation: Floating point division remainder operation

RemFOpBuilder

Builder for remf operation: Floating point division remainder operation

RemSIOp

remsi operation: Signed integer division remainder operation

RemSIOpBuilder

Builder for remsi operation: Signed integer division remainder operation

RemUIOp

remui operation: Unsigned integer division remainder operation

RemUIOpBuilder

Builder for remui operation: Unsigned integer division remainder operation

SIToFPOp

sitofp operation: Cast from integer type to floating-point

SIToFPOpBuilder

Builder for sitofp operation: Cast from integer type to floating-point

SelectOp

select operation: Select operation

SelectOpBuilder

Builder for select operation: Select operation

ShLIOp

shli operation: Integer left-shift

ShLIOpBuilder

Builder for shli operation: Integer left-shift

ShRSIOp

shrsi operation: Signed integer right-shift

ShRSIOpBuilder

Builder for shrsi operation: Signed integer right-shift

ShRUIOp

shrui operation: Unsigned integer right-shift

ShRUIOpBuilder

Builder for shrui operation: Unsigned integer right-shift

SubFOp

subf operation: Floating point subtraction operation

SubFOpBuilder

Builder for subf operation: Floating point subtraction operation

SubIOp

subi operation: Integer subtraction operation

SubIOpBuilder

Builder for subi operation: Integer subtraction operation

TruncFOp

truncf operation: Cast from floating-point to narrower floating-point

TruncFOpBuilder

Builder for truncf operation: Cast from floating-point to narrower floating-point

TruncIOp

trunci operation: Integer truncation operation

TruncIOpBuilder

Builder for trunci operation: Integer truncation operation

UIToFPOp

uitofp operation: Cast from unsigned integer type to floating-point

UIToFPOpBuilder

Builder for uitofp operation: Cast from unsigned integer type to floating-point

XOrIOp

xori operation: Integer binary xor

XOrIOpBuilder

Builder for xori operation: Integer binary xor

Functions

addf

Creates a new addf operation: Floating point addition operation

addi

Creates a new addi operation: Integer addition operation

addui_extended

Creates a new addui_extended operation: extended unsigned integer addition operation returning sum and overflow bit

andi

Creates a new andi operation: Integer binary and

bitcast

Creates a new bitcast operation: Bitcast between values of equal bit width

ceildivsi

Creates a new ceildivsi operation: Signed ceil integer division operation

ceildivui

Creates a new ceildivui operation: Unsigned ceil integer division operation

cmpf

Creates a new cmpf operation: Floating-point comparison operation

cmpi

Creates a new cmpi operation: Integer comparison operation

constant

Creates a new constant operation: Integer or floating point constant

divf

Creates a new divf operation: Floating point division operation

divsi

Creates a new divsi operation: Signed integer division operation

divui

Creates a new divui operation: Unsigned integer division operation

extf

Creates a new extf operation: Cast from floating-point to wider floating-point

extsi

Creates a new extsi operation: Integer sign extension operation

extui

Creates a new extui operation: Integer zero extension operation

floordivsi

Creates a new floordivsi operation: Signed floor integer division operation

fptosi

Creates a new fptosi operation: Cast from floating-point type to integer type

fptoui

Creates a new fptoui operation: Cast from floating-point type to integer type

index_cast

Creates a new index_cast operation: Cast between index and integer types

index_castui

Creates a new index_castui operation: Unsigned cast between index and integer types

maxf

Creates a new maxf operation: Floating-point maximum operation

maxsi

Creates a new maxsi operation: Signed integer maximum operation

maxui

Creates a new maxui operation: Unsigned integer maximum operation

minf

Creates a new minf operation: Floating-point minimum operation

minsi

Creates a new minsi operation: Signed integer minimum operation

minui

Creates a new minui operation: Unsigned integer minimum operation

mulf

Creates a new mulf operation: Floating point multiplication operation

muli

Creates a new muli operation: Integer multiplication operation

mulsi_extended

Creates a new mulsi_extended operation: extended signed integer multiplication operation

mului_extended

Creates a new mului_extended operation: extended unsigned integer multiplication operation

negf

Creates a new negf operation: Floating point negation

ori

Creates a new ori operation: Integer binary or

remf

Creates a new remf operation: Floating point division remainder operation

remsi

Creates a new remsi operation: Signed integer division remainder operation

remui

Creates a new remui operation: Unsigned integer division remainder operation

select

Creates a new select operation: Select operation

shli

Creates a new shli operation: Integer left-shift

shrsi

Creates a new shrsi operation: Signed integer right-shift

shrui

Creates a new shrui operation: Unsigned integer right-shift

sitofp

Creates a new sitofp operation: Cast from integer type to floating-point

subf

Creates a new subf operation: Floating point subtraction operation

subi

Creates a new subi operation: Integer subtraction operation

truncf

Creates a new truncf operation: Cast from floating-point to narrower floating-point

trunci

Creates a new trunci operation: Integer truncation operation

uitofp

Creates a new uitofp operation: Cast from unsigned integer type to floating-point

xori

Creates a new xori operation: Integer binary xor