Module melior::dialect::ods::async

async dialect.

This dialect contains operations for modeling asynchronous execution.

Structs

AddToGroupOp

add_to_group operation: Adds and async token or value to the group

AddToGroupOpBuilder

Builder for add_to_group operation: Adds and async token or value to the group

AwaitAllOp

await_all operation: Waits for the all async tokens or values in the group to become ready

AwaitAllOpBuilder

Builder for await_all operation: Waits for the all async tokens or values in the group to become ready

AwaitOp

await operation: Waits for the argument to become ready

AwaitOpBuilder

Builder for await operation: Waits for the argument to become ready

CallOp

call operation: Async call operation

CallOpBuilder

Builder for call operation: Async call operation

CoroBeginOp

coro.begin operation: Returns a handle to the coroutine

CoroBeginOpBuilder

Builder for coro.begin operation: Returns a handle to the coroutine

CoroEndOp

coro.end operation: Marks the end of the coroutine in the suspend block

CoroEndOpBuilder

Builder for coro.end operation: Marks the end of the coroutine in the suspend block

CoroFreeOp

coro.free operation: Deallocates the coroutine frame

CoroFreeOpBuilder

Builder for coro.free operation: Deallocates the coroutine frame

CoroIdOp

coro.id operation: Returns a switched-resume coroutine identifier

CoroIdOpBuilder

Builder for coro.id operation: Returns a switched-resume coroutine identifier

CoroSaveOp

coro.save operation: Saves the coroutine state

CoroSaveOpBuilder

Builder for coro.save operation: Saves the coroutine state

CoroSuspendOp

coro.suspend operation: Suspends the coroutine

CoroSuspendOpBuilder

Builder for coro.suspend operation: Suspends the coroutine

CreateGroupOp

create_group operation: Creates an empty async group

CreateGroupOpBuilder

Builder for create_group operation: Creates an empty async group

ExecuteOp

execute operation: Asynchronous execute operation

ExecuteOpBuilder

Builder for execute operation: Asynchronous execute operation

FuncOp

func operation: Async function operation

FuncOpBuilder

Builder for func operation: Async function operation

ReturnOp

return operation: Async function return operation

ReturnOpBuilder

Builder for return operation: Async function return operation

RuntimeAddRefOp

runtime.add_ref operation: Adds a reference to async value

RuntimeAddRefOpBuilder

Builder for runtime.add_ref operation: Adds a reference to async value

RuntimeAddToGroupOp

runtime.add_to_group operation: Adds and async token or value to the group

RuntimeAddToGroupOpBuilder

Builder for runtime.add_to_group operation: Adds and async token or value to the group

RuntimeAwaitAndResumeOp

runtime.await_and_resume operation: Awaits the async operand and resumes the coroutine

RuntimeAwaitAndResumeOpBuilder

Builder for runtime.await_and_resume operation: Awaits the async operand and resumes the coroutine

RuntimeAwaitOp

runtime.await operation: Blocks the caller thread until the operand becomes available

RuntimeAwaitOpBuilder

Builder for runtime.await operation: Blocks the caller thread until the operand becomes available

RuntimeCreateGroupOp

runtime.create_group operation: Creates an async runtime group

RuntimeCreateGroupOpBuilder

Builder for runtime.create_group operation: Creates an async runtime group

RuntimeCreateOp

runtime.create operation: Creates an async runtime token or value

RuntimeCreateOpBuilder

Builder for runtime.create operation: Creates an async runtime token or value

RuntimeDropRefOp

runtime.drop_ref operation: Drops a reference to async value

RuntimeDropRefOpBuilder

Builder for runtime.drop_ref operation: Drops a reference to async value

RuntimeIsErrorOp

runtime.is_error operation: Returns true if token, value or group is in error state

RuntimeIsErrorOpBuilder

Builder for runtime.is_error operation: Returns true if token, value or group is in error state

RuntimeLoadOp

runtime.load operation: Loads the value from the runtime async.value

RuntimeLoadOpBuilder

Builder for runtime.load operation: Loads the value from the runtime async.value

RuntimeNumWorkerThreadsOp

runtime.num_worker_threads operation: Gets the number of threads in the threadpool from the runtime

RuntimeNumWorkerThreadsOpBuilder

Builder for runtime.num_worker_threads operation: Gets the number of threads in the threadpool from the runtime

RuntimeResumeOp

runtime.resume operation: Resumes the coroutine on a thread managed by the runtime

RuntimeResumeOpBuilder

Builder for runtime.resume operation: Resumes the coroutine on a thread managed by the runtime

RuntimeSetAvailableOp

runtime.set_available operation: Switches token or value to available state

RuntimeSetAvailableOpBuilder

Builder for runtime.set_available operation: Switches token or value to available state

RuntimeSetErrorOp

runtime.set_error operation: Switches token or value to error state

RuntimeSetErrorOpBuilder

Builder for runtime.set_error operation: Switches token or value to error state

RuntimeStoreOp

runtime.store operation: Stores the value into the runtime async.value

RuntimeStoreOpBuilder

Builder for runtime.store operation: Stores the value into the runtime async.value

YieldOp

yield operation: Terminator for Async execute operation

YieldOpBuilder

Builder for yield operation: Terminator for Async execute operation

Functions

add_to_group

Creates a new add_to_group operation: Adds and async token or value to the group

await

Creates a new await operation: Waits for the argument to become ready

await_all

Creates a new await_all operation: Waits for the all async tokens or values in the group to become ready

call

Creates a new call operation: Async call operation

coro_begin

Creates a new coro.begin operation: Returns a handle to the coroutine

coro_end

Creates a new coro.end operation: Marks the end of the coroutine in the suspend block

coro_free

Creates a new coro.free operation: Deallocates the coroutine frame

coro_id

Creates a new coro.id operation: Returns a switched-resume coroutine identifier

coro_save

Creates a new coro.save operation: Saves the coroutine state

coro_suspend

Creates a new coro.suspend operation: Suspends the coroutine

create_group

Creates a new create_group operation: Creates an empty async group

execute

Creates a new execute operation: Asynchronous execute operation

func

Creates a new func operation: Async function operation

return

Creates a new return operation: Async function return operation

runtime_add_ref

Creates a new runtime.add_ref operation: Adds a reference to async value

runtime_add_to_group

Creates a new runtime.add_to_group operation: Adds and async token or value to the group

runtime_await

Creates a new runtime.await operation: Blocks the caller thread until the operand becomes available

runtime_await_and_resume

Creates a new runtime.await_and_resume operation: Awaits the async operand and resumes the coroutine

runtime_create

Creates a new runtime.create operation: Creates an async runtime token or value

runtime_create_group

Creates a new runtime.create_group operation: Creates an async runtime group

runtime_drop_ref

Creates a new runtime.drop_ref operation: Drops a reference to async value

runtime_is_error

Creates a new runtime.is_error operation: Returns true if token, value or group is in error state

runtime_load

Creates a new runtime.load operation: Loads the value from the runtime async.value

runtime_num_worker_threads

Creates a new runtime.num_worker_threads operation: Gets the number of threads in the threadpool from the runtime

runtime_resume

Creates a new runtime.resume operation: Resumes the coroutine on a thread managed by the runtime

runtime_set_available

Creates a new runtime.set_available operation: Switches token or value to available state

runtime_set_error

Creates a new runtime.set_error operation: Switches token or value to error state

runtime_store

Creates a new runtime.store operation: Stores the value into the runtime async.value

yield

Creates a new yield operation: Terminator for Async execute operation