Module melior::dialect::ods::bufferization

bufferization dialect.

Bufferization in MLIR is the process of converting the tensor type to the memref type. Simply put, bufferization is the process of converting computations on the mathematical tensor construct to computations on physical memory buffers. The bufferization dialect contains operations/interfaces specific to the bufferization passes.

An overview of the bufferization infrastructure and important conceptual details related to using the MLIR dialect conversion infrastructure can be found in bufferization and buffer deallocation.

Structs

AllocTensorOp

alloc_tensor operation. Allocate buffer for a tensor.

AllocTensorOpBuilder

Builder for alloc_tensor operation. Allocate buffer for a tensor.

CloneOp

clone operation. Clone a memref.

CloneOpBuilder

Builder for clone operation. Clone a memref.

CopyTensorOp

copy_tensor operation. Copy a tensor.

CopyTensorOpBuilder

Builder for copy_tensor operation. Copy a tensor.

DeallocOp

dealloc operation. Deallocates the given memrefs if no alias is retained.

DeallocOpBuilder

Builder for dealloc operation. Deallocates the given memrefs if no alias is retained.

DeallocTensorOp

dealloc_tensor operation. Release underlying storage format of given tensor.

DeallocTensorOpBuilder

Builder for dealloc_tensor operation. Release underlying storage format of given tensor.

ToMemrefOp

to_memref operation. Cast a tensor to memref.

ToMemrefOpBuilder

Builder for to_memref operation. Cast a tensor to memref.

ToTensorOp

to_tensor operation. Create a tensor from a memref.

ToTensorOpBuilder

Builder for to_tensor operation. Create a tensor from a memref.

Functions

alloc_tensor

Creates a new alloc_tensor operation. Allocate buffer for a tensor.

clone

Creates a new clone operation. Clone a memref.

copy_tensor

Creates a new copy_tensor operation. Copy a tensor.

dealloc

Creates a new dealloc operation. Deallocates the given memrefs if no alias is retained.

dealloc_tensor

Creates a new dealloc_tensor operation. Release underlying storage format of given tensor.

to_memref

Creates a new to_memref operation. Cast a tensor to memref.

to_tensor

Creates a new to_tensor operation. Create a tensor from a memref.