Module melior::dialect::ods::gpu

gpu dialect.

Structs

AllReduceOp

all_reduce operation. Reduce values among workgroup..

AllReduceOpBuilder

Builder for all_reduce operation. Reduce values among workgroup..

AllocOp

alloc operation. GPU memory allocation operation..

AllocOpBuilder

Builder for alloc operation. GPU memory allocation operation..

BarrierOp

barrier operation. Synchronizes all work items of a workgroup..

BarrierOpBuilder

Builder for barrier operation. Synchronizes all work items of a workgroup..

BlockDimOp

block_dim operation.

BlockDimOpBuilder

Builder for block_dim operation.

BlockIdOp

block_id operation.

BlockIdOpBuilder

Builder for block_id operation.

Create2To4SpMatOp

create_2to4_spmat operation. Create sparse matrix with 2:4 sparsity operation.

Create2To4SpMatOpBuilder

Builder for create_2to4_spmat operation. Create sparse matrix with 2:4 sparsity operation.

CreateCooAoSOp

create_coo_aos operation. Create sparse matrix in COO format operation (AoS).

CreateCooAoSOpBuilder

Builder for create_coo_aos operation. Create sparse matrix in COO format operation (AoS).

CreateCooOp

create_coo operation. Create sparse matrix in COO format operation.

CreateCooOpBuilder

Builder for create_coo operation. Create sparse matrix in COO format operation.

CreateCsrOp

create_csr operation. Create sparse matrix in CSR format operation.

CreateCsrOpBuilder

Builder for create_csr operation. Create sparse matrix in CSR format operation.

CreateDnTensorOp

create_dn_tensor operation. Create dense tensor operation.

CreateDnTensorOpBuilder

Builder for create_dn_tensor operation. Create dense tensor operation.

DeallocOp

dealloc operation. GPU memory deallocation operation.

DeallocOpBuilder

Builder for dealloc operation. GPU memory deallocation operation.

DestroyDnTensorOp

destroy_dn_tensor operation. Destroy dense tensor operation.

DestroyDnTensorOpBuilder

Builder for destroy_dn_tensor operation. Destroy dense tensor operation.

DestroySpMatOp

destroy_sp_mat operation. Destroy sparse matrix operation.

DestroySpMatOpBuilder

Builder for destroy_sp_mat operation. Destroy sparse matrix operation.

GPUFuncOp

func operation. Function executable on a GPU.

GPUFuncOpBuilder

Builder for func operation. Function executable on a GPU.

GPUModuleOp

module operation. A top level compilation unit containing code to be run on a GPU..

GPUModuleOpBuilder

Builder for module operation. A top level compilation unit containing code to be run on a GPU..

GlobalIdOp

global_id operation.

GlobalIdOpBuilder

Builder for global_id operation.

GridDimOp

grid_dim operation.

GridDimOpBuilder

Builder for grid_dim operation.

HostRegisterOp

host_register operation. Registers a memref for access from device..

HostRegisterOpBuilder

Builder for host_register operation. Registers a memref for access from device..

HostUnregisterOp

host_unregister operation. Unregisters a memref for access from device..

HostUnregisterOpBuilder

Builder for host_unregister operation. Unregisters a memref for access from device..

LaneIdOp

lane_id operation.

LaneIdOpBuilder

Builder for lane_id operation.

LaunchFuncOp

launch_func operation. Launches a function as a GPU kernel.

LaunchFuncOpBuilder

Builder for launch_func operation. Launches a function as a GPU kernel.

LaunchOp

launch operation. GPU kernel launch operation.

LaunchOpBuilder

Builder for launch operation. GPU kernel launch operation.

MemcpyOp

memcpy operation. GPU memcpy operation.

MemcpyOpBuilder

Builder for memcpy operation. GPU memcpy operation.

MemsetOp

memset operation. GPU memset operation.

MemsetOpBuilder

Builder for memset operation. GPU memset operation.

ModuleEndOp

module_end operation. A pseudo op that marks the end of a gpu.module..

ModuleEndOpBuilder

Builder for module_end operation. A pseudo op that marks the end of a gpu.module..

NumSubgroupsOp

num_subgroups operation.

NumSubgroupsOpBuilder

Builder for num_subgroups operation.

PrintfOp

printf operation. Device-side printf, as in CUDA or OpenCL, for debugging.

PrintfOpBuilder

Builder for printf operation. Device-side printf, as in CUDA or OpenCL, for debugging.

ReturnOp

return operation. Terminator for GPU functions..

ReturnOpBuilder

Builder for return operation. Terminator for GPU functions..

SDDMMBufferSizeOp

sddmm_buffer_size operation. Precompute buffersize for SDDMM operation.

SDDMMBufferSizeOpBuilder

Builder for sddmm_buffer_size operation. Precompute buffersize for SDDMM operation.

SDDMMOp

sddmm operation. SDDMM operation.

SDDMMOpBuilder

Builder for sddmm operation. SDDMM operation.

SetDefaultDeviceOp

set_default_device operation. Set default GPU for operations after this by index.

SetDefaultDeviceOpBuilder

Builder for set_default_device operation. Set default GPU for operations after this by index.

ShuffleOp

shuffle operation. Shuffles values within a subgroup..

ShuffleOpBuilder

Builder for shuffle operation. Shuffles values within a subgroup..

SpMMBufferSizeOp

spmm_buffer_size operation. Precompute buffersize for SpMM operation.

SpMMBufferSizeOpBuilder

Builder for spmm_buffer_size operation. Precompute buffersize for SpMM operation.

SpMMOp

spmm operation. SpMM operation.

SpMMOpBuilder

Builder for spmm operation. SpMM operation.

SpMVBufferSizeOp

spmv_buffer_size operation. Precompute buffersize for SpMV operation.

SpMVBufferSizeOpBuilder

Builder for spmv_buffer_size operation. Precompute buffersize for SpMV operation.

SpMVOp

spmv operation. SpMV operation.

SpMVOpBuilder

Builder for spmv operation. SpMV operation.

SubgroupIdOp

subgroup_id operation.

SubgroupIdOpBuilder

Builder for subgroup_id operation.

SubgroupMmaComputeOp

subgroup_mma_compute operation. GPU warp synchronous matrix multiply accumulate.

SubgroupMmaComputeOpBuilder

Builder for subgroup_mma_compute operation. GPU warp synchronous matrix multiply accumulate.

SubgroupMmaConstantMatrixOp

subgroup_mma_constant_matrix operation. GPU warp synchronous constant matrix.

SubgroupMmaConstantMatrixOpBuilder

Builder for subgroup_mma_constant_matrix operation. GPU warp synchronous constant matrix.

SubgroupMmaElementwiseOp

subgroup_mma_elementwise operation. GPU warp elementwise operation on a matrix.

SubgroupMmaElementwiseOpBuilder

Builder for subgroup_mma_elementwise operation. GPU warp elementwise operation on a matrix.

SubgroupMmaLoadMatrixOp

subgroup_mma_load_matrix operation. GPU warp synchronous matrix load.

SubgroupMmaLoadMatrixOpBuilder

Builder for subgroup_mma_load_matrix operation. GPU warp synchronous matrix load.

SubgroupMmaStoreMatrixOp

subgroup_mma_store_matrix operation. GPU warp synchronous matrix store.

SubgroupMmaStoreMatrixOpBuilder

Builder for subgroup_mma_store_matrix operation. GPU warp synchronous matrix store.

SubgroupReduceOp

subgroup_reduce operation. Reduce values among subgroup..

SubgroupReduceOpBuilder

Builder for subgroup_reduce operation. Reduce values among subgroup..

SubgroupSizeOp

subgroup_size operation.

SubgroupSizeOpBuilder

Builder for subgroup_size operation.

TerminatorOp

terminator operation. Terminator for GPU launch regions..

TerminatorOpBuilder

Builder for terminator operation. Terminator for GPU launch regions..

ThreadIdOp

thread_id operation.

ThreadIdOpBuilder

Builder for thread_id operation.

WaitOp

wait operation. Wait for async gpu ops to complete..

WaitOpBuilder

Builder for wait operation. Wait for async gpu ops to complete..

YieldOp

yield operation. GPU yield operation.

YieldOpBuilder

Builder for yield operation. GPU yield operation.

Functions

all_reduce

Creates a new all_reduce operation. Reduce values among workgroup..

alloc

Creates a new alloc operation. GPU memory allocation operation..

barrier

Creates a new barrier operation. Synchronizes all work items of a workgroup..

block_dim

Creates a new block_dim operation.

block_id

Creates a new block_id operation.

create_2_to_4_spmat

Creates a new create_2to4_spmat operation. Create sparse matrix with 2:4 sparsity operation.

create_coo

Creates a new create_coo operation. Create sparse matrix in COO format operation.

create_coo_aos

Creates a new create_coo_aos operation. Create sparse matrix in COO format operation (AoS).

create_csr

Creates a new create_csr operation. Create sparse matrix in CSR format operation.

create_dn_tensor

Creates a new create_dn_tensor operation. Create dense tensor operation.

dealloc

Creates a new dealloc operation. GPU memory deallocation operation.

destroy_dn_tensor

Creates a new destroy_dn_tensor operation. Destroy dense tensor operation.

destroy_sp_mat

Creates a new destroy_sp_mat operation. Destroy sparse matrix operation.

func

Creates a new func operation. Function executable on a GPU.

global_id

Creates a new global_id operation.

grid_dim

Creates a new grid_dim operation.

host_register

Creates a new host_register operation. Registers a memref for access from device..

host_unregister

Creates a new host_unregister operation. Unregisters a memref for access from device..

lane_id

Creates a new lane_id operation.

launch

Creates a new launch operation. GPU kernel launch operation.

launch_func

Creates a new launch_func operation. Launches a function as a GPU kernel.

memcpy

Creates a new memcpy operation. GPU memcpy operation.

memset

Creates a new memset operation. GPU memset operation.

module

Creates a new module operation. A top level compilation unit containing code to be run on a GPU..

module_end

Creates a new module_end operation. A pseudo op that marks the end of a gpu.module..

num_subgroups

Creates a new num_subgroups operation.

printf

Creates a new printf operation. Device-side printf, as in CUDA or OpenCL, for debugging.

return

Creates a new return operation. Terminator for GPU functions..

sddmm

Creates a new sddmm operation. SDDMM operation.

sddmm_buffer_size

Creates a new sddmm_buffer_size operation. Precompute buffersize for SDDMM operation.

set_default_device

Creates a new set_default_device operation. Set default GPU for operations after this by index.

shuffle

Creates a new shuffle operation. Shuffles values within a subgroup..

spmm

Creates a new spmm operation. SpMM operation.

spmm_buffer_size

Creates a new spmm_buffer_size operation. Precompute buffersize for SpMM operation.

spmv

Creates a new spmv operation. SpMV operation.

spmv_buffer_size

Creates a new spmv_buffer_size operation. Precompute buffersize for SpMV operation.

subgroup_id

Creates a new subgroup_id operation.

subgroup_mma_compute

Creates a new subgroup_mma_compute operation. GPU warp synchronous matrix multiply accumulate.

subgroup_mma_constant_matrix

Creates a new subgroup_mma_constant_matrix operation. GPU warp synchronous constant matrix.

subgroup_mma_elementwise

Creates a new subgroup_mma_elementwise operation. GPU warp elementwise operation on a matrix.

subgroup_mma_load_matrix

Creates a new subgroup_mma_load_matrix operation. GPU warp synchronous matrix load.

subgroup_mma_store_matrix

Creates a new subgroup_mma_store_matrix operation. GPU warp synchronous matrix store.

subgroup_reduce

Creates a new subgroup_reduce operation. Reduce values among subgroup..

subgroup_size

Creates a new subgroup_size operation.

terminator

Creates a new terminator operation. Terminator for GPU launch regions..

thread_id

Creates a new thread_id operation.

wait

Creates a new wait operation. Wait for async gpu ops to complete..

yield

Creates a new yield operation. GPU yield operation.