Module melior::dialect::ods::index

index dialect.

The Index dialect contains operations for manipulating values of the builtin index type. The index type models target-specific values of pointer width, like intptr_t. Index values are typically used as loop bounds, array subscripts, tensor dimensions, etc.

The operations in this dialect operate exclusively on scalar index types. The dialect and its operations treat the index type as signless and contains signed and unsigned versions of certain operations where the distinction is meaningful. In particular, the operations and transformations are careful to be aware of the target-independent-ness of the index type, such as when folding.

The folding semantics of the Index dialect operations ensure that folding produces the same results irrespective of the eventual target pointer width. All index constants are stored in APInts of maximum index bitwidth: 64. Operations are folded using 64-bit integer arithmetic.

For operations where the values of the upper 32 bits don’t impact the values of the lower 32 bits, no additional handling is required because if the target is 32-bit, the truncated folded result will be the same as if the operation were computed with 32-bit arithmetic, and if the target is 64-bit, the fold result is valid by default.

Consider addition: an overflow in 32-bit is the same as truncating the result computed in 64-bit. For example, add(0x800000008, 0x800000008) is 0x1000000010 in 64-bit, which truncates to 0x10, the same result as truncating the operands first: add(0x08, 0x08). Specifically, an operation f can always be folded if it satisfies the following for all 64-bit values of a and b:

trunc(f(a, b)) = f(trunc(a), trunc(b))

When materializing target-specific code, constants just need to be truncated as appropriate.

Operations where the values of the upper 32 bits do impact the values of the lower 32 bits are not folded if the results would be different in 32-bit. These are operations that right shift – division, remainder, etc. These operations are only folded for subsets of a and b for which the above property is satisfied. This is checked per fold attempt.

Consider division: the 32-bit computation will differ from 64-bit if the latter results in a high bit shifted into the lower 32 bits. For example, div(0x100000002, 2) is 0x80000001 in 64-bit but 0x01 in 32-bit; it cannot be folded. However, div(0x200000002, 2) can be folded. The 64-bit result is 0x100000001, which truncated to 32 bits is 0x01. The 32-bit result of the operation with truncated operands div(0x02, 2) which is 0x01, the same as truncating the 64-bit result.

Structs

AddOperation

An add operation. Index addition.

AddOperationBuilder

A builder for an add operation.

AndOperation

An and operation. Index bitwise and.

AndOperationBuilder

A builder for an and operation.

BoolConstantOperation

A bool.constant operation. Boolean constant.

BoolConstantOperationBuilder

A builder for a bool.constant operation.

CastSOperation

A casts operation. Index signed cast.

CastSOperationBuilder

A builder for a casts operation.

CastUOperation

A castu operation. Index unsigned cast.

CastUOperationBuilder

A builder for a castu operation.

CeilDivSOperation

A ceildivs operation. Index signed ceil division.

CeilDivSOperationBuilder

A builder for a ceildivs operation.

CeilDivUOperation

A ceildivu operation. Index unsigned ceil division.

CeilDivUOperationBuilder

A builder for a ceildivu operation.

CmpOperation

A cmp operation. Index compare.

CmpOperationBuilder

A builder for a cmp operation.

ConstantOperation

A constant operation. Index constant.

ConstantOperationBuilder

A builder for a constant operation.

DivSOperation

A divs operation. Index signed division.

DivSOperationBuilder

A builder for a divs operation.

DivUOperation

A divu operation. Index unsigned division.

DivUOperationBuilder

A builder for a divu operation.

FloorDivSOperation

A floordivs operation. Index signed floor division.

FloorDivSOperationBuilder

A builder for a floordivs operation.

MaxSOperation

A maxs operation. Index signed maximum.

MaxSOperationBuilder

A builder for a maxs operation.

MaxUOperation

A maxu operation. Index unsigned maximum.

MaxUOperationBuilder

A builder for a maxu operation.

MinSOperation

A mins operation. Index signed minimum.

MinSOperationBuilder

A builder for a mins operation.

MinUOperation

A minu operation. Index unsigned minimum.

MinUOperationBuilder

A builder for a minu operation.

MulOperation

A mul operation. Index multiplication.

MulOperationBuilder

A builder for a mul operation.

OrOperation

An or operation. Index bitwise or.

OrOperationBuilder

A builder for an or operation.

RemSOperation

A rems operation. Index signed remainder.

RemSOperationBuilder

A builder for a rems operation.

RemUOperation

A remu operation. Index unsigned remainder.

RemUOperationBuilder

A builder for a remu operation.

ShlOperation

A shl operation. Index shift left.

ShlOperationBuilder

A builder for a shl operation.

ShrSOperation

A shrs operation. Signed index shift right.

ShrSOperationBuilder

A builder for a shrs operation.

ShrUOperation

A shru operation. Unsigned index shift right.

ShrUOperationBuilder

A builder for a shru operation.

SizeOfOperation

A sizeof operation. Size in bits of the index type.

SizeOfOperationBuilder

A builder for a sizeof operation.

SubOperation

A sub operation. Index subtraction.

SubOperationBuilder

A builder for a sub operation.

XOrOperation

A xor operation. Index bitwise xor.

XOrOperationBuilder

A builder for a xor operation.

Functions

add

Creates an add operation.

and

Creates an and operation.

bool_constant

Creates a bool.constant operation.

casts

Creates a casts operation.

castu

Creates a castu operation.

ceildivs

Creates a ceildivs operation.

ceildivu

Creates a ceildivu operation.

cmp

Creates a cmp operation.

constant

Creates a constant operation.

divs

Creates a divs operation.

divu

Creates a divu operation.

floordivs

Creates a floordivs operation.

maxs

Creates a maxs operation.

maxu

Creates a maxu operation.

mins

Creates a mins operation.

minu

Creates a minu operation.

mul

Creates a mul operation.

or

Creates an or operation.

rems

Creates a rems operation.

remu

Creates a remu operation.

shl

Creates a shl operation.

shrs

Creates a shrs operation.

shru

Creates a shru operation.

sizeof

Creates a sizeof operation.

sub

Creates a sub operation.

xor

Creates a xor operation.