Module melior::dialect::ods::llvm

llvm dialect.

Structs

AShrOperation

An ashr operation.

AShrOperationBuilder

A builder for an ashr operation.

AbsOperation

An intr.abs operation.

AbsOperationBuilder

A builder for an intr.abs operation.

AddOperation

An add operation.

AddOperationBuilder

A builder for an add operation.

AddrSpaceCastOperation

An addrspacecast operation.

AddrSpaceCastOperationBuilder

A builder for an addrspacecast operation.

AddressOfOperation

A mlir.addressof operation. Creates a pointer pointing to a global or a function.

AddressOfOperationBuilder

A builder for a mlir.addressof operation.

AllocaOperation

An alloca operation.

AllocaOperationBuilder

A builder for an alloca operation.

AndOperation

An and operation.

AndOperationBuilder

A builder for an and operation.

AnnotationOperation

An intr.annotation operation.

AnnotationOperationBuilder

A builder for an intr.annotation operation.

AssumeOperation

An intr.assume operation.

AssumeOperationBuilder

A builder for an intr.assume operation.

AtomicCmpXchgOperation

A cmpxchg operation.

AtomicCmpXchgOperationBuilder

A builder for a cmpxchg operation.

AtomicRMWOperation

An atomicrmw operation.

AtomicRMWOperationBuilder

A builder for an atomicrmw operation.

BitReverseOperation

An intr.bitreverse operation.

BitReverseOperationBuilder

A builder for an intr.bitreverse operation.

BitcastOperation

A bitcast operation.

BitcastOperationBuilder

A builder for a bitcast operation.

BrOperation

A br operation.

BrOperationBuilder

A builder for a br operation.

ByteSwapOperation

An intr.bswap operation.

ByteSwapOperationBuilder

A builder for an intr.bswap operation.

CallIntrinsicOperation

A call_intrinsic operation. Call to an LLVM intrinsic function..

CallIntrinsicOperationBuilder

A builder for a call_intrinsic operation.

CallOperation

A call operation. Call to an LLVM function..

CallOperationBuilder

A builder for a call operation.

ComdatOperation

A comdat operation. LLVM dialect comdat region.

ComdatOperationBuilder

A builder for a comdat operation.

ComdatSelectorOperation

A comdat_selector operation. LLVM dialect comdat selector declaration.

ComdatSelectorOperationBuilder

A builder for a comdat_selector operation.

CondBrOperation

A cond_br operation.

CondBrOperationBuilder

A builder for a cond_br operation.

ConstantOperation

A mlir.constant operation. Defines a constant of LLVM type..

ConstantOperationBuilder

A builder for a mlir.constant operation.

CopySignOperation

An intr.copysign operation.

CopySignOperationBuilder

A builder for an intr.copysign operation.

CoroAlignOperation

An intr.coro.align operation.

CoroAlignOperationBuilder

A builder for an intr.coro.align operation.

CoroBeginOperation

An intr.coro.begin operation.

CoroBeginOperationBuilder

A builder for an intr.coro.begin operation.

CoroEndOperation

An intr.coro.end operation.

CoroEndOperationBuilder

A builder for an intr.coro.end operation.

CoroFreeOperation

An intr.coro.free operation.

CoroFreeOperationBuilder

A builder for an intr.coro.free operation.

CoroIdOperation

An intr.coro.id operation.

CoroIdOperationBuilder

A builder for an intr.coro.id operation.

CoroResumeOperation

An intr.coro.resume operation.

CoroResumeOperationBuilder

A builder for an intr.coro.resume operation.

CoroSaveOperation

An intr.coro.save operation.

CoroSaveOperationBuilder

A builder for an intr.coro.save operation.

CoroSizeOperation

An intr.coro.size operation.

CoroSizeOperationBuilder

A builder for an intr.coro.size operation.

CoroSuspendOperation

An intr.coro.suspend operation.

CoroSuspendOperationBuilder

A builder for an intr.coro.suspend operation.

CosOperation

An intr.cos operation.

CosOperationBuilder

A builder for an intr.cos operation.

CountLeadingZerosOperation

An intr.ctlz operation.

CountLeadingZerosOperationBuilder

A builder for an intr.ctlz operation.

CountTrailingZerosOperation

An intr.cttz operation.

CountTrailingZerosOperationBuilder

A builder for an intr.cttz operation.

CtPopOperation

An intr.ctpop operation.

CtPopOperationBuilder

A builder for an intr.ctpop operation.

DbgDeclareOperation

An intr.dbg.declare operation. Describes how the address relates to a source language variable..

DbgDeclareOperationBuilder

A builder for an intr.dbg.declare operation.

DbgLabelOperation

An intr.dbg.label operation. Relates the program to a debug information label..

DbgLabelOperationBuilder

A builder for an intr.dbg.label operation.

DbgValueOperation

An intr.dbg.value operation. Describes how the value relates to a source language variable..

DbgValueOperationBuilder

A builder for an intr.dbg.value operation.

DebugTrapOperation

An intr.debugtrap operation.

DebugTrapOperationBuilder

A builder for an intr.debugtrap operation.

EhTypeidForOperation

An intr.eh.typeid.for operation.

EhTypeidForOperationBuilder

A builder for an intr.eh.typeid.for operation.

Exp2Operation

An intr.exp2 operation.

Exp2OperationBuilder

A builder for an intr.exp2 operation.

ExpOperation

An intr.exp operation.

ExpOperationBuilder

A builder for an intr.exp operation.

ExpectOperation

An intr.expect operation.

ExpectOperationBuilder

A builder for an intr.expect operation.

ExpectWithProbabilityOperation

An intr.expect.with.probability operation.

ExpectWithProbabilityOperationBuilder

A builder for an intr.expect.with.probability operation.

ExtractElementOperation

An extractelement operation. Extract an element from an LLVM vector..

ExtractElementOperationBuilder

A builder for an extractelement operation.

ExtractValueOperation

An extractvalue operation. Extract a value from an LLVM struct..

ExtractValueOperationBuilder

A builder for an extractvalue operation.

FAbsOperation

An intr.fabs operation.

FAbsOperationBuilder

A builder for an intr.fabs operation.

FAddOperation

A fadd operation.

FAddOperationBuilder

A builder for a fadd operation.

FCeilOperation

An intr.ceil operation.

FCeilOperationBuilder

A builder for an intr.ceil operation.

FCmpOperation

A fcmp operation.

FCmpOperationBuilder

A builder for a fcmp operation.

FDivOperation

A fdiv operation.

FDivOperationBuilder

A builder for a fdiv operation.

FFloorOperation

An intr.floor operation.

FFloorOperationBuilder

A builder for an intr.floor operation.

FMAOperation

An intr.fma operation.

FMAOperationBuilder

A builder for an intr.fma operation.

FMulAddOperation

An intr.fmuladd operation.

FMulAddOperationBuilder

A builder for an intr.fmuladd operation.

FMulOperation

A fmul operation.

FMulOperationBuilder

A builder for a fmul operation.

FNegOperation

A fneg operation.

FNegOperationBuilder

A builder for a fneg operation.

FPExtOperation

A fpext operation.

FPExtOperationBuilder

A builder for a fpext operation.

FPToSIOperation

A fptosi operation.

FPToSIOperationBuilder

A builder for a fptosi operation.

FPToUIOperation

A fptoui operation.

FPToUIOperationBuilder

A builder for a fptoui operation.

FPTruncOperation

A fptrunc operation.

FPTruncOperationBuilder

A builder for a fptrunc operation.

FRemOperation

A frem operation.

FRemOperationBuilder

A builder for a frem operation.

FSubOperation

A fsub operation.

FSubOperationBuilder

A builder for a fsub operation.

FTruncOperation

An intr.trunc operation.

FTruncOperationBuilder

A builder for an intr.trunc operation.

FenceOperation

A fence operation.

FenceOperationBuilder

A builder for a fence operation.

FreezeOperation

A freeze operation.

FreezeOperationBuilder

A builder for a freeze operation.

FshlOperation

An intr.fshl operation.

FshlOperationBuilder

A builder for an intr.fshl operation.

FshrOperation

An intr.fshr operation.

FshrOperationBuilder

A builder for an intr.fshr operation.

GEPOperation

A getelementptr operation.

GEPOperationBuilder

A builder for a getelementptr operation.

GetActiveLaneMaskOperation

An intr.get.active.lane.mask operation.

GetActiveLaneMaskOperationBuilder

A builder for an intr.get.active.lane.mask operation.

GlobalCtorsOperation

A mlir.global_ctors operation. LLVM dialect global_ctors..

GlobalCtorsOperationBuilder

A builder for a mlir.global_ctors operation.

GlobalDtorsOperation

A mlir.global_dtors operation. LLVM dialect global_dtors..

GlobalDtorsOperationBuilder

A builder for a mlir.global_dtors operation.

GlobalOperation

A mlir.global operation. LLVM dialect global..

GlobalOperationBuilder

A builder for a mlir.global operation.

ICmpOperation

An icmp operation.

ICmpOperationBuilder

A builder for an icmp operation.

InlineAsmOperation

An inline_asm operation.

InlineAsmOperationBuilder

A builder for an inline_asm operation.

InsertElementOperation

An insertelement operation. Insert an element into an LLVM vector..

InsertElementOperationBuilder

A builder for an insertelement operation.

InsertValueOperation

An insertvalue operation. Insert a value into an LLVM struct..

InsertValueOperationBuilder

A builder for an insertvalue operation.

IntToPtrOperation

An inttoptr operation.

IntToPtrOperationBuilder

A builder for an inttoptr operation.

InvokeOperation

An invoke operation.

InvokeOperationBuilder

A builder for an invoke operation.

IsConstantOperation

An intr.is.constant operation.

IsConstantOperationBuilder

A builder for an intr.is.constant operation.

IsFPClassOperation

An intr.is.fpclass operation.

IsFPClassOperationBuilder

A builder for an intr.is.fpclass operation.

LLVMFuncOperation

A func operation. LLVM dialect function..

LLVMFuncOperationBuilder

A builder for a func operation.

LShrOperation

A lshr operation.

LShrOperationBuilder

A builder for a lshr operation.

LandingpadOperation

A landingpad operation.

LandingpadOperationBuilder

A builder for a landingpad operation.

LifetimeEndOperation

An intr.lifetime.end operation.

LifetimeEndOperationBuilder

A builder for an intr.lifetime.end operation.

LifetimeStartOperation

An intr.lifetime.start operation.

LifetimeStartOperationBuilder

A builder for an intr.lifetime.start operation.

LlrintOperation

An intr.llrint operation.

LlrintOperationBuilder

A builder for an intr.llrint operation.

LlroundOperation

An intr.llround operation.

LlroundOperationBuilder

A builder for an intr.llround operation.

LoadOperation

A load operation.

LoadOperationBuilder

A builder for a load operation.

Log2Operation

An intr.log2 operation.

Log2OperationBuilder

A builder for an intr.log2 operation.

Log10Operation

An intr.log10 operation.

Log10OperationBuilder

A builder for an intr.log10 operation.

LogOperation

An intr.log operation.

LogOperationBuilder

A builder for an intr.log operation.

LrintOperation

An intr.lrint operation.

LrintOperationBuilder

A builder for an intr.lrint operation.

LroundOperation

An intr.lround operation.

LroundOperationBuilder

A builder for an intr.lround operation.

MaskedLoadOperation

An intr.masked.load operation.

MaskedLoadOperationBuilder

A builder for an intr.masked.load operation.

MaskedStoreOperation

An intr.masked.store operation.

MaskedStoreOperationBuilder

A builder for an intr.masked.store operation.

MatrixColumnMajorLoadOperation

An intr.matrix.column.major.load operation.

MatrixColumnMajorLoadOperationBuilder

A builder for an intr.matrix.column.major.load operation.

MatrixColumnMajorStoreOperation

An intr.matrix.column.major.store operation.

MatrixColumnMajorStoreOperationBuilder

A builder for an intr.matrix.column.major.store operation.

MatrixMultiplyOperation

An intr.matrix.multiply operation.

MatrixMultiplyOperationBuilder

A builder for an intr.matrix.multiply operation.

MatrixTransposeOperation

An intr.matrix.transpose operation.

MatrixTransposeOperationBuilder

A builder for an intr.matrix.transpose operation.

MaxNumOperation

An intr.maxnum operation.

MaxNumOperationBuilder

A builder for an intr.maxnum operation.

MaximumOperation

An intr.maximum operation.

MaximumOperationBuilder

A builder for an intr.maximum operation.

MemcpyInlineOperation

An intr.memcpy.inline operation.

MemcpyInlineOperationBuilder

A builder for an intr.memcpy.inline operation.

MemcpyOperation

An intr.memcpy operation.

MemcpyOperationBuilder

A builder for an intr.memcpy operation.

MemmoveOperation

An intr.memmove operation.

MemmoveOperationBuilder

A builder for an intr.memmove operation.

MemsetOperation

An intr.memset operation.

MemsetOperationBuilder

A builder for an intr.memset operation.

MetadataOperation

A metadata operation. LLVM dialect metadata..

MetadataOperationBuilder

A builder for a metadata operation.

MinNumOperation

An intr.minnum operation.

MinNumOperationBuilder

A builder for an intr.minnum operation.

MinimumOperation

An intr.minimum operation.

MinimumOperationBuilder

A builder for an intr.minimum operation.

MulOperation

A mul operation.

MulOperationBuilder

A builder for a mul operation.

NearbyintOperation

An intr.nearbyint operation.

NearbyintOperationBuilder

A builder for an intr.nearbyint operation.

NoAliasScopeDeclOperation

An intr.experimental.noalias.scope.decl operation.

NoAliasScopeDeclOperationBuilder

A builder for an intr.experimental.noalias.scope.decl operation.

NullOperation

A mlir.null operation. Defines a value containing a null pointer to LLVM type..

NullOperationBuilder

A builder for a mlir.null operation.

OrOperation

An or operation.

OrOperationBuilder

A builder for an or operation.

PoisonOperation

A mlir.poison operation. Creates a poison value of LLVM dialect type..

PoisonOperationBuilder

A builder for a mlir.poison operation.

PowIOperation

An intr.powi operation.

PowIOperationBuilder

A builder for an intr.powi operation.

PowOperation

An intr.pow operation.

PowOperationBuilder

A builder for an intr.pow operation.

PrefetchOperation

An intr.prefetch operation.

PrefetchOperationBuilder

A builder for an intr.prefetch operation.

PtrAnnotationOperation

An intr.ptr.annotation operation.

PtrAnnotationOperationBuilder

A builder for an intr.ptr.annotation operation.

PtrToIntOperation

A ptrtoint operation.

PtrToIntOperationBuilder

A builder for a ptrtoint operation.

ResumeOperation

A resume operation.

ResumeOperationBuilder

A builder for a resume operation.

ReturnOperation

A return operation.

ReturnOperationBuilder

A builder for a return operation.

RintOperation

An intr.rint operation.

RintOperationBuilder

A builder for an intr.rint operation.

RoundEvenOperation

An intr.roundeven operation.

RoundEvenOperationBuilder

A builder for an intr.roundeven operation.

RoundOperation

An intr.round operation.

RoundOperationBuilder

A builder for an intr.round operation.

SAddSatOperation

An intr.sadd.sat operation.

SAddSatOperationBuilder

A builder for an intr.sadd.sat operation.

SAddWithOverflowOperation

An intr.sadd.with.overflow operation.

SAddWithOverflowOperationBuilder

A builder for an intr.sadd.with.overflow operation.

SDivOperation

A sdiv operation.

SDivOperationBuilder

A builder for a sdiv operation.

SExtOperation

A sext operation.

SExtOperationBuilder

A builder for a sext operation.

SIToFPOperation

A sitofp operation.

SIToFPOperationBuilder

A builder for a sitofp operation.

SMaxOperation

An intr.smax operation.

SMaxOperationBuilder

A builder for an intr.smax operation.

SMinOperation

An intr.smin operation.

SMinOperationBuilder

A builder for an intr.smin operation.

SMulWithOverflowOperation

An intr.smul.with.overflow operation.

SMulWithOverflowOperationBuilder

A builder for an intr.smul.with.overflow operation.

SRemOperation

A srem operation.

SRemOperationBuilder

A builder for a srem operation.

SSACopyOperation

An intr.ssa.copy operation.

SSACopyOperationBuilder

A builder for an intr.ssa.copy operation.

SSHLSatOperation

An intr.sshl.sat operation.

SSHLSatOperationBuilder

A builder for an intr.sshl.sat operation.

SSubSatOperation

An intr.ssub.sat operation.

SSubSatOperationBuilder

A builder for an intr.ssub.sat operation.

SSubWithOverflowOperation

An intr.ssub.with.overflow operation.

SSubWithOverflowOperationBuilder

A builder for an intr.ssub.with.overflow operation.

SelectOperation

A select operation.

SelectOperationBuilder

A builder for a select operation.

ShlOperation

A shl operation.

ShlOperationBuilder

A builder for a shl operation.

ShuffleVectorOperation

A shufflevector operation. Construct a permutation of two vectors..

ShuffleVectorOperationBuilder

A builder for a shufflevector operation.

SinOperation

An intr.sin operation.

SinOperationBuilder

A builder for an intr.sin operation.

SqrtOperation

An intr.sqrt operation.

SqrtOperationBuilder

A builder for an intr.sqrt operation.

StackRestoreOperation

An intr.stackrestore operation.

StackRestoreOperationBuilder

A builder for an intr.stackrestore operation.

StackSaveOperation

An intr.stacksave operation.

StackSaveOperationBuilder

A builder for an intr.stacksave operation.

StepVectorOperation

An intr.experimental.stepvector operation.

StepVectorOperationBuilder

A builder for an intr.experimental.stepvector operation.

StoreOperation

A store operation.

StoreOperationBuilder

A builder for a store operation.

SubOperation

A sub operation.

SubOperationBuilder

A builder for a sub operation.

SwitchOperation

A switch operation.

SwitchOperationBuilder

A builder for a switch operation.

ThreadlocalAddressOperation

An intr.threadlocal.address operation.

ThreadlocalAddressOperationBuilder

A builder for an intr.threadlocal.address operation.

TrapOperation

An intr.trap operation.

TrapOperationBuilder

A builder for an intr.trap operation.

TruncOperation

A trunc operation.

TruncOperationBuilder

A builder for a trunc operation.

UAddSatOperation

An intr.uadd.sat operation.

UAddSatOperationBuilder

A builder for an intr.uadd.sat operation.

UAddWithOverflowOperation

An intr.uadd.with.overflow operation.

UAddWithOverflowOperationBuilder

A builder for an intr.uadd.with.overflow operation.

UBSanTrapOperation

An intr.ubsantrap operation.

UBSanTrapOperationBuilder

A builder for an intr.ubsantrap operation.

UDivOperation

An udiv operation.

UDivOperationBuilder

A builder for an udiv operation.

UIToFPOperation

An uitofp operation.

UIToFPOperationBuilder

A builder for an uitofp operation.

UMaxOperation

An intr.umax operation.

UMaxOperationBuilder

A builder for an intr.umax operation.

UMinOperation

An intr.umin operation.

UMinOperationBuilder

A builder for an intr.umin operation.

UMulWithOverflowOperation

An intr.umul.with.overflow operation.

UMulWithOverflowOperationBuilder

A builder for an intr.umul.with.overflow operation.

URemOperation

An urem operation.

URemOperationBuilder

A builder for an urem operation.

USHLSatOperation

An intr.ushl.sat operation.

USHLSatOperationBuilder

A builder for an intr.ushl.sat operation.

USubSatOperation

An intr.usub.sat operation.

USubSatOperationBuilder

A builder for an intr.usub.sat operation.

USubWithOverflowOperation

An intr.usub.with.overflow operation.

USubWithOverflowOperationBuilder

A builder for an intr.usub.with.overflow operation.

UndefOperation

A mlir.undef operation. Creates an undefined value of LLVM dialect type..

UndefOperationBuilder

A builder for a mlir.undef operation.

UnreachableOperation

An unreachable operation.

UnreachableOperationBuilder

A builder for an unreachable operation.

VPAShrOperation

An intr.vp.ashr operation.

VPAShrOperationBuilder

A builder for an intr.vp.ashr operation.

VPAddOperation

An intr.vp.add operation.

VPAddOperationBuilder

A builder for an intr.vp.add operation.

VPAndOperation

An intr.vp.and operation.

VPAndOperationBuilder

A builder for an intr.vp.and operation.

VPFAddOperation

An intr.vp.fadd operation.

VPFAddOperationBuilder

A builder for an intr.vp.fadd operation.

VPFDivOperation

An intr.vp.fdiv operation.

VPFDivOperationBuilder

A builder for an intr.vp.fdiv operation.

VPFMulAddOperation

An intr.vp.fmuladd operation.

VPFMulAddOperationBuilder

A builder for an intr.vp.fmuladd operation.

VPFMulOperation

An intr.vp.fmul operation.

VPFMulOperationBuilder

A builder for an intr.vp.fmul operation.

VPFNegOperation

An intr.vp.fneg operation.

VPFNegOperationBuilder

A builder for an intr.vp.fneg operation.

VPFPExtOperation

An intr.vp.fpext operation.

VPFPExtOperationBuilder

A builder for an intr.vp.fpext operation.

VPFPToSIOperation

An intr.vp.fptosi operation.

VPFPToSIOperationBuilder

A builder for an intr.vp.fptosi operation.

VPFPToUIOperation

An intr.vp.fptoui operation.

VPFPToUIOperationBuilder

A builder for an intr.vp.fptoui operation.

VPFPTruncOperation

An intr.vp.fptrunc operation.

VPFPTruncOperationBuilder

A builder for an intr.vp.fptrunc operation.

VPFRemOperation

An intr.vp.frem operation.

VPFRemOperationBuilder

A builder for an intr.vp.frem operation.

VPFSubOperation

An intr.vp.fsub operation.

VPFSubOperationBuilder

A builder for an intr.vp.fsub operation.

VPFmaOperation

An intr.vp.fma operation.

VPFmaOperationBuilder

A builder for an intr.vp.fma operation.

VPIntToPtrOperation

An intr.vp.inttoptr operation.

VPIntToPtrOperationBuilder

A builder for an intr.vp.inttoptr operation.

VPLShrOperation

An intr.vp.lshr operation.

VPLShrOperationBuilder

A builder for an intr.vp.lshr operation.

VPLoadOperation

An intr.vp.load operation.

VPLoadOperationBuilder

A builder for an intr.vp.load operation.

VPMergeMinOperation

An intr.vp.merge operation.

VPMergeMinOperationBuilder

A builder for an intr.vp.merge operation.

VPMulOperation

An intr.vp.mul operation.

VPMulOperationBuilder

A builder for an intr.vp.mul operation.

VPOrOperation

An intr.vp.or operation.

VPOrOperationBuilder

A builder for an intr.vp.or operation.

VPPtrToIntOperation

An intr.vp.ptrtoint operation.

VPPtrToIntOperationBuilder

A builder for an intr.vp.ptrtoint operation.

VPReduceAddOperation

An intr.vp.reduce.add operation.

VPReduceAddOperationBuilder

A builder for an intr.vp.reduce.add operation.

VPReduceAndOperation

An intr.vp.reduce.and operation.

VPReduceAndOperationBuilder

A builder for an intr.vp.reduce.and operation.

VPReduceFAddOperation

An intr.vp.reduce.fadd operation.

VPReduceFAddOperationBuilder

A builder for an intr.vp.reduce.fadd operation.

VPReduceFMaxOperation

An intr.vp.reduce.fmax operation.

VPReduceFMaxOperationBuilder

A builder for an intr.vp.reduce.fmax operation.

VPReduceFMinOperation

An intr.vp.reduce.fmin operation.

VPReduceFMinOperationBuilder

A builder for an intr.vp.reduce.fmin operation.

VPReduceFMulOperation

An intr.vp.reduce.fmul operation.

VPReduceFMulOperationBuilder

A builder for an intr.vp.reduce.fmul operation.

VPReduceMulOperation

An intr.vp.reduce.mul operation.

VPReduceMulOperationBuilder

A builder for an intr.vp.reduce.mul operation.

VPReduceOrOperation

An intr.vp.reduce.or operation.

VPReduceOrOperationBuilder

A builder for an intr.vp.reduce.or operation.

VPReduceSMaxOperation

An intr.vp.reduce.smax operation.

VPReduceSMaxOperationBuilder

A builder for an intr.vp.reduce.smax operation.

VPReduceSMinOperation

An intr.vp.reduce.smin operation.

VPReduceSMinOperationBuilder

A builder for an intr.vp.reduce.smin operation.

VPReduceUMaxOperation

An intr.vp.reduce.umax operation.

VPReduceUMaxOperationBuilder

A builder for an intr.vp.reduce.umax operation.

VPReduceUMinOperation

An intr.vp.reduce.umin operation.

VPReduceUMinOperationBuilder

A builder for an intr.vp.reduce.umin operation.

VPReduceXorOperation

An intr.vp.reduce.xor operation.

VPReduceXorOperationBuilder

A builder for an intr.vp.reduce.xor operation.

VPSDivOperation

An intr.vp.sdiv operation.

VPSDivOperationBuilder

A builder for an intr.vp.sdiv operation.

VPSExtOperation

An intr.vp.sext operation.

VPSExtOperationBuilder

A builder for an intr.vp.sext operation.

VPSIToFPOperation

An intr.vp.sitofp operation.

VPSIToFPOperationBuilder

A builder for an intr.vp.sitofp operation.

VPSRemOperation

An intr.vp.srem operation.

VPSRemOperationBuilder

A builder for an intr.vp.srem operation.

VPSelectMinOperation

An intr.vp.select operation.

VPSelectMinOperationBuilder

A builder for an intr.vp.select operation.

VPShlOperation

An intr.vp.shl operation.

VPShlOperationBuilder

A builder for an intr.vp.shl operation.

VPStoreOperation

An intr.vp.store operation.

VPStoreOperationBuilder

A builder for an intr.vp.store operation.

VPStridedLoadOperation

An intr.experimental.vp.strided.load operation.

VPStridedLoadOperationBuilder

A builder for an intr.experimental.vp.strided.load operation.

VPStridedStoreOperation

An intr.experimental.vp.strided.store operation.

VPStridedStoreOperationBuilder

A builder for an intr.experimental.vp.strided.store operation.

VPSubOperation

An intr.vp.sub operation.

VPSubOperationBuilder

A builder for an intr.vp.sub operation.

VPTruncOperation

An intr.vp.trunc operation.

VPTruncOperationBuilder

A builder for an intr.vp.trunc operation.

VPUDivOperation

An intr.vp.udiv operation.

VPUDivOperationBuilder

A builder for an intr.vp.udiv operation.

VPUIToFPOperation

An intr.vp.uitofp operation.

VPUIToFPOperationBuilder

A builder for an intr.vp.uitofp operation.

VPURemOperation

An intr.vp.urem operation.

VPURemOperationBuilder

A builder for an intr.vp.urem operation.

VPXorOperation

An intr.vp.xor operation.

VPXorOperationBuilder

A builder for an intr.vp.xor operation.

VPZExtOperation

An intr.vp.zext operation.

VPZExtOperationBuilder

A builder for an intr.vp.zext operation.

VaCopyOperation

An intr.vacopy operation. Copies the current argument position from src_list to dest_list..

VaCopyOperationBuilder

A builder for an intr.vacopy operation.

VaEndOperation

An intr.vaend operation. Destroys arg_list, which has been initialized by intr.vastart or intr.vacopy..

VaEndOperationBuilder

A builder for an intr.vaend operation.

VaStartOperation

An intr.vastart operation. Initializes arg_list for subsequent variadic argument extractions..

VaStartOperationBuilder

A builder for an intr.vastart operation.

VarAnnotationOperation

An intr.var.annotation operation.

VarAnnotationOperationBuilder

A builder for an intr.var.annotation operation.

XOrOperation

A xor operation.

XOrOperationBuilder

A builder for a xor operation.

ZExtOperation

A zext operation.

ZExtOperationBuilder

A builder for a zext operation.

masked_compressstoreOperation

An intr.masked.compressstore operation.

masked_compressstoreOperationBuilder

A builder for an intr.masked.compressstore operation.

masked_expandloadOperation

An intr.masked.expandload operation.

masked_expandloadOperationBuilder

A builder for an intr.masked.expandload operation.

masked_gatherOperation

An intr.masked.gather operation.

masked_gatherOperationBuilder

A builder for an intr.masked.gather operation.

masked_scatterOperation

An intr.masked.scatter operation.

masked_scatterOperationBuilder

A builder for an intr.masked.scatter operation.

vector_extractOperation

An intr.vector.extract operation.

vector_extractOperationBuilder

A builder for an intr.vector.extract operation.

vector_insertOperation

An intr.vector.insert operation.

vector_insertOperationBuilder

A builder for an intr.vector.insert operation.

vector_reduce_addOperation

An intr.vector.reduce.add operation.

vector_reduce_addOperationBuilder

A builder for an intr.vector.reduce.add operation.

vector_reduce_andOperation

An intr.vector.reduce.and operation.

vector_reduce_andOperationBuilder

A builder for an intr.vector.reduce.and operation.

vector_reduce_faddOperation

An intr.vector.reduce.fadd operation.

vector_reduce_faddOperationBuilder

A builder for an intr.vector.reduce.fadd operation.

vector_reduce_fmaxOperation

An intr.vector.reduce.fmax operation.

vector_reduce_fmaxOperationBuilder

A builder for an intr.vector.reduce.fmax operation.

vector_reduce_fmaximumOperation

An intr.vector.reduce.fmaximum operation.

vector_reduce_fmaximumOperationBuilder

A builder for an intr.vector.reduce.fmaximum operation.

vector_reduce_fminOperation

An intr.vector.reduce.fmin operation.

vector_reduce_fminOperationBuilder

A builder for an intr.vector.reduce.fmin operation.

vector_reduce_fminimumOperation

An intr.vector.reduce.fminimum operation.

vector_reduce_fminimumOperationBuilder

A builder for an intr.vector.reduce.fminimum operation.

vector_reduce_fmulOperation

An intr.vector.reduce.fmul operation.

vector_reduce_fmulOperationBuilder

A builder for an intr.vector.reduce.fmul operation.

vector_reduce_mulOperation

An intr.vector.reduce.mul operation.

vector_reduce_mulOperationBuilder

A builder for an intr.vector.reduce.mul operation.

vector_reduce_orOperation

An intr.vector.reduce.or operation.

vector_reduce_orOperationBuilder

A builder for an intr.vector.reduce.or operation.

vector_reduce_smaxOperation

An intr.vector.reduce.smax operation.

vector_reduce_smaxOperationBuilder

A builder for an intr.vector.reduce.smax operation.

vector_reduce_sminOperation

An intr.vector.reduce.smin operation.

vector_reduce_sminOperationBuilder

A builder for an intr.vector.reduce.smin operation.

vector_reduce_umaxOperation

An intr.vector.reduce.umax operation.

vector_reduce_umaxOperationBuilder

A builder for an intr.vector.reduce.umax operation.

vector_reduce_uminOperation

An intr.vector.reduce.umin operation.

vector_reduce_uminOperationBuilder

A builder for an intr.vector.reduce.umin operation.

vector_reduce_xorOperation

An intr.vector.reduce.xor operation.

vector_reduce_xorOperationBuilder

A builder for an intr.vector.reduce.xor operation.

vscaleOperation

An intr.vscale operation.

vscaleOperationBuilder

A builder for an intr.vscale operation.

Functions

add

Creates an add operation.

addrspacecast

Creates an addrspacecast operation.

alloca

Creates an alloca operation.

and

Creates an and operation.

ashr

Creates an ashr operation.

atomicrmw

Creates an atomicrmw operation.

bitcast

Creates a bitcast operation.

br

Creates a br operation.

call

Creates a call operation.

call_intrinsic

Creates a call_intrinsic operation.

cmpxchg

Creates a cmpxchg operation.

comdat

Creates a comdat operation.

comdat_selector

Creates a comdat_selector operation.

cond_br

Creates a cond_br operation.

extractelement

Creates an extractelement operation.

extractvalue

Creates an extractvalue operation.

fadd

Creates a fadd operation.

fcmp

Creates a fcmp operation.

fdiv

Creates a fdiv operation.

fence

Creates a fence operation.

fmul

Creates a fmul operation.

fneg

Creates a fneg operation.

fpext

Creates a fpext operation.

fptosi

Creates a fptosi operation.

fptoui

Creates a fptoui operation.

fptrunc

Creates a fptrunc operation.

freeze

Creates a freeze operation.

frem

Creates a frem operation.

fsub

Creates a fsub operation.

func

Creates a func operation.

getelementptr

Creates a getelementptr operation.

icmp

Creates an icmp operation.

inline_asm

Creates an inline_asm operation.

insertelement

Creates an insertelement operation.

insertvalue

Creates an insertvalue operation.

intr_abs

Creates an intr.abs operation.

intr_annotation

Creates an intr.annotation operation.

intr_assume

Creates an intr.assume operation.

intr_bitreverse

Creates an intr.bitreverse operation.

intr_bswap

Creates an intr.bswap operation.

intr_ceil

Creates an intr.ceil operation.

intr_copysign

Creates an intr.copysign operation.

intr_coro_align

Creates an intr.coro.align operation.

intr_coro_begin

Creates an intr.coro.begin operation.

intr_coro_end

Creates an intr.coro.end operation.

intr_coro_free

Creates an intr.coro.free operation.

intr_coro_id

Creates an intr.coro.id operation.

intr_coro_resume

Creates an intr.coro.resume operation.

intr_coro_save

Creates an intr.coro.save operation.

intr_coro_size

Creates an intr.coro.size operation.

intr_coro_suspend

Creates an intr.coro.suspend operation.

intr_cos

Creates an intr.cos operation.

intr_ctlz

Creates an intr.ctlz operation.

intr_ctpop

Creates an intr.ctpop operation.

intr_cttz

Creates an intr.cttz operation.

intr_dbg_declare

Creates an intr.dbg.declare operation.

intr_dbg_label

Creates an intr.dbg.label operation.

intr_dbg_value

Creates an intr.dbg.value operation.

intr_debugtrap

Creates an intr.debugtrap operation.

intr_eh_typeid_for

Creates an intr.eh.typeid.for operation.

intr_exp

Creates an intr.exp operation.

intr_exp_2

Creates an intr.exp2 operation.

intr_expect

Creates an intr.expect operation.

intr_expect_with_probability

Creates an intr.expect.with.probability operation.

intr_experimental_noalias_scope_decl

Creates an intr.experimental.noalias.scope.decl operation.

intr_experimental_stepvector

Creates an intr.experimental.stepvector operation.

intr_experimental_vp_strided_load

Creates an intr.experimental.vp.strided.load operation.

intr_experimental_vp_strided_store

Creates an intr.experimental.vp.strided.store operation.

intr_fabs

Creates an intr.fabs operation.

intr_floor

Creates an intr.floor operation.

intr_fma

Creates an intr.fma operation.

intr_fmuladd

Creates an intr.fmuladd operation.

intr_fshl

Creates an intr.fshl operation.

intr_fshr

Creates an intr.fshr operation.

intr_get_active_lane_mask

Creates an intr.get.active.lane.mask operation.

intr_is_constant

Creates an intr.is.constant operation.

intr_is_fpclass

Creates an intr.is.fpclass operation.

intr_lifetime_end

Creates an intr.lifetime.end operation.

intr_lifetime_start

Creates an intr.lifetime.start operation.

intr_llrint

Creates an intr.llrint operation.

intr_llround

Creates an intr.llround operation.

intr_log

Creates an intr.log operation.

intr_log_2

Creates an intr.log2 operation.

intr_log_10

Creates an intr.log10 operation.

intr_lrint

Creates an intr.lrint operation.

intr_lround

Creates an intr.lround operation.

intr_masked_compressstore

Creates an intr.masked.compressstore operation.

intr_masked_expandload

Creates an intr.masked.expandload operation.

intr_masked_gather

Creates an intr.masked.gather operation.

intr_masked_load

Creates an intr.masked.load operation.

intr_masked_scatter

Creates an intr.masked.scatter operation.

intr_masked_store

Creates an intr.masked.store operation.

intr_matrix_column_major_load

Creates an intr.matrix.column.major.load operation.

intr_matrix_column_major_store

Creates an intr.matrix.column.major.store operation.

intr_matrix_multiply

Creates an intr.matrix.multiply operation.

intr_matrix_transpose

Creates an intr.matrix.transpose operation.

intr_maximum

Creates an intr.maximum operation.

intr_maxnum

Creates an intr.maxnum operation.

intr_memcpy

Creates an intr.memcpy operation.

intr_memcpy_inline

Creates an intr.memcpy.inline operation.

intr_memmove

Creates an intr.memmove operation.

intr_memset

Creates an intr.memset operation.

intr_minimum

Creates an intr.minimum operation.

intr_minnum

Creates an intr.minnum operation.

intr_nearbyint

Creates an intr.nearbyint operation.

intr_pow

Creates an intr.pow operation.

intr_powi

Creates an intr.powi operation.

intr_prefetch

Creates an intr.prefetch operation.

intr_ptr_annotation

Creates an intr.ptr.annotation operation.

intr_rint

Creates an intr.rint operation.

intr_round

Creates an intr.round operation.

intr_roundeven

Creates an intr.roundeven operation.

intr_sadd_sat

Creates an intr.sadd.sat operation.

intr_sadd_with_overflow

Creates an intr.sadd.with.overflow operation.

intr_sin

Creates an intr.sin operation.

intr_smax

Creates an intr.smax operation.

intr_smin

Creates an intr.smin operation.

intr_smul_with_overflow

Creates an intr.smul.with.overflow operation.

intr_sqrt

Creates an intr.sqrt operation.

intr_ssa_copy

Creates an intr.ssa.copy operation.

intr_sshl_sat

Creates an intr.sshl.sat operation.

intr_ssub_sat

Creates an intr.ssub.sat operation.

intr_ssub_with_overflow

Creates an intr.ssub.with.overflow operation.

intr_stackrestore

Creates an intr.stackrestore operation.

intr_stacksave

Creates an intr.stacksave operation.

intr_threadlocal_address

Creates an intr.threadlocal.address operation.

intr_trap

Creates an intr.trap operation.

intr_trunc

Creates an intr.trunc operation.

intr_uadd_sat

Creates an intr.uadd.sat operation.

intr_uadd_with_overflow

Creates an intr.uadd.with.overflow operation.

intr_ubsantrap

Creates an intr.ubsantrap operation.

intr_umax

Creates an intr.umax operation.

intr_umin

Creates an intr.umin operation.

intr_umul_with_overflow

Creates an intr.umul.with.overflow operation.

intr_ushl_sat

Creates an intr.ushl.sat operation.

intr_usub_sat

Creates an intr.usub.sat operation.

intr_usub_with_overflow

Creates an intr.usub.with.overflow operation.

intr_vacopy

Creates an intr.vacopy operation.

intr_vaend

Creates an intr.vaend operation.

intr_var_annotation

Creates an intr.var.annotation operation.

intr_vastart

Creates an intr.vastart operation.

intr_vector_extract

Creates an intr.vector.extract operation.

intr_vector_insert

Creates an intr.vector.insert operation.

intr_vector_reduce_add

Creates an intr.vector.reduce.add operation.

intr_vector_reduce_and

Creates an intr.vector.reduce.and operation.

intr_vector_reduce_fadd

Creates an intr.vector.reduce.fadd operation.

intr_vector_reduce_fmax

Creates an intr.vector.reduce.fmax operation.

intr_vector_reduce_fmaximum

Creates an intr.vector.reduce.fmaximum operation.

intr_vector_reduce_fmin

Creates an intr.vector.reduce.fmin operation.

intr_vector_reduce_fminimum

Creates an intr.vector.reduce.fminimum operation.

intr_vector_reduce_fmul

Creates an intr.vector.reduce.fmul operation.

intr_vector_reduce_mul

Creates an intr.vector.reduce.mul operation.

intr_vector_reduce_or

Creates an intr.vector.reduce.or operation.

intr_vector_reduce_smax

Creates an intr.vector.reduce.smax operation.

intr_vector_reduce_smin

Creates an intr.vector.reduce.smin operation.

intr_vector_reduce_umax

Creates an intr.vector.reduce.umax operation.

intr_vector_reduce_umin

Creates an intr.vector.reduce.umin operation.

intr_vector_reduce_xor

Creates an intr.vector.reduce.xor operation.

intr_vp_add

Creates an intr.vp.add operation.

intr_vp_and

Creates an intr.vp.and operation.

intr_vp_ashr

Creates an intr.vp.ashr operation.

intr_vp_fadd

Creates an intr.vp.fadd operation.

intr_vp_fdiv

Creates an intr.vp.fdiv operation.

intr_vp_fma

Creates an intr.vp.fma operation.

intr_vp_fmul

Creates an intr.vp.fmul operation.

intr_vp_fmuladd

Creates an intr.vp.fmuladd operation.

intr_vp_fneg

Creates an intr.vp.fneg operation.

intr_vp_fpext

Creates an intr.vp.fpext operation.

intr_vp_fptosi

Creates an intr.vp.fptosi operation.

intr_vp_fptoui

Creates an intr.vp.fptoui operation.

intr_vp_fptrunc

Creates an intr.vp.fptrunc operation.

intr_vp_frem

Creates an intr.vp.frem operation.

intr_vp_fsub

Creates an intr.vp.fsub operation.

intr_vp_inttoptr

Creates an intr.vp.inttoptr operation.

intr_vp_load

Creates an intr.vp.load operation.

intr_vp_lshr

Creates an intr.vp.lshr operation.

intr_vp_merge

Creates an intr.vp.merge operation.

intr_vp_mul

Creates an intr.vp.mul operation.

intr_vp_or

Creates an intr.vp.or operation.

intr_vp_ptrtoint

Creates an intr.vp.ptrtoint operation.

intr_vp_reduce_add

Creates an intr.vp.reduce.add operation.

intr_vp_reduce_and

Creates an intr.vp.reduce.and operation.

intr_vp_reduce_fadd

Creates an intr.vp.reduce.fadd operation.

intr_vp_reduce_fmax

Creates an intr.vp.reduce.fmax operation.

intr_vp_reduce_fmin

Creates an intr.vp.reduce.fmin operation.

intr_vp_reduce_fmul

Creates an intr.vp.reduce.fmul operation.

intr_vp_reduce_mul

Creates an intr.vp.reduce.mul operation.

intr_vp_reduce_or

Creates an intr.vp.reduce.or operation.

intr_vp_reduce_smax

Creates an intr.vp.reduce.smax operation.

intr_vp_reduce_smin

Creates an intr.vp.reduce.smin operation.

intr_vp_reduce_umax

Creates an intr.vp.reduce.umax operation.

intr_vp_reduce_umin

Creates an intr.vp.reduce.umin operation.

intr_vp_reduce_xor

Creates an intr.vp.reduce.xor operation.

intr_vp_sdiv

Creates an intr.vp.sdiv operation.

intr_vp_select

Creates an intr.vp.select operation.

intr_vp_sext

Creates an intr.vp.sext operation.

intr_vp_shl

Creates an intr.vp.shl operation.

intr_vp_sitofp

Creates an intr.vp.sitofp operation.

intr_vp_srem

Creates an intr.vp.srem operation.

intr_vp_store

Creates an intr.vp.store operation.

intr_vp_sub

Creates an intr.vp.sub operation.

intr_vp_trunc

Creates an intr.vp.trunc operation.

intr_vp_udiv

Creates an intr.vp.udiv operation.

intr_vp_uitofp

Creates an intr.vp.uitofp operation.

intr_vp_urem

Creates an intr.vp.urem operation.

intr_vp_xor

Creates an intr.vp.xor operation.

intr_vp_zext

Creates an intr.vp.zext operation.

intr_vscale

Creates an intr.vscale operation.

inttoptr

Creates an inttoptr operation.

invoke

Creates an invoke operation.

landingpad

Creates a landingpad operation.

load

Creates a load operation.

lshr

Creates a lshr operation.

metadata

Creates a metadata operation.

mlir_addressof

Creates a mlir.addressof operation.

mlir_constant

Creates a mlir.constant operation.

mlir_global

Creates a mlir.global operation.

mlir_global_ctors

Creates a mlir.global_ctors operation.

mlir_global_dtors

Creates a mlir.global_dtors operation.

mlir_null

Creates a mlir.null operation.

mlir_poison

Creates a mlir.poison operation.

mlir_undef

Creates a mlir.undef operation.

mul

Creates a mul operation.

or

Creates an or operation.

ptrtoint

Creates a ptrtoint operation.

resume

Creates a resume operation.

return

Creates a return operation.

sdiv

Creates a sdiv operation.

select

Creates a select operation.

sext

Creates a sext operation.

shl

Creates a shl operation.

shufflevector

Creates a shufflevector operation.

sitofp

Creates a sitofp operation.

srem

Creates a srem operation.

store

Creates a store operation.

sub

Creates a sub operation.

switch

Creates a switch operation.

trunc

Creates a trunc operation.

udiv

Creates an udiv operation.

uitofp

Creates an uitofp operation.

unreachable

Creates an unreachable operation.

urem

Creates an urem operation.

xor

Creates a xor operation.

zext

Creates a zext operation.