Module melior::dialect::ods::llvm

llvm dialect.

Structs

AShrOp

ashr operation

AShrOpBuilder

Builder for ashr operation

AddOp

add operation

AddOpBuilder

Builder for add operation

AddrSpaceCastOp

addrspacecast operation

AddrSpaceCastOpBuilder

Builder for addrspacecast operation

AddressOfOp

mlir.addressof operation: Creates a pointer pointing to a global or a function

AddressOfOpBuilder

Builder for mlir.addressof operation: Creates a pointer pointing to a global or a function

AllocaOp

alloca operation

AllocaOpBuilder

Builder for alloca operation

AndOp

and operation

AndOpBuilder

Builder for and operation

AtomicCmpXchgOp

cmpxchg operation

AtomicCmpXchgOpBuilder

Builder for cmpxchg operation

AtomicRMWOp

atomicrmw operation

AtomicRMWOpBuilder

Builder for atomicrmw operation

BitcastOp

bitcast operation

BitcastOpBuilder

Builder for bitcast operation

BrOp

br operation

BrOpBuilder

Builder for br operation

CallOp

call operation: Call to an LLVM function.

CallOpBuilder

Builder for call operation: Call to an LLVM function.

ComdatOp

comdat operation: LLVM dialect comdat region

ComdatOpBuilder

Builder for comdat operation: LLVM dialect comdat region

ComdatSelectorOp

comdat_selector operation: LLVM dialect comdat selector declaration

ComdatSelectorOpBuilder

Builder for comdat_selector operation: LLVM dialect comdat selector declaration

CondBrOp

cond_br operation

CondBrOpBuilder

Builder for cond_br operation

ConstantOp

mlir.constant operation: Defines a constant of LLVM type.

ConstantOpBuilder

Builder for mlir.constant operation: Defines a constant of LLVM type.

ExtractElementOp

extractelement operation: Extract an element from an LLVM vector.

ExtractElementOpBuilder

Builder for extractelement operation: Extract an element from an LLVM vector.

ExtractValueOp

extractvalue operation: Extract a value from an LLVM struct.

ExtractValueOpBuilder

Builder for extractvalue operation: Extract a value from an LLVM struct.

FAddOp

fadd operation

FAddOpBuilder

Builder for fadd operation

FCmpOp

fcmp operation

FCmpOpBuilder

Builder for fcmp operation

FDivOp

fdiv operation

FDivOpBuilder

Builder for fdiv operation

FMulOp

fmul operation

FMulOpBuilder

Builder for fmul operation

FNegOp

fneg operation

FNegOpBuilder

Builder for fneg operation

FPExtOp

fpext operation

FPExtOpBuilder

Builder for fpext operation

FPToSIOp

fptosi operation

FPToSIOpBuilder

Builder for fptosi operation

FPToUIOp

fptoui operation

FPToUIOpBuilder

Builder for fptoui operation

FPTruncOp

fptrunc operation

FPTruncOpBuilder

Builder for fptrunc operation

FRemOp

frem operation

FRemOpBuilder

Builder for frem operation

FSubOp

fsub operation

FSubOpBuilder

Builder for fsub operation

FenceOp

fence operation

FenceOpBuilder

Builder for fence operation

FreezeOp

freeze operation

FreezeOpBuilder

Builder for freeze operation

GEPOp

getelementptr operation

GEPOpBuilder

Builder for getelementptr operation

GlobalCtorsOp

mlir.global_ctors operation: LLVM dialect global_ctors.

GlobalCtorsOpBuilder

Builder for mlir.global_ctors operation: LLVM dialect global_ctors.

GlobalDtorsOp

mlir.global_dtors operation: LLVM dialect global_dtors.

GlobalDtorsOpBuilder

Builder for mlir.global_dtors operation: LLVM dialect global_dtors.

GlobalOp

mlir.global operation: LLVM dialect global.

GlobalOpBuilder

Builder for mlir.global operation: LLVM dialect global.

ICmpOp

icmp operation

ICmpOpBuilder

Builder for icmp operation

InlineAsmOp

inline_asm operation

InlineAsmOpBuilder

Builder for inline_asm operation

InsertElementOp

insertelement operation: Insert an element into an LLVM vector.

InsertElementOpBuilder

Builder for insertelement operation: Insert an element into an LLVM vector.

InsertValueOp

insertvalue operation: Insert a value into an LLVM struct.

InsertValueOpBuilder

Builder for insertvalue operation: Insert a value into an LLVM struct.

IntToPtrOp

inttoptr operation

IntToPtrOpBuilder

Builder for inttoptr operation

InvokeOp

invoke operation

InvokeOpBuilder

Builder for invoke operation

LLVMFuncOp

func operation: LLVM dialect function.

LLVMFuncOpBuilder

Builder for func operation: LLVM dialect function.

LShrOp

lshr operation

LShrOpBuilder

Builder for lshr operation

LandingpadOp

landingpad operation

LandingpadOpBuilder

Builder for landingpad operation

LoadOp

load operation

LoadOpBuilder

Builder for load operation

MetadataOp

metadata operation: LLVM dialect metadata.

MetadataOpBuilder

Builder for metadata operation: LLVM dialect metadata.

MulOp

mul operation

MulOpBuilder

Builder for mul operation

NullOp

mlir.null operation: Defines a value containing a null pointer to LLVM type.

NullOpBuilder

Builder for mlir.null operation: Defines a value containing a null pointer to LLVM type.

OrOp

or operation

OrOpBuilder

Builder for or operation

PoisonOp

mlir.poison operation: Creates a poison value of LLVM dialect type.

PoisonOpBuilder

Builder for mlir.poison operation: Creates a poison value of LLVM dialect type.

PtrToIntOp

ptrtoint operation

PtrToIntOpBuilder

Builder for ptrtoint operation

ResumeOp

resume operation

ResumeOpBuilder

Builder for resume operation

ReturnOp

return operation

ReturnOpBuilder

Builder for return operation

SDivOp

sdiv operation

SDivOpBuilder

Builder for sdiv operation

SExtOp

sext operation

SExtOpBuilder

Builder for sext operation

SIToFPOp

sitofp operation

SIToFPOpBuilder

Builder for sitofp operation

SRemOp

srem operation

SRemOpBuilder

Builder for srem operation

SelectOp

select operation

SelectOpBuilder

Builder for select operation

ShlOp

shl operation

ShlOpBuilder

Builder for shl operation

ShuffleVectorOp

shufflevector operation: Construct a permutation of two vectors.

ShuffleVectorOpBuilder

Builder for shufflevector operation: Construct a permutation of two vectors.

StoreOp

store operation

StoreOpBuilder

Builder for store operation

SubOp

sub operation

SubOpBuilder

Builder for sub operation

SwitchOp

switch operation

SwitchOpBuilder

Builder for switch operation

TruncOp

trunc operation

TruncOpBuilder

Builder for trunc operation

UDivOp

udiv operation

UDivOpBuilder

Builder for udiv operation

UIToFPOp

uitofp operation

UIToFPOpBuilder

Builder for uitofp operation

URemOp

urem operation

URemOpBuilder

Builder for urem operation

UndefOp

mlir.undef operation: Creates an undefined value of LLVM dialect type.

UndefOpBuilder

Builder for mlir.undef operation: Creates an undefined value of LLVM dialect type.

UnreachableOp

unreachable operation

UnreachableOpBuilder

Builder for unreachable operation

XOrOp

xor operation

XOrOpBuilder

Builder for xor operation

ZExtOp

zext operation

ZExtOpBuilder

Builder for zext operation

Functions

add

Creates a new add operation

addrspacecast

Creates a new addrspacecast operation

alloca

Creates a new alloca operation

and

Creates a new and operation

ashr

Creates a new ashr operation

atomicrmw

Creates a new atomicrmw operation

bitcast

Creates a new bitcast operation

br

Creates a new br operation

call

Creates a new call operation: Call to an LLVM function.

cmpxchg

Creates a new cmpxchg operation

comdat

Creates a new comdat operation: LLVM dialect comdat region

comdat_selector

Creates a new comdat_selector operation: LLVM dialect comdat selector declaration

cond_br

Creates a new cond_br operation

extractelement

Creates a new extractelement operation: Extract an element from an LLVM vector.

extractvalue

Creates a new extractvalue operation: Extract a value from an LLVM struct.

fadd

Creates a new fadd operation

fcmp

Creates a new fcmp operation

fdiv

Creates a new fdiv operation

fence

Creates a new fence operation

fmul

Creates a new fmul operation

fneg

Creates a new fneg operation

fpext

Creates a new fpext operation

fptosi

Creates a new fptosi operation

fptoui

Creates a new fptoui operation

fptrunc

Creates a new fptrunc operation

freeze

Creates a new freeze operation

frem

Creates a new frem operation

fsub

Creates a new fsub operation

func

Creates a new func operation: LLVM dialect function.

getelementptr

Creates a new getelementptr operation

icmp

Creates a new icmp operation

inline_asm

Creates a new inline_asm operation

insertelement

Creates a new insertelement operation: Insert an element into an LLVM vector.

insertvalue

Creates a new insertvalue operation: Insert a value into an LLVM struct.

inttoptr

Creates a new inttoptr operation

invoke

Creates a new invoke operation

landingpad

Creates a new landingpad operation

load

Creates a new load operation

lshr

Creates a new lshr operation

metadata

Creates a new metadata operation: LLVM dialect metadata.

mlir_addressof

Creates a new mlir.addressof operation: Creates a pointer pointing to a global or a function

mlir_constant

Creates a new mlir.constant operation: Defines a constant of LLVM type.

mlir_global

Creates a new mlir.global operation: LLVM dialect global.

mlir_global_ctors

Creates a new mlir.global_ctors operation: LLVM dialect global_ctors.

mlir_global_dtors

Creates a new mlir.global_dtors operation: LLVM dialect global_dtors.

mlir_null

Creates a new mlir.null operation: Defines a value containing a null pointer to LLVM type.

mlir_poison

Creates a new mlir.poison operation: Creates a poison value of LLVM dialect type.

mlir_undef

Creates a new mlir.undef operation: Creates an undefined value of LLVM dialect type.

mul

Creates a new mul operation

or

Creates a new or operation

ptrtoint

Creates a new ptrtoint operation

resume

Creates a new resume operation

return

Creates a new return operation

sdiv

Creates a new sdiv operation

select

Creates a new select operation

sext

Creates a new sext operation

shl

Creates a new shl operation

shufflevector

Creates a new shufflevector operation: Construct a permutation of two vectors.

sitofp

Creates a new sitofp operation

srem

Creates a new srem operation

store

Creates a new store operation

sub

Creates a new sub operation

switch

Creates a new switch operation

trunc

Creates a new trunc operation

udiv

Creates a new udiv operation

uitofp

Creates a new uitofp operation

unreachable

Creates a new unreachable operation

urem

Creates a new urem operation

xor

Creates a new xor operation

zext

Creates a new zext operation