Module melior::dialect::ods::math

math dialect.

The math dialect is intended to hold mathematical operations on integer and floating types beyond simple arithmetics. Each operation works on scalar, vector or tensor type. On vector and tensor type operations apply elementwise unless explicitly specified otherwise. As an example, the floating point absolute value can be expressed as:

// Scalar absolute value.
%a = math.absf %b : f64

// Vector elementwise absolute value.
%f = math.absf %g : vector<4xf32>

// Tensor elementwise absolute value.
%x = math.absf %y : tensor<4x?xf8>

Structs

AbsFOp

absf operation. Floating point absolute-value operation.

AbsFOpBuilder

Builder for absf operation. Floating point absolute-value operation.

AbsIOp

absi operation. Integer absolute-value operation.

AbsIOpBuilder

Builder for absi operation. Integer absolute-value operation.

Atan2Op

atan2 operation. 2-argument arcus tangent of the given values.

Atan2OpBuilder

Builder for atan2 operation. 2-argument arcus tangent of the given values.

AtanOp

atan operation. Arcus tangent of the given value.

AtanOpBuilder

Builder for atan operation. Arcus tangent of the given value.

CbrtOp

cbrt operation. Cube root of the specified value.

CbrtOpBuilder

Builder for cbrt operation. Cube root of the specified value.

CeilOp

ceil operation. Ceiling of the specified value.

CeilOpBuilder

Builder for ceil operation. Ceiling of the specified value.

CopySignOp

copysign operation. A copysign operation.

CopySignOpBuilder

Builder for copysign operation. A copysign operation.

CosOp

cos operation. Cosine of the specified value.

CosOpBuilder

Builder for cos operation. Cosine of the specified value.

CountLeadingZerosOp

ctlz operation. Counts the leading zeros an integer value.

CountLeadingZerosOpBuilder

Builder for ctlz operation. Counts the leading zeros an integer value.

CountTrailingZerosOp

cttz operation. Counts the trailing zeros an integer value.

CountTrailingZerosOpBuilder

Builder for cttz operation. Counts the trailing zeros an integer value.

CtPopOp

ctpop operation. Counts the number of set bits of an integer value.

CtPopOpBuilder

Builder for ctpop operation. Counts the number of set bits of an integer value.

ErfOp

erf operation. Error function of the specified value.

ErfOpBuilder

Builder for erf operation. Error function of the specified value.

Exp2Op

exp2 operation. Base-2 exponential of the specified value.

Exp2OpBuilder

Builder for exp2 operation. Base-2 exponential of the specified value.

ExpM1Op

expm1 operation. Base-e exponential of the specified value minus 1.

ExpM1OpBuilder

Builder for expm1 operation. Base-e exponential of the specified value minus 1.

ExpOp

exp operation. Base-e exponential of the specified value.

ExpOpBuilder

Builder for exp operation. Base-e exponential of the specified value.

FPowIOp

fpowi operation. Floating point raised to the signed integer power.

FPowIOpBuilder

Builder for fpowi operation. Floating point raised to the signed integer power.

FloorOp

floor operation. Floor of the specified value.

FloorOpBuilder

Builder for floor operation. Floor of the specified value.

FmaOp

fma operation. Floating point fused multipy-add operation.

FmaOpBuilder

Builder for fma operation. Floating point fused multipy-add operation.

IPowIOp

ipowi operation. Signed integer raised to the power of operation.

IPowIOpBuilder

Builder for ipowi operation. Signed integer raised to the power of operation.

Log1pOp

log1p operation. Computes the natural logarithm of one plus the given value.

Log1pOpBuilder

Builder for log1p operation. Computes the natural logarithm of one plus the given value.

Log2Op

log2 operation. Base-2 logarithm of the specified value.

Log2OpBuilder

Builder for log2 operation. Base-2 logarithm of the specified value.

Log10Op

log10 operation. Base-10 logarithm of the specified value.

Log10OpBuilder

Builder for log10 operation. Base-10 logarithm of the specified value.

LogOp

log operation. Base-e logarithm of the specified value.

LogOpBuilder

Builder for log operation. Base-e logarithm of the specified value.

PowFOp

powf operation. Floating point raised to the power of operation.

PowFOpBuilder

Builder for powf operation. Floating point raised to the power of operation.

RoundEvenOp

roundeven operation. Round of the specified value with halfway cases to even.

RoundEvenOpBuilder

Builder for roundeven operation. Round of the specified value with halfway cases to even.

RoundOp

round operation. Round of the specified value.

RoundOpBuilder

Builder for round operation. Round of the specified value.

RsqrtOp

rsqrt operation. Reciprocal of sqrt (1 / sqrt of the specified value).

RsqrtOpBuilder

Builder for rsqrt operation. Reciprocal of sqrt (1 / sqrt of the specified value).

SinOp

sin operation. Sine of the specified value.

SinOpBuilder

Builder for sin operation. Sine of the specified value.

SqrtOp

sqrt operation. Sqrt of the specified value.

SqrtOpBuilder

Builder for sqrt operation. Sqrt of the specified value.

TanOp

tan operation. Tangent of the specified value.

TanOpBuilder

Builder for tan operation. Tangent of the specified value.

TanhOp

tanh operation. Hyperbolic tangent of the specified value.

TanhOpBuilder

Builder for tanh operation. Hyperbolic tangent of the specified value.

TruncOp

trunc operation. Trunc of the specified value.

TruncOpBuilder

Builder for trunc operation. Trunc of the specified value.

Functions

absf

Creates a new absf operation. Floating point absolute-value operation.

absi

Creates a new absi operation. Integer absolute-value operation.

atan

Creates a new atan operation. Arcus tangent of the given value.

atan_2

Creates a new atan2 operation. 2-argument arcus tangent of the given values.

cbrt

Creates a new cbrt operation. Cube root of the specified value.

ceil

Creates a new ceil operation. Ceiling of the specified value.

copysign

Creates a new copysign operation. A copysign operation.

cos

Creates a new cos operation. Cosine of the specified value.

ctlz

Creates a new ctlz operation. Counts the leading zeros an integer value.

ctpop

Creates a new ctpop operation. Counts the number of set bits of an integer value.

cttz

Creates a new cttz operation. Counts the trailing zeros an integer value.

erf

Creates a new erf operation. Error function of the specified value.

exp

Creates a new exp operation. Base-e exponential of the specified value.

exp_2

Creates a new exp2 operation. Base-2 exponential of the specified value.

expm_1

Creates a new expm1 operation. Base-e exponential of the specified value minus 1.

floor

Creates a new floor operation. Floor of the specified value.

fma

Creates a new fma operation. Floating point fused multipy-add operation.

fpowi

Creates a new fpowi operation. Floating point raised to the signed integer power.

ipowi

Creates a new ipowi operation. Signed integer raised to the power of operation.

log

Creates a new log operation. Base-e logarithm of the specified value.

log_1_p

Creates a new log1p operation. Computes the natural logarithm of one plus the given value.

log_2

Creates a new log2 operation. Base-2 logarithm of the specified value.

log_10

Creates a new log10 operation. Base-10 logarithm of the specified value.

powf

Creates a new powf operation. Floating point raised to the power of operation.

round

Creates a new round operation. Round of the specified value.

roundeven

Creates a new roundeven operation. Round of the specified value with halfway cases to even.

rsqrt

Creates a new rsqrt operation. Reciprocal of sqrt (1 / sqrt of the specified value).

sin

Creates a new sin operation. Sine of the specified value.

sqrt

Creates a new sqrt operation. Sqrt of the specified value.

tan

Creates a new tan operation. Tangent of the specified value.

tanh

Creates a new tanh operation. Hyperbolic tangent of the specified value.

trunc

Creates a new trunc operation. Trunc of the specified value.