Module melior::dialect::ods::pdl

pdl dialect.

PDL presents a high level abstraction for the rewrite pattern infrastructure available in MLIR. This abstraction allows for representing patterns transforming MLIR, as MLIR. This allows for applying all of the benefits that the general MLIR infrastructure provides, to the infrastructure itself. This means that pattern matching can be more easily verified for correctness, targeted by frontends, and optimized.

PDL abstracts over various different aspects of patterns and core MLIR data structures. Patterns are specified via a pdl.pattern operation. These operations contain a region body for the “matcher” code, and terminate with a pdl.rewrite that either dispatches to an external rewriter or contains a region for the rewrite specified via pdl. The types of values in pdl are handle types to MLIR C++ types, with !pdl.attribute, !pdl.operation, !pdl.value, and !pdl.type directly mapping to mlir::Attribute, mlir::Operation*, mlir::Value, and mlir::Type respectively.

An example pattern is shown below:

// pdl.pattern contains metadata similarly to a `RewritePattern`.
pdl.pattern : benefit(1) {
  // External input operand values are specified via `pdl.operand` operations.
  // Result types are constrainted via `pdl.type` operations.

  %resultType = pdl.type
  %inputOperand = pdl.operand
  %root = pdl.operation "foo.op"(%inputOperand) -> %resultType
  pdl.rewrite %root {
    pdl.replace %root with (%inputOperand)
  }
}

The above pattern simply replaces an operation with its first operand. Note how the input operation is specified structurally, similarly to how it would look in memory. This is a simple example and pdl provides support for many other features such as applying external constraints or external generator methods. These features and more are detailed below.

Structs

ApplyNativeConstraintOperation

An apply_native_constraint operation. Apply a native constraint to a set of provided entities.

ApplyNativeConstraintOperationBuilder

A builder for an apply_native_constraint operation.

ApplyNativeRewriteOperation

An apply_native_rewrite operation. Apply a native rewrite method inside of pdl.rewrite region.

ApplyNativeRewriteOperationBuilder

A builder for an apply_native_rewrite operation.

AttributeOperation

An attribute operation. Define an input attribute in a pattern.

AttributeOperationBuilder

A builder for an attribute operation.

EraseOperation

An erase operation. Mark an input operation as erased.

EraseOperationBuilder

A builder for an erase operation.

OperandOperation

An operand operation. Define an external input operand in a pattern.

OperandOperationBuilder

A builder for an operand operation.

OperandsOperation

An operands operation. Define a range of input operands in a pattern.

OperandsOperationBuilder

A builder for an operands operation.

OperationOperation

An operation operation. Define an operation within a pattern.

OperationOperationBuilder

A builder for an operation operation.

PatternOperation

A pattern operation. Define a rewrite pattern.

PatternOperationBuilder

A builder for a pattern operation.

RangeOperation

A range operation. Construct a range of pdl entities.

RangeOperationBuilder

A builder for a range operation.

ReplaceOperation

A replace operation. Mark an input operation as replaced.

ReplaceOperationBuilder

A builder for a replace operation.

ResultOperation

A result operation. Extract a result from an operation.

ResultOperationBuilder

A builder for a result operation.

ResultsOperation

A results operation. Extract a result group from an operation.

ResultsOperationBuilder

A builder for a results operation.

RewriteOperation

A rewrite operation. Specify the rewrite of a matched pattern.

RewriteOperationBuilder

A builder for a rewrite operation.

TypeOperation

A type operation. Define a type handle within a pattern.

TypeOperationBuilder

A builder for a type operation.

TypesOperation

A types operation. Define a range of type handles within a pattern.

TypesOperationBuilder

A builder for a types operation.

Functions

_operation

Creates an operation operation.

apply_native_constraint

Creates an apply_native_constraint operation.

apply_native_rewrite

Creates an apply_native_rewrite operation.

attribute

Creates an attribute operation.

erase

Creates an erase operation.

operand

Creates an operand operation.

operands

Creates an operands operation.

pattern

Creates a pattern operation.

range

Creates a range operation.

replace

Creates a replace operation.

result

Creates a result operation.

results

Creates a results operation.

rewrite

Creates a rewrite operation.

type

Creates a type operation.

types

Creates a types operation.