Module melior::dialect::ods::pdl_interp

pdl_interp dialect.

The PDL Interpreter dialect provides a lower level abstraction compared to the PDL dialect, and is targeted towards low level optimization and interpreter code generation. The dialect operations encapsulates low-level pattern match and rewrite “primitives”, such as navigating the IR (Operation::getOperand), creating new operations (OpBuilder::create), etc. Many of the operations within this dialect also fuse branching control flow with some form of a predicate comparison operation. This type of fusion reduces the amount of work that an interpreter must do when executing.

Structs

ApplyConstraintOp

apply_constraint operation. Apply a constraint to a set of positional values.

ApplyConstraintOpBuilder

Builder for apply_constraint operation. Apply a constraint to a set of positional values.

ApplyRewriteOp

apply_rewrite operation. Invoke and apply an externally registered rewrite method.

ApplyRewriteOpBuilder

Builder for apply_rewrite operation. Invoke and apply an externally registered rewrite method.

AreEqualOp

are_equal operation. Check if two positional values or ranges are equivalent.

AreEqualOpBuilder

Builder for are_equal operation. Check if two positional values or ranges are equivalent.

BranchOp

branch operation. General branch operation.

BranchOpBuilder

Builder for branch operation. General branch operation.

CheckAttributeOp

check_attribute operation. Check the value of an Attribute.

CheckAttributeOpBuilder

Builder for check_attribute operation. Check the value of an Attribute.

CheckOperandCountOp

check_operand_count operation. Check the number of operands of an Operation.

CheckOperandCountOpBuilder

Builder for check_operand_count operation. Check the number of operands of an Operation.

CheckOperationNameOp

check_operation_name operation. Check the OperationName of an Operation.

CheckOperationNameOpBuilder

Builder for check_operation_name operation. Check the OperationName of an Operation.

CheckResultCountOp

check_result_count operation. Check the number of results of an Operation.

CheckResultCountOpBuilder

Builder for check_result_count operation. Check the number of results of an Operation.

CheckTypeOp

check_type operation. Compare a type to a known value.

CheckTypeOpBuilder

Builder for check_type operation. Compare a type to a known value.

CheckTypesOp

check_types operation. Compare a range of types to a range of known values.

CheckTypesOpBuilder

Builder for check_types operation. Compare a range of types to a range of known values.

ContinueOp

continue operation. Breaks the current iteration.

ContinueOpBuilder

Builder for continue operation. Breaks the current iteration.

CreateAttributeOp

create_attribute operation. Create an interpreter handle to a constant Attribute.

CreateAttributeOpBuilder

Builder for create_attribute operation. Create an interpreter handle to a constant Attribute.

CreateOperationOp

create_operation operation. Create an instance of a specific Operation.

CreateOperationOpBuilder

Builder for create_operation operation. Create an instance of a specific Operation.

CreateRangeOp

create_range operation. Construct a range of PDL entities.

CreateRangeOpBuilder

Builder for create_range operation. Construct a range of PDL entities.

CreateTypeOp

create_type operation. Create an interpreter handle to a constant Type.

CreateTypeOpBuilder

Builder for create_type operation. Create an interpreter handle to a constant Type.

CreateTypesOp

create_types operation. Create an interpreter handle to a range of constant Types.

CreateTypesOpBuilder

Builder for create_types operation. Create an interpreter handle to a range of constant Types.

EraseOp

erase operation. Mark an operation as erased.

EraseOpBuilder

Builder for erase operation. Mark an operation as erased.

ExtractOp

extract operation. Extract the item at the specified index in a range.

ExtractOpBuilder

Builder for extract operation. Extract the item at the specified index in a range.

FinalizeOp

finalize operation. Finalize a pattern match or rewrite sequence.

FinalizeOpBuilder

Builder for finalize operation. Finalize a pattern match or rewrite sequence.

ForEachOp

foreach operation. Iterates over a range of values or ranges.

ForEachOpBuilder

Builder for foreach operation. Iterates over a range of values or ranges.

FuncOp

func operation. PDL Interpreter Function Operation.

FuncOpBuilder

Builder for func operation. PDL Interpreter Function Operation.

GetAttributeOp

get_attribute operation. Get a specified attribute value from an Operation.

GetAttributeOpBuilder

Builder for get_attribute operation. Get a specified attribute value from an Operation.

GetAttributeTypeOp

get_attribute_type operation. Get the result type of a specified Attribute.

GetAttributeTypeOpBuilder

Builder for get_attribute_type operation. Get the result type of a specified Attribute.

GetDefiningOpOp

get_defining_op operation. Get the defining operation of a Value.

GetDefiningOpOpBuilder

Builder for get_defining_op operation. Get the defining operation of a Value.

GetOperandOp

get_operand operation. Get a specified operand from an Operation.

GetOperandOpBuilder

Builder for get_operand operation. Get a specified operand from an Operation.

GetOperandsOp

get_operands operation. Get a specified operand group from an Operation.

GetOperandsOpBuilder

Builder for get_operands operation. Get a specified operand group from an Operation.

GetResultOp

get_result operation. Get a specified result from an Operation.

GetResultOpBuilder

Builder for get_result operation. Get a specified result from an Operation.

GetResultsOp

get_results operation. Get a specified result group from an Operation.

GetResultsOpBuilder

Builder for get_results operation. Get a specified result group from an Operation.

GetUsersOp

get_users operation. Get the users of a Value.

GetUsersOpBuilder

Builder for get_users operation. Get the users of a Value.

GetValueTypeOp

get_value_type operation. Get the result type of a specified Value.

GetValueTypeOpBuilder

Builder for get_value_type operation. Get the result type of a specified Value.

IsNotNullOp

is_not_null operation. Check if a positional value is non-null.

IsNotNullOpBuilder

Builder for is_not_null operation. Check if a positional value is non-null.

RecordMatchOp

record_match operation. Record the metadata for a successful pattern match.

RecordMatchOpBuilder

Builder for record_match operation. Record the metadata for a successful pattern match.

ReplaceOp

replace operation. Mark an operation as replaced.

ReplaceOpBuilder

Builder for replace operation. Mark an operation as replaced.

SwitchAttributeOp

switch_attribute operation. Switch on the value of an Attribute.

SwitchAttributeOpBuilder

Builder for switch_attribute operation. Switch on the value of an Attribute.

SwitchOperandCountOp

switch_operand_count operation. Switch on the operand count of an Operation.

SwitchOperandCountOpBuilder

Builder for switch_operand_count operation. Switch on the operand count of an Operation.

SwitchOperationNameOp

switch_operation_name operation. Switch on the OperationName of an Operation.

SwitchOperationNameOpBuilder

Builder for switch_operation_name operation. Switch on the OperationName of an Operation.

SwitchResultCountOp

switch_result_count operation. Switch on the result count of an Operation.

SwitchResultCountOpBuilder

Builder for switch_result_count operation. Switch on the result count of an Operation.

SwitchTypeOp

switch_type operation. Switch on a Type value.

SwitchTypeOpBuilder

Builder for switch_type operation. Switch on a Type value.

SwitchTypesOp

switch_types operation. Switch on a range of Type values.

SwitchTypesOpBuilder

Builder for switch_types operation. Switch on a range of Type values.

Functions

apply_constraint

Creates a new apply_constraint operation. Apply a constraint to a set of positional values.

apply_rewrite

Creates a new apply_rewrite operation. Invoke and apply an externally registered rewrite method.

are_equal

Creates a new are_equal operation. Check if two positional values or ranges are equivalent.

branch

Creates a new branch operation. General branch operation.

check_attribute

Creates a new check_attribute operation. Check the value of an Attribute.

check_operand_count

Creates a new check_operand_count operation. Check the number of operands of an Operation.

check_operation_name

Creates a new check_operation_name operation. Check the OperationName of an Operation.

check_result_count

Creates a new check_result_count operation. Check the number of results of an Operation.

check_type

Creates a new check_type operation. Compare a type to a known value.

check_types

Creates a new check_types operation. Compare a range of types to a range of known values.

continue

Creates a new continue operation. Breaks the current iteration.

create_attribute

Creates a new create_attribute operation. Create an interpreter handle to a constant Attribute.

create_operation

Creates a new create_operation operation. Create an instance of a specific Operation.

create_range

Creates a new create_range operation. Construct a range of PDL entities.

create_type

Creates a new create_type operation. Create an interpreter handle to a constant Type.

create_types

Creates a new create_types operation. Create an interpreter handle to a range of constant Types.

erase

Creates a new erase operation. Mark an operation as erased.

extract

Creates a new extract operation. Extract the item at the specified index in a range.

finalize

Creates a new finalize operation. Finalize a pattern match or rewrite sequence.

foreach

Creates a new foreach operation. Iterates over a range of values or ranges.

func

Creates a new func operation. PDL Interpreter Function Operation.

get_attribute

Creates a new get_attribute operation. Get a specified attribute value from an Operation.

get_attribute_type

Creates a new get_attribute_type operation. Get the result type of a specified Attribute.

get_defining_op

Creates a new get_defining_op operation. Get the defining operation of a Value.

get_operand

Creates a new get_operand operation. Get a specified operand from an Operation.

get_operands

Creates a new get_operands operation. Get a specified operand group from an Operation.

get_result

Creates a new get_result operation. Get a specified result from an Operation.

get_results

Creates a new get_results operation. Get a specified result group from an Operation.

get_users

Creates a new get_users operation. Get the users of a Value.

get_value_type

Creates a new get_value_type operation. Get the result type of a specified Value.

is_not_null

Creates a new is_not_null operation. Check if a positional value is non-null.

record_match

Creates a new record_match operation. Record the metadata for a successful pattern match.

replace

Creates a new replace operation. Mark an operation as replaced.

switch_attribute

Creates a new switch_attribute operation. Switch on the value of an Attribute.

switch_operand_count

Creates a new switch_operand_count operation. Switch on the operand count of an Operation.

switch_operation_name

Creates a new switch_operation_name operation. Switch on the OperationName of an Operation.

switch_result_count

Creates a new switch_result_count operation. Switch on the result count of an Operation.

switch_type

Creates a new switch_type operation. Switch on a Type value.

switch_types

Creates a new switch_types operation. Switch on a range of Type values.