Module melior::dialect::ods::shape

shape dialect.

This dialect contains operations for shape inference.

Note: Unless explicitly stated, all functions that return a shape and take shapes as input, return the invalid shape if one of its operands is an invalid shape. This avoids flagging multiple errors for one verification failure. The dialect itself does not specify how errors should be combined (there are multiple different options, from always choosing first operand, concatting etc. on how to combine them).

Structs

AddOp

add operation. Addition of sizes and indices.

AddOpBuilder

Builder for add operation. Addition of sizes and indices.

AnyOp

any operation. Return any combination of the input shapes.

AnyOpBuilder

Builder for any operation. Return any combination of the input shapes.

AssumingAllOp

assuming_all operation. Return a logical AND of all witnesses.

AssumingAllOpBuilder

Builder for assuming_all operation. Return a logical AND of all witnesses.

AssumingOp

assuming operation. Execute the region.

AssumingOpBuilder

Builder for assuming operation. Execute the region.

AssumingYieldOp

assuming_yield operation. Yield operation.

AssumingYieldOpBuilder

Builder for assuming_yield operation. Yield operation.

BroadcastOp

broadcast operation. Returns the broadcasted output shape of two or more inputs.

BroadcastOpBuilder

Builder for broadcast operation. Returns the broadcasted output shape of two or more inputs.

ConcatOp

concat operation. Concatenates two shapes.

ConcatOpBuilder

Builder for concat operation. Concatenates two shapes.

ConstShapeOp

const_shape operation. Creates a constant shape or extent tensor.

ConstShapeOpBuilder

Builder for const_shape operation. Creates a constant shape or extent tensor.

ConstSizeOp

const_size operation. Creates a constant of type shape.size.

ConstSizeOpBuilder

Builder for const_size operation. Creates a constant of type shape.size.

ConstWitnessOp

const_witness operation. An operation that returns a statically known witness value.

ConstWitnessOpBuilder

Builder for const_witness operation. An operation that returns a statically known witness value.

CstrBroadcastableOp

cstr_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

CstrBroadcastableOpBuilder

Builder for cstr_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

CstrEqOp

cstr_eq operation. Determines if all input shapes are equal.

CstrEqOpBuilder

Builder for cstr_eq operation. Determines if all input shapes are equal.

CstrRequireOp

cstr_require operation. Represents a runtime assertion that an i1 is true.

CstrRequireOpBuilder

Builder for cstr_require operation. Represents a runtime assertion that an i1 is true.

DebugPrintOp

debug_print operation. Prints the input shape or size.

DebugPrintOpBuilder

Builder for debug_print operation. Prints the input shape or size.

DimOp

dim operation. Gets the specified extent from the shape of a shaped input.

DimOpBuilder

Builder for dim operation. Gets the specified extent from the shape of a shaped input.

DivOp

div operation. Division of sizes and indices.

DivOpBuilder

Builder for div operation. Division of sizes and indices.

FromExtentTensorOp

from_extent_tensor operation. Creates a shape from a tensor of extents.

FromExtentTensorOpBuilder

Builder for from_extent_tensor operation. Creates a shape from a tensor of extents.

FromExtentsOp

from_extents operation. Creates a shape from extents.

FromExtentsOpBuilder

Builder for from_extents operation. Creates a shape from extents.

FuncOp

func operation. Shape function.

FuncOpBuilder

Builder for func operation. Shape function.

FunctionLibraryOp

function_library operation. Represents shape functions and corresponding ops.

FunctionLibraryOpBuilder

Builder for function_library operation. Represents shape functions and corresponding ops.

GetExtentOp

get_extent operation. Gets the specified extent from a shape or extent tensor.

GetExtentOpBuilder

Builder for get_extent operation. Gets the specified extent from a shape or extent tensor.

IndexToSizeOp

index_to_size operation. Converts a standard index to a shape size.

IndexToSizeOpBuilder

Builder for index_to_size operation. Converts a standard index to a shape size.

IsBroadcastableOp

is_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

IsBroadcastableOpBuilder

Builder for is_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

MaxOp

max operation. Elementwise maximum.

MaxOpBuilder

Builder for max operation. Elementwise maximum.

MeetOp

meet operation. Returns the least general shape or size of its operands.

MeetOpBuilder

Builder for meet operation. Returns the least general shape or size of its operands.

MinOp

min operation. Elementwise minimum.

MinOpBuilder

Builder for min operation. Elementwise minimum.

MulOp

mul operation. Multiplication of sizes and indices.

MulOpBuilder

Builder for mul operation. Multiplication of sizes and indices.

NumElementsOp

num_elements operation. Returns the number of elements for a given shape.

NumElementsOpBuilder

Builder for num_elements operation. Returns the number of elements for a given shape.

RankOp

rank operation. Gets the rank of a shape.

RankOpBuilder

Builder for rank operation. Gets the rank of a shape.

ReduceOp

reduce operation. Returns an expression reduced over a shape or extent tensor.

ReduceOpBuilder

Builder for reduce operation. Returns an expression reduced over a shape or extent tensor.

ReturnOp

return operation. Shape function return operation.

ReturnOpBuilder

Builder for return operation. Shape function return operation.

ShapeEqOp

shape_eq operation. Returns whether the input shapes or extent tensors are equal.

ShapeEqOpBuilder

Builder for shape_eq operation. Returns whether the input shapes or extent tensors are equal.

ShapeOfOp

shape_of operation. Returns shape of a value or shaped type operand.

ShapeOfOpBuilder

Builder for shape_of operation. Returns shape of a value or shaped type operand.

SizeToIndexOp

size_to_index operation. Casts between index types of the shape and standard dialect.

SizeToIndexOpBuilder

Builder for size_to_index operation. Casts between index types of the shape and standard dialect.

SplitAtOp

split_at operation. Splits a shape at a given index.

SplitAtOpBuilder

Builder for split_at operation. Splits a shape at a given index.

ToExtentTensorOp

to_extent_tensor operation. Creates a dimension tensor from a shape.

ToExtentTensorOpBuilder

Builder for to_extent_tensor operation. Creates a dimension tensor from a shape.

ValueAsShapeOp

value_as_shape operation. Returns value as a shape.

ValueAsShapeOpBuilder

Builder for value_as_shape operation. Returns value as a shape.

ValueOfOp

value_of operation. Returns value of a !shape.value_shape operand.

ValueOfOpBuilder

Builder for value_of operation. Returns value of a !shape.value_shape operand.

WithOp

with_shape operation. Returns ValueShape with given shape.

WithOpBuilder

Builder for with_shape operation. Returns ValueShape with given shape.

YieldOp

yield operation. Returns the value to parent op.

YieldOpBuilder

Builder for yield operation. Returns the value to parent op.

Functions

add

Creates a new add operation. Addition of sizes and indices.

any

Creates a new any operation. Return any combination of the input shapes.

assuming

Creates a new assuming operation. Execute the region.

assuming_all

Creates a new assuming_all operation. Return a logical AND of all witnesses.

assuming_yield

Creates a new assuming_yield operation. Yield operation.

broadcast

Creates a new broadcast operation. Returns the broadcasted output shape of two or more inputs.

concat

Creates a new concat operation. Concatenates two shapes.

const_shape

Creates a new const_shape operation. Creates a constant shape or extent tensor.

const_size

Creates a new const_size operation. Creates a constant of type shape.size.

const_witness

Creates a new const_witness operation. An operation that returns a statically known witness value.

cstr_broadcastable

Creates a new cstr_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

cstr_eq

Creates a new cstr_eq operation. Determines if all input shapes are equal.

cstr_require

Creates a new cstr_require operation. Represents a runtime assertion that an i1 is true.

debug_print

Creates a new debug_print operation. Prints the input shape or size.

dim

Creates a new dim operation. Gets the specified extent from the shape of a shaped input.

div

Creates a new div operation. Division of sizes and indices.

from_extent_tensor

Creates a new from_extent_tensor operation. Creates a shape from a tensor of extents.

from_extents

Creates a new from_extents operation. Creates a shape from extents.

func

Creates a new func operation. Shape function.

function_library

Creates a new function_library operation. Represents shape functions and corresponding ops.

get_extent

Creates a new get_extent operation. Gets the specified extent from a shape or extent tensor.

index_to_size

Creates a new index_to_size operation. Converts a standard index to a shape size.

is_broadcastable

Creates a new is_broadcastable operation. Determines if 2+ shapes can be successfully broadcasted.

max

Creates a new max operation. Elementwise maximum.

meet

Creates a new meet operation. Returns the least general shape or size of its operands.

min

Creates a new min operation. Elementwise minimum.

mul

Creates a new mul operation. Multiplication of sizes and indices.

num_elements

Creates a new num_elements operation. Returns the number of elements for a given shape.

rank

Creates a new rank operation. Gets the rank of a shape.

reduce

Creates a new reduce operation. Returns an expression reduced over a shape or extent tensor.

return

Creates a new return operation. Shape function return operation.

shape_eq

Creates a new shape_eq operation. Returns whether the input shapes or extent tensors are equal.

shape_of

Creates a new shape_of operation. Returns shape of a value or shaped type operand.

size_to_index

Creates a new size_to_index operation. Casts between index types of the shape and standard dialect.

split_at

Creates a new split_at operation. Splits a shape at a given index.

to_extent_tensor

Creates a new to_extent_tensor operation. Creates a dimension tensor from a shape.

value_as_shape

Creates a new value_as_shape operation. Returns value as a shape.

value_of

Creates a new value_of operation. Returns value of a !shape.value_shape operand.

with_shape

Creates a new with_shape operation. Returns ValueShape with given shape.

yield

Creates a new yield operation. Returns the value to parent op.