Module melior::dialect::ods::sparse_tensor

sparse_tensor dialect.

The SparseTensor dialect supports all the attributes, types, operations, and passes that are required to make sparse tensor types first class citizens within the MLIR compiler infrastructure. The dialect forms a bridge between high-level operations on sparse tensors types and lower-level operations on the actual sparse storage schemes consisting of positions, coordinates, and values. Lower-level support may consist of fully generated code or may be provided by means of a small sparse runtime support library.

The concept of treating sparsity as a property, not a tedious implementation detail, by letting a sparse compiler generate sparse code automatically was pioneered for linear algebra by [Bik96] in MT1 (see https://www.aartbik.com/sparse.php) and formalized to tensor algebra by [Kjolstad17,Kjolstad20] in the Sparse Tensor Algebra Compiler (TACO) project (see http://tensor-compiler.org).

The MLIR implementation [Biketal22] closely follows the “sparse iteration theory” that forms the foundation of TACO. A rewriting rule is applied to each tensor expression in the Linalg dialect (MLIR’s tensor index notation) where the sparsity of tensors is indicated using the per-level level-types (e.g., dense, compressed, singleton) together with a specification of the order on the levels (see [Chou18] for an in-depth discussions and possible extensions to these level-types). Subsequently, a topologically sorted iteration graph, reflecting the required order on coordinates with respect to the levels of each tensor, is constructed to ensure that all tensors are visited in natural level-coordinate order. Next, iteration lattices are constructed for the tensor expression for every index in topological order. Each iteration lattice point consists of a conjunction of tensor coordinates together with a tensor (sub)expression that needs to be evaluated for that conjunction. Within the lattice, iteration points are ordered according to the way coordinates are exhausted. As such these iteration lattices drive actual sparse code generation, which consists of a relatively straightforward one-to-one mapping from iteration lattices to combinations of for-loops, while-loops, and if-statements. Sparse tensor outputs that materialize uninitialized are handled with direct insertions if all parallel loops are outermost or insertions that indirectly go through a 1-dimensional access pattern expansion (a.k.a. workspace) where feasible [Gustavson72,Bik96,Kjolstad19].

[Bik96] Aart J.C. Bik. Compiler Support for Sparse Matrix Computations. PhD thesis, Leiden University, May 1996.

[Biketal22] Aart J.C. Bik, Penporn Koanantakool, Tatiana Shpeisman, Nicolas Vasilache, Bixia Zheng, and Fredrik Kjolstad. Compiler Support for Sparse Tensor Computations in MLIR. ACM Transactions on Architecture and Code Optimization, June, 2022. See: https://dl.acm.org/doi/10.1145/3544559

[Chou18] Stephen Chou, Fredrik Berg Kjolstad, and Saman Amarasinghe. Format Abstraction for Sparse Tensor Algebra Compilers. Proceedings of the ACM on Programming Languages, October 2018.

[Chou20] Stephen Chou, Fredrik Berg Kjolstad, and Saman Amarasinghe. Automatic Generation of Efficient Sparse Tensor Format Conversion Routines. Proceedings of the 41st ACM SIGPLAN Conference on Programming Language Design and Implementation, June, 2020.

[Gustavson72] Fred G. Gustavson. Some basic techniques for solving sparse systems of linear equations. In Sparse Matrices and Their Applications, pages 41–52. Plenum Press, New York, 1972.

[Kjolstad17] Fredrik Berg Kjolstad, Shoaib Ashraf Kamil, Stephen Chou, David Lugato, and Saman Amarasinghe. The Tensor Algebra Compiler. Proceedings of the ACM on Programming Languages, October 2017.

[Kjolstad19] Fredrik Berg Kjolstad, Peter Ahrens, Shoaib Ashraf Kamil, and Saman Amarasinghe. Tensor Algebra Compilation with Workspaces, Proceedings of the IEEE/ACM International Symposium on Code Generation and Optimization, 2019.

[Kjolstad20] Fredrik Berg Kjolstad. Sparse Tensor Algebra Compilation. PhD thesis, MIT, February, 2020.

Structs

BinaryOperation

A(n) binary operation. Binary set operation utilized within linalg.generic.

BinaryOperationBuilder

A builder for a(n) binary operation.

CompressOperation

A(n) compress operation. Compressed an access pattern for insertion.

CompressOperationBuilder

A builder for a(n) compress operation.

ConcatenateOperation

A(n) concatenate operation. Concatenates a list of tensors into a single tensor..

ConcatenateOperationBuilder

A builder for a(n) concatenate operation.

ConvertOperation

A(n) convert operation. Converts between different tensor types.

ConvertOperationBuilder

A builder for a(n) convert operation.

ExpandOperation

A(n) expand operation. Expands an access pattern for insertion.

ExpandOperationBuilder

A builder for a(n) expand operation.

ForeachOperation

A(n) foreach operation. Iterates over elements in a tensor.

ForeachOperationBuilder

A builder for a(n) foreach operation.

GetStorageSpecifierOperation

A(n) storage_specifier.get operation.

GetStorageSpecifierOperationBuilder

A builder for a(n) storage_specifier.get operation.

InsertOperation

A(n) insert operation. Inserts a value into the sparse tensor.

InsertOperationBuilder

A builder for a(n) insert operation.

LoadOperation

A(n) load operation. Rematerializes tensor from underlying sparse storage format.

LoadOperationBuilder

A builder for a(n) load operation.

NewOperation

A(n) new operation. Materializes a new sparse tensor from given source.

NewOperationBuilder

A builder for a(n) new operation.

NumberOfEntriesOperation

A(n) number_of_entries operation. Returns the number of entries that are stored in the tensor..

NumberOfEntriesOperationBuilder

A builder for a(n) number_of_entries operation.

OutOperation

A(n) out operation. Outputs a sparse tensor to the given destination.

OutOperationBuilder

A builder for a(n) out operation.

PackOperation

A(n) pack operation. Returns a sparse tensor from the given values, levels.

PackOperationBuilder

A builder for a(n) pack operation.

PushBackOperation

A(n) push_back operation. Pushes a value to the back of a given buffer.

PushBackOperationBuilder

A builder for a(n) push_back operation.

ReduceOperation

A(n) reduce operation. Custom reduction operation utilized within linalg.generic.

ReduceOperationBuilder

A builder for a(n) reduce operation.

SelectOperation

A(n) select operation. Select operation utilized within linalg.generic.

SelectOperationBuilder

A builder for a(n) select operation.

SetStorageSpecifierOperation

A(n) storage_specifier.set operation.

SetStorageSpecifierOperationBuilder

A builder for a(n) storage_specifier.set operation.

SortCooOperation

A(n) sort_coo operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

SortCooOperationBuilder

A builder for a(n) sort_coo operation.

SortOperation

A(n) sort operation. Sorts the arrays in xs and ys lexicographically on the integral values found in the xs list.

SortOperationBuilder

A builder for a(n) sort operation.

StorageSpecifierInitOperation

A(n) storage_specifier.init operation.

StorageSpecifierInitOperationBuilder

A builder for a(n) storage_specifier.init operation.

ToCoordinatesBufferOperation

A(n) coordinates_buffer operation. Extracts the linear coordinates array from a tensor.

ToCoordinatesBufferOperationBuilder

A builder for a(n) coordinates_buffer operation.

ToCoordinatesOperation

A(n) coordinates operation. Extracts the level-th coordinates array of the tensor.

ToCoordinatesOperationBuilder

A builder for a(n) coordinates operation.

ToPositionsOperation

A(n) positions operation. Extracts the level-th positions array of the tensor.

ToPositionsOperationBuilder

A builder for a(n) positions operation.

ToSliceOffsetOperation

A(n) slice.offset operation. Extracts the offset of the sparse tensor slice at the given dimension.

ToSliceOffsetOperationBuilder

A builder for a(n) slice.offset operation.

ToSliceStrideOperation

A(n) slice.stride operation. Extracts the stride of the sparse tensor slice at the given dimension.

ToSliceStrideOperationBuilder

A builder for a(n) slice.stride operation.

ToValuesOperation

A(n) values operation. Extracts numerical values array from a tensor.

ToValuesOperationBuilder

A builder for a(n) values operation.

UnaryOperation

A(n) unary operation. Unary set operation utilized within linalg.generic.

UnaryOperationBuilder

A builder for a(n) unary operation.

UnpackOperation

A(n) unpack operation. Returns the (values, coordinates) pair unpacked from the input tensor.

UnpackOperationBuilder

A builder for a(n) unpack operation.

YieldOperation

A(n) yield operation. Yield from sparse_tensor set-like operations.

YieldOperationBuilder

A builder for a(n) yield operation.

Functions

binary

Creates a(n) binary operation.

compress

Creates a(n) compress operation.

concatenate

Creates a(n) concatenate operation.

convert

Creates a(n) convert operation.

coordinates

Creates a(n) coordinates operation.

coordinates_buffer

Creates a(n) coordinates_buffer operation.

expand

Creates a(n) expand operation.

foreach

Creates a(n) foreach operation.

insert

Creates a(n) insert operation.

load

Creates a(n) load operation.

new

Creates a(n) new operation.

number_of_entries

Creates a(n) number_of_entries operation.

out

Creates a(n) out operation.

pack

Creates a(n) pack operation.

positions

Creates a(n) positions operation.

push_back

Creates a(n) push_back operation.

reduce

Creates a(n) reduce operation.

select

Creates a(n) select operation.

slice_offset

Creates a(n) slice.offset operation.

slice_stride

Creates a(n) slice.stride operation.

sort

Creates a(n) sort operation.

sort_coo

Creates a(n) sort_coo operation.

storage_specifier_get

Creates a(n) storage_specifier.get operation.

storage_specifier_init

Creates a(n) storage_specifier.init operation.

storage_specifier_set

Creates a(n) storage_specifier.set operation.

unary

Creates a(n) unary operation.

unpack

Creates a(n) unpack operation.

values

Creates a(n) values operation.

yield

Creates a(n) yield operation.