Module melior::dialect::ods::tensor

tensor dialect.

The tensor dialect is intended to hold core tensor creation and manipulation ops, which are not strongly associated with any particular other dialect or domain abstraction. The primary smoke test of this is ops that make sense for any tensor element type.

We leave it to other dialects to hold the vast swath of possible computations one might want to do on a tensor.

The tensor type is (for better or for worse) used to represent all kinds of things, and supports an open-ended set of element types. Examples:

representing large, dense aggregations of primitive types, suitable for high-performance numerical computing.

representing shapes in the shape dialect, which consist of small 1D tensors of index data type.

representing aggregations of strings or “variant” types.

representing large, sparse aggregations of primitive types, suitable for high-performance numerical computing.

Thus, for the tensor dialect, we prefer for now to constrain the scope as much as possible. The expectation is that at some point in the future, the tensor dialect’s scope may be broadened through a careful discussion of the tradeoffs.

The tensor type is actually a builtin type (it lives in the builtin dialect), and does not live in this dialect.

Structs

CastOp

cast operation: Tensor cast operation

CastOpBuilder

Builder for cast operation: Tensor cast operation

CollapseShapeOp

collapse_shape operation: Operation to produce a tensor with a smaller rank

CollapseShapeOpBuilder

Builder for collapse_shape operation: Operation to produce a tensor with a smaller rank

DimOp

dim operation: Dimension index operation

DimOpBuilder

Builder for dim operation: Dimension index operation

EmptyOp

empty operation: Empty tensor operation

EmptyOpBuilder

Builder for empty operation: Empty tensor operation

ExpandShapeOp

expand_shape operation: Operation to produce a tensor with a higher rank

ExpandShapeOpBuilder

Builder for expand_shape operation: Operation to produce a tensor with a higher rank

ExtractOp

extract operation: Element extraction operation

ExtractOpBuilder

Builder for extract operation: Element extraction operation

ExtractSliceOp

extract_slice operation: Extract slice operation

ExtractSliceOpBuilder

Builder for extract_slice operation: Extract slice operation

FromElementsOp

from_elements operation: Tensor from elements operation.

FromElementsOpBuilder

Builder for from_elements operation: Tensor from elements operation.

GatherOp

gather operation: Gather a subset of a tensor at specified indices

GatherOpBuilder

Builder for gather operation: Gather a subset of a tensor at specified indices

GenerateOp

generate operation: Creates a dynamically sized tensor from elements

GenerateOpBuilder

Builder for generate operation: Creates a dynamically sized tensor from elements

InsertOp

insert operation: Element insertion operation

InsertOpBuilder

Builder for insert operation: Element insertion operation

InsertSliceOp

insert_slice operation: Insert_slice operation

InsertSliceOpBuilder

Builder for insert_slice operation: Insert_slice operation

PackOp

pack operation: Tensor pack operation

PackOpBuilder

Builder for pack operation: Tensor pack operation

PadOp

pad operation: Tensor pad operation

PadOpBuilder

Builder for pad operation: Tensor pad operation

ParallelInsertSliceOp

parallel_insert_slice operation: Specify the tensor slice update of a single thread of a parent ParallelCombiningOpInterface op.

ParallelInsertSliceOpBuilder

Builder for parallel_insert_slice operation: Specify the tensor slice update of a single thread of a parent ParallelCombiningOpInterface op.

RankOp

rank operation: Rank operation

RankOpBuilder

Builder for rank operation: Rank operation

ReshapeOp

reshape operation: Tensor reshape operation

ReshapeOpBuilder

Builder for reshape operation: Tensor reshape operation

ScatterOp

scatter operation: Scatter a tensor into a destination tensor at specified indices

ScatterOpBuilder

Builder for scatter operation: Scatter a tensor into a destination tensor at specified indices

SplatOp

splat operation: Tensor splat or broadcast operation

SplatOpBuilder

Builder for splat operation: Tensor splat or broadcast operation

UnPackOp

unpack operation: Tensor unpack operation

UnPackOpBuilder

Builder for unpack operation: Tensor unpack operation

YieldOp

yield operation: Yield a value from a region

YieldOpBuilder

Builder for yield operation: Yield a value from a region

Functions

cast

Creates a new cast operation: Tensor cast operation

collapse_shape

Creates a new collapse_shape operation: Operation to produce a tensor with a smaller rank

dim

Creates a new dim operation: Dimension index operation

empty

Creates a new empty operation: Empty tensor operation

expand_shape

Creates a new expand_shape operation: Operation to produce a tensor with a higher rank

extract

Creates a new extract operation: Element extraction operation

extract_slice

Creates a new extract_slice operation: Extract slice operation

from_elements

Creates a new from_elements operation: Tensor from elements operation.

gather

Creates a new gather operation: Gather a subset of a tensor at specified indices

generate

Creates a new generate operation: Creates a dynamically sized tensor from elements

insert

Creates a new insert operation: Element insertion operation

insert_slice

Creates a new insert_slice operation: Insert_slice operation

pack

Creates a new pack operation: Tensor pack operation

pad

Creates a new pad operation: Tensor pad operation

parallel_insert_slice

Creates a new parallel_insert_slice operation: Specify the tensor slice update of a single thread of a parent ParallelCombiningOpInterface op.

rank

Creates a new rank operation: Rank operation

reshape

Creates a new reshape operation: Tensor reshape operation

scatter

Creates a new scatter operation: Scatter a tensor into a destination tensor at specified indices

splat

Creates a new splat operation: Tensor splat or broadcast operation

unpack

Creates a new unpack operation: Tensor unpack operation

yield

Creates a new yield operation: Yield a value from a region