Module melior::dialect::ods::tosa

tosa dialect.

The Tensor Operator Set Architecture (TOSA) dialect.

This dialect implements the TOSA standard described at https://developer.mlplatform.org/w/tosa/ .

Tensor Operator Set Architecture (TOSA) provides a set of whole-tensor operations commonly employed by Deep Neural Networks. The intent is to enable a variety of implementations running on a diverse range of processors, with the results at the TOSA level consistent across those implementations. Applications or frameworks which target TOSA can therefore be deployed on a wide range of different processors, such as CPUs or GPUs, with defined accuracy and compatibility constraints. Most operators from the common ML frameworks should be expressible in TOSA. It is expected that there will be tools to lower from the ML frameworks into TOSA.

Structs

AbsOp

abs operation: Elementwise abs op

AbsOpBuilder

Builder for abs operation: Elementwise abs op

AddOp

add operation: Elementwise addition operator

AddOpBuilder

Builder for add operation: Elementwise addition operator

ApplyScaleOp

apply_scale operation: Rescale scalar operator for Tosa tensor operators

ApplyScaleOpBuilder

Builder for apply_scale operation: Rescale scalar operator for Tosa tensor operators

ArgMaxOp

argmax operation: Perform argmax on the input.

ArgMaxOpBuilder

Builder for argmax operation: Perform argmax on the input.

ArithmeticRightShiftOp

arithmetic_right_shift operation: Elementwise Arithmetic Right Shift

ArithmeticRightShiftOpBuilder

Builder for arithmetic_right_shift operation: Elementwise Arithmetic Right Shift

AvgPool2dOp

avg_pool2d operation: Performs max pooling on the input.

AvgPool2dOpBuilder

Builder for avg_pool2d operation: Performs max pooling on the input.

BitwiseAndOp

bitwise_and operation: Bitwise AND operator

BitwiseAndOpBuilder

Builder for bitwise_and operation: Bitwise AND operator

BitwiseNotOp

bitwise_not operation: Bitwise NOT operator

BitwiseNotOpBuilder

Builder for bitwise_not operation: Bitwise NOT operator

BitwiseOrOp

bitwise_or operation: Bitwise OR operator

BitwiseOrOpBuilder

Builder for bitwise_or operation: Bitwise OR operator

BitwiseXorOp

bitwise_xor operation: Bitwise XOR operator

BitwiseXorOpBuilder

Builder for bitwise_xor operation: Bitwise XOR operator

CastOp

cast operation: Cast operation

CastOpBuilder

Builder for cast operation: Cast operation

CeilOp

ceil operation: Elementwise ceil op

CeilOpBuilder

Builder for ceil operation: Elementwise ceil op

ClampOp

clamp operation: Computes clamp(features, min, max).

ClampOpBuilder

Builder for clamp operation: Computes clamp(features, min, max).

ClzOp

clz operation: Elementwise count leading zero op

ClzOpBuilder

Builder for clz operation: Elementwise count leading zero op

ConcatOp

concat operation: Concatenates tensors along one dimension.

ConcatOpBuilder

Builder for concat operation: Concatenates tensors along one dimension.

ConstOp

const operation: Constant op.

ConstOpBuilder

Builder for const operation: Constant op.

Conv2DOp

conv2d operation: 2D Convolution Operator

Conv2DOpBuilder

Builder for conv2d operation: 2D Convolution Operator

Conv3DOp

conv3d operation: 3D Convolution operator

Conv3DOpBuilder

Builder for conv3d operation: 3D Convolution operator

CustomOp

custom operation: Custom operator wrapper for Tosa

CustomOpBuilder

Builder for custom operation: Custom operator wrapper for Tosa

DepthwiseConv2DOp

depthwise_conv2d operation: Depthwise 2D Convolution operator

DepthwiseConv2DOpBuilder

Builder for depthwise_conv2d operation: Depthwise 2D Convolution operator

DivOp

div operation: Integer divide operator

DivOpBuilder

Builder for div operation: Integer divide operator

EqualOp

equal operation: Returns the truth value of (x == y) element-wise.

EqualOpBuilder

Builder for equal operation: Returns the truth value of (x == y) element-wise.

ExpOp

exp operation: Elementwise exp op

ExpOpBuilder

Builder for exp operation: Elementwise exp op

FloorOp

floor operation: Elementwise floor op

FloorOpBuilder

Builder for floor operation: Elementwise floor op

FullyConnectedOp

fully_connected operation: Fully Connected operator

FullyConnectedOpBuilder

Builder for fully_connected operation: Fully Connected operator

GatherOp

gather operation: Gather operation,

GatherOpBuilder

Builder for gather operation: Gather operation,

GreaterEqualOp

greater_equal operation: Returns the truth value of (x >= y) element-wise.

GreaterEqualOpBuilder

Builder for greater_equal operation: Returns the truth value of (x >= y) element-wise.

GreaterOp

greater operation: Returns the truth value of (x > y) element-wise.

GreaterOpBuilder

Builder for greater operation: Returns the truth value of (x > y) element-wise.

IdentityOp

identity operation: Identity operator

IdentityOpBuilder

Builder for identity operation: Identity operator

IfOp

cond_if operation: Conditional if operator

IfOpBuilder

Builder for cond_if operation: Conditional if operator

LogOp

log operation: Elementwise log op

LogOpBuilder

Builder for log operation: Elementwise log op

LogicalAndOp

logical_and operation: Returns the truth value of x AND y element-wise.

LogicalAndOpBuilder

Builder for logical_and operation: Returns the truth value of x AND y element-wise.

LogicalLeftShiftOp

logical_left_shift operation: Elementwise Logical Left Shift

LogicalLeftShiftOpBuilder

Builder for logical_left_shift operation: Elementwise Logical Left Shift

LogicalNotOp

logical_not operation: Returns the truth value of NOT x element-wise.

LogicalNotOpBuilder

Builder for logical_not operation: Returns the truth value of NOT x element-wise.

LogicalOrOp

logical_or operation: Returns the truth value of x OR y element-wise.

LogicalOrOpBuilder

Builder for logical_or operation: Returns the truth value of x OR y element-wise.

LogicalRightShiftOp

logical_right_shift operation: Elementwise Logical Right Shift

LogicalRightShiftOpBuilder

Builder for logical_right_shift operation: Elementwise Logical Right Shift

LogicalXorOp

logical_xor operation: Returns the truth value of x XOR y element-wise.

LogicalXorOpBuilder

Builder for logical_xor operation: Returns the truth value of x XOR y element-wise.

MatMulOp

matmul operation: Matrix multiplication with bias

MatMulOpBuilder

Builder for matmul operation: Matrix multiplication with bias

MaxPool2dOp

max_pool2d operation: Performs max pooling on the input.

MaxPool2dOpBuilder

Builder for max_pool2d operation: Performs max pooling on the input.

MaximumOp

maximum operation: Elementwise Maximum

MaximumOpBuilder

Builder for maximum operation: Elementwise Maximum

MinimumOp

minimum operation: Elementwise Minimum

MinimumOpBuilder

Builder for minimum operation: Elementwise Minimum

MulOp

mul operation: Multiplication operator

MulOpBuilder

Builder for mul operation: Multiplication operator

NegateOp

negate operation: Elementwise negate op

NegateOpBuilder

Builder for negate operation: Elementwise negate op

PadOp

pad operation: Pads a tensor with value specified.

PadOpBuilder

Builder for pad operation: Pads a tensor with value specified.

PowOp

pow operation: Computes the power of one value to another.

PowOpBuilder

Builder for pow operation: Computes the power of one value to another.

RFFT2dOp

rfft2d operation: Performs RFFT2D operation on the input.

RFFT2dOpBuilder

Builder for rfft2d operation: Performs RFFT2D operation on the input.

ReciprocalOp

reciprocal operation: Elementwise reciprocal op

ReciprocalOpBuilder

Builder for reciprocal operation: Elementwise reciprocal op

ReduceAllOp

reduce_all operation: Reduce All operator

ReduceAllOpBuilder

Builder for reduce_all operation: Reduce All operator

ReduceAnyOp

reduce_any operation: Reduce Any operator

ReduceAnyOpBuilder

Builder for reduce_any operation: Reduce Any operator

ReduceMaxOp

reduce_max operation: Reduce Max operator

ReduceMaxOpBuilder

Builder for reduce_max operation: Reduce Max operator

ReduceMinOp

reduce_min operation: Reduce Min operator

ReduceMinOpBuilder

Builder for reduce_min operation: Reduce Min operator

ReduceProdOp

reduce_prod operation: Reduce Prod operator

ReduceProdOpBuilder

Builder for reduce_prod operation: Reduce Prod operator

ReduceSumOp

reduce_sum operation: Reduce Sum operator

ReduceSumOpBuilder

Builder for reduce_sum operation: Reduce Sum operator

RescaleOp

rescale operation: Tosa rescale operator

RescaleOpBuilder

Builder for rescale operation: Tosa rescale operator

ReshapeOp

reshape operation: Reshape operator

ReshapeOpBuilder

Builder for reshape operation: Reshape operator

ResizeOp

resize operation: Resize operation, supports various resize/upsample modes

ResizeOpBuilder

Builder for resize operation: Resize operation, supports various resize/upsample modes

ReverseOp

reverse operation: Reverse operator

ReverseOpBuilder

Builder for reverse operation: Reverse operator

RsqrtOp

rsqrt operation: Elementwise 1/sqrt op

RsqrtOpBuilder

Builder for rsqrt operation: Elementwise 1/sqrt op

ScatterOp

scatter operation: Scatter operation,

ScatterOpBuilder

Builder for scatter operation: Scatter operation,

SelectOp

select operation: Elementwise select operator

SelectOpBuilder

Builder for select operation: Elementwise select operator

SigmoidOp

sigmoid operation: Computes elementwise sigmoid of input.

SigmoidOpBuilder

Builder for sigmoid operation: Computes elementwise sigmoid of input.

SliceOp

slice operation: Slice operator

SliceOpBuilder

Builder for slice operation: Slice operator

SubOp

sub operation: Elementwise subtraction operator

SubOpBuilder

Builder for sub operation: Elementwise subtraction operator

TableOp

table operation: Table lookup op

TableOpBuilder

Builder for table operation: Table lookup op

TanhOp

tanh operation: Computes elementwise hyperbolic tangent of input

TanhOpBuilder

Builder for tanh operation: Computes elementwise hyperbolic tangent of input

TileOp

tile operation: Tile operator

TileOpBuilder

Builder for tile operation: Tile operator

TransposeConv2DOp

transpose_conv2d operation: Transpose 2D Convolution operator.

TransposeConv2DOpBuilder

Builder for transpose_conv2d operation: Transpose 2D Convolution operator.

TransposeOp

transpose operation: Transpose operator

TransposeOpBuilder

Builder for transpose operation: Transpose operator

WhileOp

while_loop operation: Output = input; While (Cond(output)) {output = Body(output)}

WhileOpBuilder

Builder for while_loop operation: Output = input; While (Cond(output)) {output = Body(output)}

YieldOp

yield operation: Yield operator

YieldOpBuilder

Builder for yield operation: Yield operator

Functions

abs

Create a new abs operation: Elementwise abs op

add

Create a new add operation: Elementwise addition operator

apply_scale

Create a new apply_scale operation: Rescale scalar operator for Tosa tensor operators

argmax

Create a new argmax operation: Perform argmax on the input.

arithmetic_right_shift

Create a new arithmetic_right_shift operation: Elementwise Arithmetic Right Shift

avg_pool_2_d

Create a new avg_pool2d operation: Performs max pooling on the input.

bitwise_and

Create a new bitwise_and operation: Bitwise AND operator

bitwise_not

Create a new bitwise_not operation: Bitwise NOT operator

bitwise_or

Create a new bitwise_or operation: Bitwise OR operator

bitwise_xor

Create a new bitwise_xor operation: Bitwise XOR operator

cast

Create a new cast operation: Cast operation

ceil

Create a new ceil operation: Elementwise ceil op

clamp

Create a new clamp operation: Computes clamp(features, min, max).

clz

Create a new clz operation: Elementwise count leading zero op

concat

Create a new concat operation: Concatenates tensors along one dimension.

cond_if

Create a new cond_if operation: Conditional if operator

const

Create a new const operation: Constant op.

conv_2_d

Create a new conv2d operation: 2D Convolution Operator

conv_3_d

Create a new conv3d operation: 3D Convolution operator

custom

Create a new custom operation: Custom operator wrapper for Tosa

depthwise_conv_2_d

Create a new depthwise_conv2d operation: Depthwise 2D Convolution operator

div

Create a new div operation: Integer divide operator

equal

Create a new equal operation: Returns the truth value of (x == y) element-wise.

exp

Create a new exp operation: Elementwise exp op

floor

Create a new floor operation: Elementwise floor op

fully_connected

Create a new fully_connected operation: Fully Connected operator

gather

Create a new gather operation: Gather operation,

greater

Create a new greater operation: Returns the truth value of (x > y) element-wise.

greater_equal

Create a new greater_equal operation: Returns the truth value of (x >= y) element-wise.

identity

Create a new identity operation: Identity operator

log

Create a new log operation: Elementwise log op

logical_and

Create a new logical_and operation: Returns the truth value of x AND y element-wise.

logical_left_shift

Create a new logical_left_shift operation: Elementwise Logical Left Shift

logical_not

Create a new logical_not operation: Returns the truth value of NOT x element-wise.

logical_or

Create a new logical_or operation: Returns the truth value of x OR y element-wise.

logical_right_shift

Create a new logical_right_shift operation: Elementwise Logical Right Shift

logical_xor

Create a new logical_xor operation: Returns the truth value of x XOR y element-wise.

matmul

Create a new matmul operation: Matrix multiplication with bias

max_pool_2_d

Create a new max_pool2d operation: Performs max pooling on the input.

maximum

Create a new maximum operation: Elementwise Maximum

minimum

Create a new minimum operation: Elementwise Minimum

mul

Create a new mul operation: Multiplication operator

negate

Create a new negate operation: Elementwise negate op

pad

Create a new pad operation: Pads a tensor with value specified.

pow

Create a new pow operation: Computes the power of one value to another.

reciprocal

Create a new reciprocal operation: Elementwise reciprocal op

reduce_all

Create a new reduce_all operation: Reduce All operator

reduce_any

Create a new reduce_any operation: Reduce Any operator

reduce_max

Create a new reduce_max operation: Reduce Max operator

reduce_min

Create a new reduce_min operation: Reduce Min operator

reduce_prod

Create a new reduce_prod operation: Reduce Prod operator

reduce_sum

Create a new reduce_sum operation: Reduce Sum operator

rescale

Create a new rescale operation: Tosa rescale operator

reshape

Create a new reshape operation: Reshape operator

resize

Create a new resize operation: Resize operation, supports various resize/upsample modes

reverse

Create a new reverse operation: Reverse operator

rfft_2_d

Create a new rfft2d operation: Performs RFFT2D operation on the input.

rsqrt

Create a new rsqrt operation: Elementwise 1/sqrt op

scatter

Create a new scatter operation: Scatter operation,

select

Create a new select operation: Elementwise select operator

sigmoid

Create a new sigmoid operation: Computes elementwise sigmoid of input.

slice

Create a new slice operation: Slice operator

sub

Create a new sub operation: Elementwise subtraction operator

table

Create a new table operation: Table lookup op

tanh

Create a new tanh operation: Computes elementwise hyperbolic tangent of input

tile

Create a new tile operation: Tile operator

transpose

Create a new transpose operation: Transpose operator

transpose_conv_2_d

Create a new transpose_conv2d operation: Transpose 2D Convolution operator.

while_loop

Create a new while_loop operation: Output = input; While (Cond(output)) {output = Body(output)}

yield

Create a new yield operation: Yield operator