Module melior::dialect::ods::transform

transform dialect.

Structs

AlternativesOp

alternatives operation. Attempts sequences of transforms until one succeeds.

AlternativesOpBuilder

Builder for alternatives operation. Attempts sequences of transforms until one succeeds.

AnnotateOp

annotate operation. Annotates the target operation with an attribute by name.

AnnotateOpBuilder

Builder for annotate operation. Annotates the target operation with an attribute by name.

ApplyCanonicalizationPatternsOp

apply_patterns.canonicalization operation. Populates canonicalization patterns.

ApplyCanonicalizationPatternsOpBuilder

Builder for apply_patterns.canonicalization operation. Populates canonicalization patterns.

ApplyCommonSubexpressionEliminationOp

apply_cse operation. Eliminate common subexpressions in the body of the target op.

ApplyCommonSubexpressionEliminationOpBuilder

Builder for apply_cse operation. Eliminate common subexpressions in the body of the target op.

ApplyDeadCodeEliminationOp

apply_dce operation. Eliminate dead operations in the body of the target op.

ApplyDeadCodeEliminationOpBuilder

Builder for apply_dce operation. Eliminate dead operations in the body of the target op.

ApplyLoopInvariantCodeMotionOp

apply_licm operation. Move loop-invariant code out of a loop-like op.

ApplyLoopInvariantCodeMotionOpBuilder

Builder for apply_licm operation. Move loop-invariant code out of a loop-like op.

ApplyPatternsOp

apply_patterns operation. Greedily applies patterns to the body of the targeted op.

ApplyPatternsOpBuilder

Builder for apply_patterns operation. Greedily applies patterns to the body of the targeted op.

ApplyRegisteredPassOp

apply_registered_pass operation. Applies the specified registered pass.

ApplyRegisteredPassOpBuilder

Builder for apply_registered_pass operation. Applies the specified registered pass.

CastOp

cast operation.

CastOpBuilder

Builder for cast operation.

ForeachMatchOp

foreach_match operation. Applies named sequences when a named matcher succeeds.

ForeachMatchOpBuilder

Builder for foreach_match operation. Applies named sequences when a named matcher succeeds.

ForeachOp

foreach operation. Executes the body for each payload op.

ForeachOpBuilder

Builder for foreach operation. Executes the body for each payload op.

GetConsumersOfResult

get_consumers_of_result operation. Get handle to the consumers of this operation’s result number.

GetConsumersOfResultBuilder

Builder for get_consumers_of_result operation. Get handle to the consumers of this operation’s result number.

GetDefiningOp

get_defining_op operation. Get handle to the defining op of a value.

GetDefiningOpBuilder

Builder for get_defining_op operation. Get handle to the defining op of a value.

GetParentOp

get_parent_op operation. Gets handles to the closest isolated-from-above parents.

GetParentOpBuilder

Builder for get_parent_op operation. Gets handles to the closest isolated-from-above parents.

GetProducerOfOperand

get_producer_of_operand operation. Get handle to the producer of this operation’s operand number.

GetProducerOfOperandBuilder

Builder for get_producer_of_operand operation. Get handle to the producer of this operation’s operand number.

GetResultOp

get_result operation. Get handle to the a result of the targeted op.

GetResultOpBuilder

Builder for get_result operation. Get handle to the a result of the targeted op.

GetTypeOp

get_type operation. Get a parameter containing the type of the given value.

GetTypeOpBuilder

Builder for get_type operation. Get a parameter containing the type of the given value.

IncludeOp

include operation. Includes a named transform sequence.

IncludeOpBuilder

Builder for include operation. Includes a named transform sequence.

MatchOperationNameOp

match.operation_name operation. Matches a single operation of one of the given kinds.

MatchOperationNameOpBuilder

Builder for match.operation_name operation. Matches a single operation of one of the given kinds.

MatchParamCmpIOp

match.param.cmpi operation. Matches if two parameter lists are associated with the same value.

MatchParamCmpIOpBuilder

Builder for match.param.cmpi operation. Matches if two parameter lists are associated with the same value.

MergeHandlesOp

merge_handles operation. Merges handles into one pointing to the union of payload ops.

MergeHandlesOpBuilder

Builder for merge_handles operation. Merges handles into one pointing to the union of payload ops.

NamedSequenceOp

named_sequence operation. Named transform sequence that can be included elsewhere.

NamedSequenceOpBuilder

Builder for named_sequence operation. Named transform sequence that can be included elsewhere.

ParamConstantOp

param.constant operation. Produces a new transform dialect parameter value associated with the given attribute.

ParamConstantOpBuilder

Builder for param.constant operation. Produces a new transform dialect parameter value associated with the given attribute.

PrintOp

print operation. Dump each payload op.

PrintOpBuilder

Builder for print operation. Dump each payload op.

ReplicateOp

replicate operation. Lists payload ops multiple times in the new handle.

ReplicateOpBuilder

Builder for replicate operation. Lists payload ops multiple times in the new handle.

SelectOp

select operation. Select payload ops by name.

SelectOpBuilder

Builder for select operation. Select payload ops by name.

SequenceOp

sequence operation. Contains a sequence of other transform ops to apply.

SequenceOpBuilder

Builder for sequence operation. Contains a sequence of other transform ops to apply.

SplitHandleOp

split_handle operation. Splits a handle of payload ops into handles with a single op.

SplitHandleOpBuilder

Builder for split_handle operation. Splits a handle of payload ops into handles with a single op.

VerifyOp

verify operation. Verifies the targeted ops.

VerifyOpBuilder

Builder for verify operation. Verifies the targeted ops.

YieldOp

yield operation. Yields operation handles from a transform IR region.

YieldOpBuilder

Builder for yield operation. Yields operation handles from a transform IR region.

Functions

alternatives

Creates a new alternatives operation. Attempts sequences of transforms until one succeeds.

annotate

Creates a new annotate operation. Annotates the target operation with an attribute by name.

apply_cse

Creates a new apply_cse operation. Eliminate common subexpressions in the body of the target op.

apply_dce

Creates a new apply_dce operation. Eliminate dead operations in the body of the target op.

apply_licm

Creates a new apply_licm operation. Move loop-invariant code out of a loop-like op.

apply_patterns

Creates a new apply_patterns operation. Greedily applies patterns to the body of the targeted op.

apply_patterns_canonicalization

Creates a new apply_patterns.canonicalization operation. Populates canonicalization patterns.

apply_registered_pass

Creates a new apply_registered_pass operation. Applies the specified registered pass.

cast

Creates a new cast operation.

foreach

Creates a new foreach operation. Executes the body for each payload op.

foreach_match

Creates a new foreach_match operation. Applies named sequences when a named matcher succeeds.

get_consumers_of_result

Creates a new get_consumers_of_result operation. Get handle to the consumers of this operation’s result number.

get_defining_op

Creates a new get_defining_op operation. Get handle to the defining op of a value.

get_parent_op

Creates a new get_parent_op operation. Gets handles to the closest isolated-from-above parents.

get_producer_of_operand

Creates a new get_producer_of_operand operation. Get handle to the producer of this operation’s operand number.

get_result

Creates a new get_result operation. Get handle to the a result of the targeted op.

get_type

Creates a new get_type operation. Get a parameter containing the type of the given value.

include

Creates a new include operation. Includes a named transform sequence.

match_operation_name

Creates a new match.operation_name operation. Matches a single operation of one of the given kinds.

match_param_cmpi

Creates a new match.param.cmpi operation. Matches if two parameter lists are associated with the same value.

merge_handles

Creates a new merge_handles operation. Merges handles into one pointing to the union of payload ops.

named_sequence

Creates a new named_sequence operation. Named transform sequence that can be included elsewhere.

param_constant

Creates a new param.constant operation. Produces a new transform dialect parameter value associated with the given attribute.

print

Creates a new print operation. Dump each payload op.

replicate

Creates a new replicate operation. Lists payload ops multiple times in the new handle.

select

Creates a new select operation. Select payload ops by name.

sequence

Creates a new sequence operation. Contains a sequence of other transform ops to apply.

split_handle

Creates a new split_handle operation. Splits a handle of payload ops into handles with a single op.

verify

Creates a new verify operation. Verifies the targeted ops.

yield

Creates a new yield operation. Yields operation handles from a transform IR region.