Module melior::dialect::ods::transform

transform dialect.

Structs

AlternativesOperation

An alternatives operation. Attempts sequences of transforms until one succeeds.

AlternativesOperationBuilder

A builder for an alternatives operation.

AnnotateOperation

An annotate operation. Annotates the target operation with an attribute by name.

AnnotateOperationBuilder

A builder for an annotate operation.

ApplyCanonicalizationPatternsOperation

An apply_patterns.canonicalization operation. Populates canonicalization patterns.

ApplyCanonicalizationPatternsOperationBuilder

A builder for an apply_patterns.canonicalization operation.

ApplyCommonSubexpressionEliminationOperation

An apply_cse operation. Eliminate common subexpressions in the body of the target op.

ApplyCommonSubexpressionEliminationOperationBuilder

A builder for an apply_cse operation.

ApplyDeadCodeEliminationOperation

An apply_dce operation. Eliminate dead operations in the body of the target op.

ApplyDeadCodeEliminationOperationBuilder

A builder for an apply_dce operation.

ApplyLoopInvariantCodeMotionOperation

An apply_licm operation. Move loop-invariant code out of a loop-like op.

ApplyLoopInvariantCodeMotionOperationBuilder

A builder for an apply_licm operation.

ApplyPatternsOperation

An apply_patterns operation. Greedily applies patterns to the body of the targeted op.

ApplyPatternsOperationBuilder

A builder for an apply_patterns operation.

ApplyRegisteredPassOperation

An apply_registered_pass operation. Applies the specified registered pass.

ApplyRegisteredPassOperationBuilder

A builder for an apply_registered_pass operation.

CastOperation

A cast operation.

CastOperationBuilder

A builder for a cast operation.

ForeachMatchOperation

A foreach_match operation. Applies named sequences when a named matcher succeeds.

ForeachMatchOperationBuilder

A builder for a foreach_match operation.

ForeachOperation

A foreach operation. Executes the body for each payload op.

ForeachOperationBuilder

A builder for a foreach operation.

GetConsumersOfResultOperation

A get_consumers_of_result operation. Get handle to the consumers of this operation’s result number.

GetConsumersOfResultOperationBuilder

A builder for a get_consumers_of_result operation.

GetDefiningOperation

A get_defining_op operation. Get handle to the defining op of a value.

GetDefiningOperationBuilder

A builder for a get_defining_op operation.

GetParentOperation

A get_parent_op operation. Gets handles to the closest isolated-from-above parents.

GetParentOperationBuilder

A builder for a get_parent_op operation.

GetProducerOfOperandOperation

A get_producer_of_operand operation. Get handle to the producer of this operation’s operand number.

GetProducerOfOperandOperationBuilder

A builder for a get_producer_of_operand operation.

GetResultOperation

A get_result operation. Get handle to the a result of the targeted op.

GetResultOperationBuilder

A builder for a get_result operation.

GetTypeOperation

A get_type operation. Get a parameter containing the type of the given value.

GetTypeOperationBuilder

A builder for a get_type operation.

IncludeOperation

An include operation. Includes a named transform sequence.

IncludeOperationBuilder

A builder for an include operation.

MatchOperationNameOperation

A match.operation_name operation. Matches a single operation of one of the given kinds.

MatchOperationNameOperationBuilder

A builder for a match.operation_name operation.

MatchParamCmpIOperation

A match.param.cmpi operation. Matches if two parameter lists are associated with the same value.

MatchParamCmpIOperationBuilder

A builder for a match.param.cmpi operation.

MergeHandlesOperation

A merge_handles operation. Merges handles into one pointing to the union of payload ops.

MergeHandlesOperationBuilder

A builder for a merge_handles operation.

NamedSequenceOperation

A named_sequence operation. Named transform sequence that can be included elsewhere.

NamedSequenceOperationBuilder

A builder for a named_sequence operation.

ParamConstantOperation

A param.constant operation. Produces a new transform dialect parameter value associated with the given attribute.

ParamConstantOperationBuilder

A builder for a param.constant operation.

PrintOperation

A print operation. Dump each payload op.

PrintOperationBuilder

A builder for a print operation.

ReplicateOperation

A replicate operation. Lists payload ops multiple times in the new handle.

ReplicateOperationBuilder

A builder for a replicate operation.

SelectOperation

A select operation. Select payload ops by name.

SelectOperationBuilder

A builder for a select operation.

SequenceOperation

A sequence operation. Contains a sequence of other transform ops to apply.

SequenceOperationBuilder

A builder for a sequence operation.

SplitHandleOperation

A split_handle operation. Splits a handle of payload ops into handles with a single op.

SplitHandleOperationBuilder

A builder for a split_handle operation.

VerifyOperation

A verify operation. Verifies the targeted ops.

VerifyOperationBuilder

A builder for a verify operation.

YieldOperation

A yield operation. Yields operation handles from a transform IR region.

YieldOperationBuilder

A builder for a yield operation.

Functions

alternatives

Creates an alternatives operation.

annotate

Creates an annotate operation.

apply_cse

Creates an apply_cse operation.

apply_dce

Creates an apply_dce operation.

apply_licm

Creates an apply_licm operation.

apply_patterns

Creates an apply_patterns operation.

apply_patterns_canonicalization

Creates an apply_patterns.canonicalization operation.

apply_registered_pass

Creates an apply_registered_pass operation.

cast

Creates a cast operation.

foreach

Creates a foreach operation.

foreach_match

Creates a foreach_match operation.

get_consumers_of_result

Creates a get_consumers_of_result operation.

get_defining_op

Creates a get_defining_op operation.

get_parent_op

Creates a get_parent_op operation.

get_producer_of_operand

Creates a get_producer_of_operand operation.

get_result

Creates a get_result operation.

get_type

Creates a get_type operation.

include

Creates an include operation.

match_operation_name

Creates a match.operation_name operation.

match_param_cmpi

Creates a match.param.cmpi operation.

merge_handles

Creates a merge_handles operation.

named_sequence

Creates a named_sequence operation.

param_constant

Creates a param.constant operation.

print

Creates a print operation.

replicate

Creates a replicate operation.

select

Creates a select operation.

sequence

Creates a sequence operation.

split_handle

Creates a split_handle operation.

verify

Creates a verify operation.

yield

Creates a yield operation.