Module melior::dialect::ods::vector

vector dialect.

Structs

BitCastOp

bitcast operation: Bitcast casts between vectors

BitCastOpBuilder

Builder for bitcast operation: Bitcast casts between vectors

BroadcastOp

broadcast operation: Broadcast operation

BroadcastOpBuilder

Builder for broadcast operation: Broadcast operation

CompressStoreOp

compressstore operation: Writes elements selectively from a vector as defined by a mask

CompressStoreOpBuilder

Builder for compressstore operation: Writes elements selectively from a vector as defined by a mask

ConstantMaskOp

constant_mask operation: Creates a constant vector mask

ConstantMaskOpBuilder

Builder for constant_mask operation: Creates a constant vector mask

ContractionOp

contract operation: Vector contraction operation

ContractionOpBuilder

Builder for contract operation: Vector contraction operation

CreateMaskOp

create_mask operation: Creates a vector mask

CreateMaskOpBuilder

Builder for create_mask operation: Creates a vector mask

ExpandLoadOp

expandload operation: Reads elements from memory and spreads them into a vector as defined by a mask

ExpandLoadOpBuilder

Builder for expandload operation: Reads elements from memory and spreads them into a vector as defined by a mask

ExtractElementOp

extractelement operation: Extractelement operation

ExtractElementOpBuilder

Builder for extractelement operation: Extractelement operation

ExtractOp

extract operation: Extract operation

ExtractOpBuilder

Builder for extract operation: Extract operation

ExtractStridedSliceOp

extract_strided_slice operation: Extract_strided_slice operation

ExtractStridedSliceOpBuilder

Builder for extract_strided_slice operation: Extract_strided_slice operation

FMAOp

fma operation: Vector fused multiply-add

FMAOpBuilder

Builder for fma operation: Vector fused multiply-add

FlatTransposeOp

flat_transpose operation: Vector matrix transposition on flattened 1-D MLIR vectors

FlatTransposeOpBuilder

Builder for flat_transpose operation: Vector matrix transposition on flattened 1-D MLIR vectors

GatherOp

gather operation: gathers elements from memory or ranked tensor into a vector as defined by an index vector and a mask vector

GatherOpBuilder

Builder for gather operation: gathers elements from memory or ranked tensor into a vector as defined by an index vector and a mask vector

InsertElementOp

insertelement operation: Insertelement operation

InsertElementOpBuilder

Builder for insertelement operation: Insertelement operation

InsertOp

insert operation: Insert operation

InsertOpBuilder

Builder for insert operation: Insert operation

InsertStridedSliceOp

insert_strided_slice operation: Strided_slice operation

InsertStridedSliceOpBuilder

Builder for insert_strided_slice operation: Strided_slice operation

LoadOp

load operation: Reads an n-D slice of memory into an n-D vector

LoadOpBuilder

Builder for load operation: Reads an n-D slice of memory into an n-D vector

MaskOp

mask operation: Predicates a maskable vector operation

MaskOpBuilder

Builder for mask operation: Predicates a maskable vector operation

MaskedLoadOp

maskedload operation: Loads elements from memory into a vector as defined by a mask vector

MaskedLoadOpBuilder

Builder for maskedload operation: Loads elements from memory into a vector as defined by a mask vector

MaskedStoreOp

maskedstore operation: Stores elements from a vector into memory as defined by a mask vector

MaskedStoreOpBuilder

Builder for maskedstore operation: Stores elements from a vector into memory as defined by a mask vector

MatmulOp

matrix_multiply operation: Vector matrix multiplication op that operates on flattened 1-D MLIR vectors

MatmulOpBuilder

Builder for matrix_multiply operation: Vector matrix multiplication op that operates on flattened 1-D MLIR vectors

MultiDimReductionOp

multi_reduction operation: Multi-dimensional reduction operation

MultiDimReductionOpBuilder

Builder for multi_reduction operation: Multi-dimensional reduction operation

OuterProductOp

outerproduct operation: Vector outerproduct with optional fused add

OuterProductOpBuilder

Builder for outerproduct operation: Vector outerproduct with optional fused add

PrintOp

print operation: Print operation (for testing and debugging)

PrintOpBuilder

Builder for print operation: Print operation (for testing and debugging)

ReductionOp

reduction operation: Reduction operation

ReductionOpBuilder

Builder for reduction operation: Reduction operation

ReshapeOp

reshape operation: Vector reshape operation

ReshapeOpBuilder

Builder for reshape operation: Vector reshape operation

ScalableExtractOp

scalable.extract operation: Extract subvector from scalable vector operation

ScalableExtractOpBuilder

Builder for scalable.extract operation: Extract subvector from scalable vector operation

ScalableInsertOp

scalable.insert operation: Insert subvector into scalable vector operation

ScalableInsertOpBuilder

Builder for scalable.insert operation: Insert subvector into scalable vector operation

ScanOp

scan operation: Scan operation

ScanOpBuilder

Builder for scan operation: Scan operation

ScatterOp

scatter operation: scatters elements from a vector into memory as defined by an index vector and a mask vector

ScatterOpBuilder

Builder for scatter operation: scatters elements from a vector into memory as defined by an index vector and a mask vector

ShapeCastOp

shape_cast operation: Shape_cast casts between vector shapes

ShapeCastOpBuilder

Builder for shape_cast operation: Shape_cast casts between vector shapes

ShuffleOp

shuffle operation: Shuffle operation

ShuffleOpBuilder

Builder for shuffle operation: Shuffle operation

SplatOp

splat operation: Vector splat or broadcast operation

SplatOpBuilder

Builder for splat operation: Vector splat or broadcast operation

StoreOp

store operation: Writes an n-D vector to an n-D slice of memory

StoreOpBuilder

Builder for store operation: Writes an n-D vector to an n-D slice of memory

TransferReadOp

transfer_read operation: Reads a supervector from memory into an SSA vector value.

TransferReadOpBuilder

Builder for transfer_read operation: Reads a supervector from memory into an SSA vector value.

TransferWriteOp

transfer_write operation: The vector.transfer_write op writes a supervector to memory.

TransferWriteOpBuilder

Builder for transfer_write operation: The vector.transfer_write op writes a supervector to memory.

TransposeOp

transpose operation: Vector transpose operation

TransposeOpBuilder

Builder for transpose operation: Vector transpose operation

TypeCastOp

type_cast operation: Type_cast op converts a scalar memref to a vector memref

TypeCastOpBuilder

Builder for type_cast operation: Type_cast op converts a scalar memref to a vector memref

VectorScaleOp

vscale operation: Load vector scale size

VectorScaleOpBuilder

Builder for vscale operation: Load vector scale size

WarpExecuteOnLane0Op

warp_execute_on_lane_0 operation: Executes operations in the associated region on thread #0 of aSPMD program

WarpExecuteOnLane0OpBuilder

Builder for warp_execute_on_lane_0 operation: Executes operations in the associated region on thread #0 of aSPMD program

YieldOp

yield operation: Terminates and yields values from vector regions.

YieldOpBuilder

Builder for yield operation: Terminates and yields values from vector regions.

Functions

bitcast

Creates a new bitcast operation: Bitcast casts between vectors

broadcast

Creates a new broadcast operation: Broadcast operation

compressstore

Creates a new compressstore operation: Writes elements selectively from a vector as defined by a mask

constant_mask

Creates a new constant_mask operation: Creates a constant vector mask

contract

Creates a new contract operation: Vector contraction operation

create_mask

Creates a new create_mask operation: Creates a vector mask

expandload

Creates a new expandload operation: Reads elements from memory and spreads them into a vector as defined by a mask

extract

Creates a new extract operation: Extract operation

extract_strided_slice

Creates a new extract_strided_slice operation: Extract_strided_slice operation

extractelement

Creates a new extractelement operation: Extractelement operation

flat_transpose

Creates a new flat_transpose operation: Vector matrix transposition on flattened 1-D MLIR vectors

fma

Creates a new fma operation: Vector fused multiply-add

gather

Creates a new gather operation: gathers elements from memory or ranked tensor into a vector as defined by an index vector and a mask vector

insert

Creates a new insert operation: Insert operation

insert_strided_slice

Creates a new insert_strided_slice operation: Strided_slice operation

insertelement

Creates a new insertelement operation: Insertelement operation

load

Creates a new load operation: Reads an n-D slice of memory into an n-D vector

mask

Creates a new mask operation: Predicates a maskable vector operation

maskedload

Creates a new maskedload operation: Loads elements from memory into a vector as defined by a mask vector

maskedstore

Creates a new maskedstore operation: Stores elements from a vector into memory as defined by a mask vector

matrix_multiply

Creates a new matrix_multiply operation: Vector matrix multiplication op that operates on flattened 1-D MLIR vectors

multi_reduction

Creates a new multi_reduction operation: Multi-dimensional reduction operation

outerproduct

Creates a new outerproduct operation: Vector outerproduct with optional fused add

print

Creates a new print operation: Print operation (for testing and debugging)

reduction

Creates a new reduction operation: Reduction operation

reshape

Creates a new reshape operation: Vector reshape operation

scalable_extract

Creates a new scalable.extract operation: Extract subvector from scalable vector operation

scalable_insert

Creates a new scalable.insert operation: Insert subvector into scalable vector operation

scan

Creates a new scan operation: Scan operation

scatter

Creates a new scatter operation: scatters elements from a vector into memory as defined by an index vector and a mask vector

shape_cast

Creates a new shape_cast operation: Shape_cast casts between vector shapes

shuffle

Creates a new shuffle operation: Shuffle operation

splat

Creates a new splat operation: Vector splat or broadcast operation

store

Creates a new store operation: Writes an n-D vector to an n-D slice of memory

transfer_read

Creates a new transfer_read operation: Reads a supervector from memory into an SSA vector value.

transfer_write

Creates a new transfer_write operation: The vector.transfer_write op writes a supervector to memory.

transpose

Creates a new transpose operation: Vector transpose operation

type_cast

Creates a new type_cast operation: Type_cast op converts a scalar memref to a vector memref

vscale

Creates a new vscale operation: Load vector scale size

warp_execute_on_lane_0

Creates a new warp_execute_on_lane_0 operation: Executes operations in the associated region on thread #0 of aSPMD program

yield

Creates a new yield operation: Terminates and yields values from vector regions.