Struct melior::ir::operation::Operation

pub struct Operation<'c> { /* private fields */ }

An operation.

Implementations

impl<'c> Operation<'c>

pub fn context(&self) -> ContextRef<'c>

Gets a context.

pub fn name(&self) -> Identifier<'c>

Gets a name.

pub fn block(&self) -> Option<BlockRef<'c, '_>>

Gets a block.

pub fn operand_count(&self) -> usize

Gets the number of operands.

pub fn operand(&self, index: usize) -> Result<Value<'c, '_>, Error>

Gets the operand at a position.

pub fn operands(&self) -> impl Iterator<Item = Value<'c, '_>>

Gets all operands.

pub fn result_count(&self) -> usize

Gets the number of results.

pub fn result(&self, index: usize) -> Result<OperationResult<'c, '_>, Error>

Gets a result at a position.

pub fn results(&self) -> impl Iterator<Item = OperationResult<'c, '_>>

Gets all results.

pub fn region_count(&self) -> usize

Gets the number of regions.

pub fn region(&self, index: usize) -> Result<RegionRef<'c, '_>, Error>

Gets a region at a position.

pub fn regions(&self) -> impl Iterator<Item = RegionRef<'c, '_>>

Gets all regions.

pub fn successor_count(&self) -> usize

Gets the number of successors.

pub fn successor(&self, index: usize) -> Result<BlockRef<'c, '_>, Error>

Gets a successor at a position.

pub fn successors(&self) -> impl Iterator<Item = BlockRef<'c, '_>>

Gets all successors.

pub fn attribute_count(&self) -> usize

Gets the number of attributes.

pub fn attribute_at( &self, index: usize ) -> Result<(Identifier<'c>, Attribute<'c>), Error>

Gets a attribute at a position.

pub fn attributes( &self ) -> impl Iterator<Item = (Identifier<'c>, Attribute<'c>)> + '_

Gets all attributes.

pub fn attribute(&self, name: &str) -> Result<Attribute<'c>, Error>

Gets a attribute with the given name.

pub fn has_attribute(&self, name: &str) -> bool

Checks if the operation has a attribute with the given name.

pub fn set_attribute(&mut self, name: &str, attribute: &Attribute<'c>)

Sets the attribute with the given name to the given attribute.

pub fn remove_attribute(&mut self, name: &str) -> Result<(), Error>

Removes the attribute with the given name.

pub fn next_in_block(&self) -> Option<OperationRef<'c, '_>>

Gets the next operation in the same block.

pub fn verify(&self) -> bool

Verifies an operation.

pub fn dump(&self)

Dumps an operation.

pub fn to_string_with_flags( &self, flags: OperationPrintingFlags ) -> Result<String, Error>

Prints an operation with flags.

pub unsafe fn from_raw(raw: MlirOperation) -> Self

Creates an operation from a raw object.

Safety

A raw object must be valid.

pub unsafe fn from_option_raw(raw: MlirOperation) -> Option<Self>

Creates an optional operation from a raw object.

Safety

A raw object must be valid.

pub fn into_raw(self) -> MlirOperation

Converts an operation into a raw object.