Trait melior::ir::ValueLike

pub trait ValueLike {
    // Required method
    fn to_raw(&self) -> MlirValue;

    // Provided methods
    fn type(&self) -> Type<'_> { ... }
    fn is_block_argument(&self) -> bool { ... }
    fn is_operation_result(&self) -> bool { ... }
    fn dump(&self) { ... }
}

Trait for value-like types.

Required Methods

fn to_raw(&self) -> MlirValue

Converts a value into a raw value.

Provided Methods

fn type(&self) -> Type<'_>

Gets a type.

fn is_block_argument(&self) -> bool

Returns true if a value is a block argument.

fn is_operation_result(&self) -> bool

Returns true if a value is an operation result.

fn dump(&self)

Dumps a value.